==> old_versions/0_Joined_PIN_level_data_taxyear2021.Rmd <==
---
title: "Create Joined PIN Level Data for 2021"
author: "AWM"
date: "`r Sys.Date()`"
output: html_document
---

```{r setup, warning=FALSE, message=FALSE, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, warning = FALSE, message = FALSE)

library(tidyverse)
library(data.table)
library(ptaxsim)
library(glue)
library(DBI)

```

# Bring in data


```{r}
  
ptaxsim_db_conn <- DBI::dbConnect(RSQLite::SQLite(), "./ptaxsim.db/ptaxsim-2023.0.0.db")

# Identify relevant agency names, numbers, and minor types from agency_info table

muni_agency_names <- DBI::dbGetQuery(
  ptaxsim_db_conn,
  "SELECT DISTINCT agency_num, agency_name, minor_type
  FROM agency_info
  WHERE minor_type = 'MUNI'
  OR agency_num = '020060000'  
  "
  )

# Identify tax codes associated with relevant agencies

muni_tax_codes <- DBI::dbGetQuery(
  ptaxsim_db_conn,
  glue_sql("
  SELECT DISTINCT agency_num, tax_code_num
  FROM tax_code
  WHERE agency_num IN ({muni_agency_names$agency_num*})
  AND year = 2021
  ",
  .con = ptaxsim_db_conn
  ))

# Pull in the tax code rates

tax_codes <- DBI::dbGetQuery(
  ptaxsim_db_conn,
  glue_sql("
  SELECT DISTINCT tax_code_num, tax_code_rate
  FROM tax_code
  WHERE year = 2021
  ",
  .con = ptaxsim_db_conn
  )
  )

# Identify PINs associated with tax codes

cook_pins <- DBI::dbGetQuery(
    ptaxsim_db_conn, 
    glue_sql("
             SELECT DISTINCT pin, class, tax_code_num 
             FROM pin 
             WHERE tax_code_num IN ({tax_codes$tax_code_num*}) 
             AND year = 2021", 
             .con = ptaxsim_db_conn))
```

```{r}
year <- 2021

nicknames <- readxl::read_excel("./Necessary_Files/muni_shortnames.xlsx")


## All tax codes.
## tax codes within municipalities have additional info
tc_muninames <- tax_codes %>%
  left_join(muni_tax_codes) %>%
  left_join(muni_agency_names) %>%
  #select(-agency_rate) %>%
  left_join(nicknames) %>%
  select(-c(minor_type, short_name,
           # `Column1`, `Most recent reassessed`,
            agency_number))
ccao_loa <- read_csv("./inputs/ccao_loa.csv") %>% 
  mutate(class_code = as.character(class_code))

  
# tif_distribution table contains eav 

tif_distrib <- DBI::dbGetQuery(
    ptaxsim_db_conn,
    glue_sql(
  'SELECT *
  FROM tif_distribution
  WHERE year = 2021'), .con = ptaxsim_db_conn)
          
eq_factor <- DBI::dbGetQuery(
    ptaxsim_db_conn,
    glue_sql(
  'SELECT *
  FROM eq_factor
  WHERE year = 2021'), .con = ptaxsim_db_conn)   
```



Equalization factor for 2021 is 3.02 Use this to calculate the taxable AV from EAV. 

```{r}
# 2021 Tax Bills ---------------------------------------------------------------

# Create value for 2021 equalization factor

eq2021 = 3.0027

# Use PTAXSIM::tax_bill to pull tax bill information at PIN level
# simplify = FALSE dramatically increases the data for each PIN

taxbills <- tax_bill(2021,  cook_pins$pin, simplify = FALSE)

# Summarizes information from PIN-agency level to PIN level

pin14_bills <- taxbills %>%
    
  mutate(total_bill = final_tax_to_dist + final_tax_to_tif,
                     eq_av = av/eq2021) %>% # from each taxing agency
  
  group_by(tax_code, class, pin) %>%

  
  summarize(total_billed = sum(total_bill, na.rm = TRUE), # total on someone's property tax bill
            av = first(av),            # PINs appear multiple times, uses first av appearance
            eq_av = first(eq_av),      # equalized AV = potential taxable value for homeowners
            eav = first(eav),          # EAV after exemptions are subtracted. 
            taxing_agency_count = n(), # number of taxing agencies that tax the pin
            final_tax_to_dist = sum(final_tax_to_dist, na.rm = TRUE), # portion of all levies paid by the pin
            final_tax_to_tif = sum(final_tax_to_tif, na.rm = TRUE), 
            tax_amt_exe = sum(tax_amt_exe, na.rm = TRUE),           # revenue lost due to exemptions
            tax_amt_pre_exe = sum(tax_amt_pre_exe, na.rm = TRUE),   # total rev before all exemptions
            tax_amt_post_exe = sum(tax_amt_post_exe, na.rm = TRUE), # total rev after all exemptions
            rpm_tif_to_cps = sum(rpm_tif_to_cps, na.rm = TRUE),     # not used
            rpm_tif_to_rpm = sum(rpm_tif_to_rpm, na.rm=TRUE),       # not used
            rpm_tif_to_dist = sum(rpm_tif_to_dist, na.rm=TRUE),     # not used
            tif_share = mean(tif_share, na.rm=TRUE),                # not used
  )  %>% 
  mutate(propclass_1dig = str_sub(class, 1, 1))

rm(taxbills)



# Exemptions at PIN level --------------------------------------------------


# finds all pins within Cook county and data on their exemptions
# joins tax code variable by pin
exemption_data <- lookup_pin(2021, cook_pins$pin) %>%
  left_join(cook_pins, by = c("pin", "class")) %>%
  
# Future preps for potential future obtained loa's
  left_join(ccao_loa, by = c("year", "class" = "class_code")) %>%
  
  mutate(all_exemptions = exe_homeowner + exe_senior + exe_freeze + exe_longtime_homeowner + 
         exe_disabled + exe_vet_returning + exe_vet_dis_lt50 + exe_vet_dis_50_69 + exe_vet_dis_ge70 ,
         abatements = exe_abate, #abatements get their own variable
         fmv = av / loa) %>%
  
# Create binary variables for exemptions
  
  mutate(zero_bill = ifelse(eav <= all_exemptions, 1, 0),
         has_HO_exemp = ifelse(exe_homeowner > 0, 1, 0),
         has_SF_exemp = ifelse(exe_senior > 0, 1, 0),
         has_FR_exemp = ifelse(exe_freeze > 0, 1, 0),
         has_LTHO_exemp = ifelse(exe_longtime_homeowner > 0, 1, 0),
         has_DI_exemp = ifelse(exe_disabled > 0, 1, 0),
         has_VR_exemp = ifelse(exe_vet_returning > 0, 1, 0),
         has_DV_exemp = ifelse(exe_vet_dis_lt50 + exe_vet_dis_50_69 + exe_vet_dis_ge70 > 0, 1, 0),
         has_AB_exemp = ifelse(exe_abate > 0, 1, 0),
         in_tif = ifelse(tax_code_num %in% tif_distrib$tax_code_num, 1, 0)
  )

rm(cook_pins)


#  change variable type to character so the join works.
  
#class_dict$class_code <- as.character(class_dict$class_code)

  # use the property class to make the major property types
  # joins the class_dict file to the pin_data classes
  # exemption_data <- class_dict %>%
  #   right_join(exemption_data
  #              #, 
  #             # by = "class",
  #             # relationship = "many-to-many"
  #             )
```

> The tax code level summaries created below are currently not being used. 

```{r eval = FALSE}
# Summarize PIN Exemptions ------------------------------------------------
  
  ## summarize pin level data to the tax code level for each type of property class
  exemptions_inCook_perTC <- exemption_data %>%
    group_by(tax_code_num, class) %>%
    summarize(year = first(year), 
              av = sum(av, na.rm = TRUE),
              eav = sum(eav, na.rm=TRUE),   # EAV after exemptions and abatements, from taxbill command
              fmv = sum(fmv, na.rm=TRUE),
              pin_count = n(),  # number of pins within each tax code and property combo
              
              all_exemptions = sum(all_exemptions),
              
              exe_homeowner = sum(exe_homeowner, na.rm=TRUE),
              exe_senior = sum(exe_senior, na.rm=TRUE),
              exe_freeze = sum(exe_freeze, na.rm=TRUE),
              exe_longtime_homeowner = sum(exe_longtime_homeowner, na.rm=TRUE),
              exe_disabled = sum(exe_disabled, na.rm=TRUE),
              exe_vet_returning = sum(exe_vet_returning, na.rm=TRUE),
              exe_vet_dis_lt50 = sum(exe_vet_dis_lt50, na.rm=TRUE),
              exe_vet_dis_50_69 = sum(exe_vet_dis_50_69, na.rm=TRUE),
              exe_vet_dis_ge70 = sum(exe_vet_dis_ge70, na.rm=TRUE),
              exe_vet_dis = sum(exe_vet_dis_lt50 + exe_vet_dis_50_69 + exe_vet_dis_ge70, na.rm=TRUE),
              exe_abate = sum(exe_abate, na.rm=TRUE),
              
              zero_bills = sum(zero_bill),
              has_HO_exemp = sum(has_HO_exemp),
              has_SF_exemp = sum(has_SF_exemp),
              has_FR_exemp = sum(has_FR_exemp),
              has_LTHO_exemp = sum(has_LTHO_exemp),
              has_DI_exemp = sum(has_DI_exemp),
              has_VR_exemp = sum(has_VR_exemp),
              has_DV_exemp = sum(has_DV_exemp),
              has_AB_exemp = sum(has_AB_exemp),
              in_tif = ifelse(tax_code_num %in% tif_distrib$tax_code_num, 1, 0)

    )
  
  
  exemptions_inCook_perTC <- exemptions_inCook_perTC %>% 
    left_join(tc_muninames)
```


```{r}
# Join Bills and Exemptions  ----------------------------------------------
  
  ## Add exemption types and values to the tax bill data at PIN level
  joined_pin_data <- left_join(pin14_bills, exemption_data, by = c("pin", "class" ))  %>%
    rename(av = av.x,
           eav =  eav.x,
           equalized_av = eav.y)
  
rm(exemption_data)
  
  
  ## Add tax code tax rate to PIN level data
joined_pin_data <- left_join(joined_pin_data, tc_muninames, by = c("tax_code" = "tax_code_num"))
```


```{r}
write_csv(joined_pin_data, "./Output/Dont_Upload/0_joined_PIN_data_2021.csv")
```


Need to merge PIN level data of tax bills and exemptions per PIN if the goal is to calculate new tax rates for different scenarios AND view individual level tax bill changes. (i.e. if we want to look at the "median taxpayer" then need to use this method).

Note: The EAV from the pin table is the original EAV (AV \* equalizer) without considering TIFs or exemptions.
