==> old_versions/2_Summed_Bills_by_TaxCode.Rmd <==
---
title: "Calculate all Pin Bills for 2021"
author: "Alea Wilbur"
date: "`r Sys.Date()`"
output:
  html_document:
    toc: yes
    toc_float: yes
    df_print: paged
    code_download: true
---

# Get Pin and Tax Bill Data

Run file `1_Get_All_Pin_Bills.Rmd` to replicate pulling the 2021 bills.

```{r setup}
knitr::opts_chunk$set(echo = TRUE, warning = FALSE, message = FALSE)


library(tidyverse)
library(DBI)
library(data.table)
library(ggspatial)
library(gstat)
library(here)
library(httr)
library(jsonlite)
library(ptaxsim)
library(sf)
library(stars)
library(glue)

# Create the DB connection with the default name expected by PTAXSIM functions
ptaxsim_db_conn <- DBI::dbConnect(RSQLite::SQLite(), "./ptaxsim.db/ptaxsim-2021.0.4.db")


options(digits=4, scipen = 999)

library(sf)
library(jsonlite)
library(httr)

# link to the API output as a JSON file
#muni_shp <- read_sf("https://gis.cookcountyil.gov/traditional/rest/services/politicalBoundary/MapServer/2/query?outFields=*&where=1%3D1&f=geojson")

#muni_shp <- read_json("muni_shp.json")
nicknames <- readxl::read_excel("./Necessary_Files/muni_shortnames.xlsx")

class_dict <- read_csv("./Necessary_Files/class_dict.csv")



```


```{r agency-dt}
# has EAV values, extensions by agency_num
agency_dt <- DBI::dbGetQuery(
  ptaxsim_db_conn,
  "SELECT *
  FROM agency
  WHERE year = 2021
  "
)

# grabs all unique muni names & numbs
# don't forget Cicero
muni_agency_names <- DBI::dbGetQuery(
  ptaxsim_db_conn,
  "SELECT DISTINCT agency_num, agency_name, minor_type
  FROM agency_info
  WHERE minor_type = 'MUNI'
  OR agency_num = '020060000'  

  "
)

muni_agency_names <- muni_agency_names %>% 
    mutate(first6 = str_sub(agency_num,1,6),
         first5 = str_sub(agency_num,1,5)) %>% 
  select(-minor_type)



#Makes a list of ALL taxing agencies, including TIFs, SSAs, etc.

# all agency names, numbers, and types
# includes TIF and non-TIF agencies
all_taxing_agencies <- DBI::dbGetQuery(
  ptaxsim_db_conn,
  "SELECT agency_num, agency_name, major_type, minor_type
  FROM agency_info
  "
) %>% mutate(first6 = str_sub(agency_num,1,6),
         first5 = str_sub(agency_num,1,5))


muni_agency_nums<- all_taxing_agencies %>% 
  filter(minor_type %in% c("MUNI") | 
           agency_num == "020060000") %>%
   select(agency_num)


# list of all taxcodes in municipalities. 
# This does NOT include unincorporated tax codes!!
muni_tax_codes <- DBI::dbGetQuery(
  ptaxsim_db_conn,
  glue_sql("
  SELECT*
  FROM tax_code
  WHERE agency_num IN ({muni_agency_names$agency_num*})
  AND year = 2021
  ",
  .con = ptaxsim_db_conn
  )
)# %>% select(-agency_rate)

# Agency number and agency name for all TIFs
TIF_agencies <- DBI::dbGetQuery(
  ptaxsim_db_conn,
  "SELECT DISTINCT agency_num, agency_name, major_type, minor_type
  FROM agency_info
  WHERE minor_type = 'TIF'
  "
)

unique_tif_taxcodes <- DBI::dbGetQuery(
  ptaxsim_db_conn, 
  glue_sql("
  SELECT DISTINCT tax_code_num
  FROM tax_code
  WHERE agency_num IN ({TIF_agencies$agency_num*})
  AND year = 2021
  ",
  .con = ptaxsim_db_conn
  )
)


tif_distrib <- DBI::dbGetQuery(
  ptaxsim_db_conn, 
  glue_sql("
  SELECT *
  FROM tif_distribution
  WHERE tax_code_num IN ({muni_tax_codes$tax_code_num*})
  AND year = 2021
  ",
  .con = ptaxsim_db_conn
  )
) %>% mutate(tax_code_num = as.character(tax_code_num))



all_taxing_agencies <- all_taxing_agencies %>% 
  left_join(muni_agency_names, by = c("first5", "first6")) %>% 
  rename(muni_name =  agency_name.y,
        muni_num = agency_num.y,
        agency_name = agency_name.x,
        agency_num = agency_num.x)


# combine taxing agency names and agency type to data table that has eav and extension values
agency_data <- right_join(agency_dt, all_taxing_agencies) %>% 
  # get rid of unneeded columns to make table outputs smaller
  select(-c(cty_dupage_eav:cty_livingston_eav)) %>% # drop some of the unused variables
  arrange(agency_num)
```

## Summing Tax Code level Values

using the pin data from ptaxsim, I calculate the eav of all properties, exempt EAV, the current tax base, and tax base if there were not exemptions

-   exempt EAV (summed from all exemption types)

-   current taxbase (using the `tif_distrib` table and percent of taxcode rev that goes to the tif). If a taxcode is a TIF taxcode, then do (EAV-exempt EAV) \* (1-%rev that goes to TIF). If the tax code is not a TIF tax code, then use the EAV-exempt EAV.

-   tax base without exemptions: If a taxcode is a TIF taxcode, then do EAV \* (1-%rev that goes to TIF). If the tax code is not a TIF tax code, then use the eav




```{r}
taxbills_current <- read_csv("./Output/1_Get_All_Pins-AllPinTaxbills_2021_Actual.csv")
# 22,453,875 tax bills in 2021 in municipalities. 
# DOES NOT INCLUDE unincorporated tax bills based on how we pulled the data in Step 1.

sapply(taxbills_current, function(x) sum(is.na(x)))


# 1,825,816 billed properties with 14-digit PINs  
pin14_bills_current <- taxbills_current %>%
  group_by(tax_code, class, pin) %>%
  
  mutate(total_bill = final_tax_to_dist + final_tax_to_tif) %>% # from each taxing agency
  
  summarize(total_billed = sum(total_bill, na.rm = TRUE), # total on someone's property tax bill
   av = first(av),
   eav = first(eav),
   taxing_agency_count = n(), # number of taxing agencies that tax the pin
   final_tax_to_dist = sum(final_tax_to_dist, na.rm = TRUE), # portion of all levies paid by the pin
   final_tax_to_tif = sum(final_tax_to_tif, na.rm = TRUE), 
   tax_amt_exe = sum(tax_amt_exe, na.rm = TRUE),           # revenue lost due to exemptions
   tax_amt_pre_exe = sum(tax_amt_pre_exe, na.rm = TRUE),   # total rev before all exemptions
   tax_amt_post_exe = sum(tax_amt_post_exe, na.rm = TRUE), # total rev after all exemptions
   rpm_tif_to_cps = sum(rpm_tif_to_cps, na.rm = TRUE),     # not used
   rpm_tif_to_rpm = sum(rpm_tif_to_rpm, na.rm=TRUE),       # not used
   rpm_tif_to_dist = sum(rpm_tif_to_dist, na.rm=TRUE),     # not used
   tif_share = mean(tif_share, na.rm=TRUE),                # not used
  )  %>% 
  mutate(propclass_1dig = str_sub(class, 1, 1))

head(pin14_bills_current)

sapply(pin14_bills_current, function(x) sum(is.na(x)))


write_csv(pin14_bills_current, "./Output/2_Summed_Bills_output-14digitPinSums.csv")
```

```{r}
table(taxbills_current$class)
```


```{r}
# 1,386,803 10-digit pins. Equivalent to the parcel or building footprint 
# values grouped like this are needed to map the parcels 

pin10_bills_current <- pin14_bills_current %>%
  mutate(pin10 = str_sub(pin, 1, 10)) %>%
  group_by(tax_code, class, pin10) %>%
  mutate(total_bill = final_tax_to_dist + final_tax_to_tif) %>%
  summarize(
    total_billed = sum(total_bill, na.rm = TRUE),
    av = sum(av), # changed from first() to sum() on 7/28/2023. Probably results in different output.......
    eav = sum(eav), # these would be parcel sums. All eav in the building or parcel for multi-unit buildings
    pin_count_in_parcel = n(),
    final_tax_to_dist = sum(final_tax_to_dist, na.rm = TRUE),
    final_tax_to_tif = sum(final_tax_to_tif, na.rm = TRUE),
    tax_amt_exe = sum(tax_amt_exe, na.rm = TRUE), # revenue lost due to exemptions
    tax_amt_pre_exe = sum(tax_amt_pre_exe, na.rm = TRUE), # total rev before all exemptions
    tax_amt_post_exe = sum(tax_amt_post_exe, na.rm = TRUE), # total rev after all exemptions
    rpm_tif_to_cps = sum(rpm_tif_to_cps, na.rm = TRUE), # not used
    rpm_tif_to_rpm = sum(rpm_tif_to_rpm, na.rm=TRUE), # not used
    rpm_tif_to_dist = sum(rpm_tif_to_dist, na.rm=TRUE), # not used
    tif_share = mean(tif_share, na.rm=TRUE))  %>% 
  mutate(propclass_2dig = str_sub(class, 1, 2)) 


head(pin10_bills_current)

pin10_bills_current <- pin10_bills_current %>%  
  mutate(tax_code = as.character(tax_code)) %>%
  full_join(muni_tax_codes, 
                      by = c("tax_code" = "tax_code_num")) %>%
  left_join(muni_agency_names) %>%
  left_join(nicknames) %>%

  mutate(parcel_bill_total = final_tax_to_tif + final_tax_to_dist)

write_csv(pin10_bills_current, "./Output/2_Summed_Bills_output-10digitParcelSums.csv")

head(pin10_bills_current)
tail(pin10_bills_current)


# numbe of missing values in each variable:
sapply(pin10_bills_current, function(x) sum(is.na(x)))

```


Now group again and sum values for tax code and each property class!

There are 28,381 property class - tax code group combinations. 

```{r}
taxcodes_by_class_current <- pin10_bills_current %>% 
  ungroup() %>%
  group_by(tax_code, class)  %>%
  summarize(
    av = sum(av), # changed from first() to sum() on 7/28/2023. Probably results in different output.......
    eav = sum(eav),
    pins_in_class = sum(pin_count_in_parcel),
    parcel_count_in_class = n(),
    final_tax_to_dist = sum(final_tax_to_dist, na.rm = TRUE),
    final_tax_to_tif = sum(final_tax_to_tif, na.rm = TRUE),
    tax_amt_exe = sum(tax_amt_exe, na.rm = TRUE), # revenue lost due to exemptions
    tax_amt_pre_exe = sum(tax_amt_pre_exe, na.rm = TRUE), # total rev before all exemptions
    tax_amt_post_exe = sum(tax_amt_post_exe, na.rm = TRUE), # total rev after all exemptions
    rpm_tif_to_cps = sum(rpm_tif_to_cps, na.rm = TRUE), # not used
    rpm_tif_to_rpm = sum(rpm_tif_to_rpm, na.rm=TRUE), # not used
    rpm_tif_to_dist = sum(rpm_tif_to_dist, na.rm=TRUE), # not used
    tif_share = mean(tif_share, na.rm=TRUE)) %>% 
  mutate(total_bill = final_tax_to_dist + final_tax_to_tif)

head(taxcodes_by_class_current)

taxcodes_by_class_current %>% write_csv("./Output/2_Summed_Bills_by_Taxcode_and_Class.csv")
```


```{r}
taxcodes_by_majorclass_current <- taxcodes_by_class_current %>% 
  ungroup() %>%
  left_join(class_dict, by = c("class" = "class_code")) %>%
  group_by(tax_code, major_class_code, major_class_type)  %>%
  
  summarize(
    av = sum(av), 
    eav = sum(eav),
    pins_in_majorclass = sum(pins_in_class),
    parcel_count_in_majorclass = sum(parcel_count_in_class),
    final_tax_to_dist = sum(final_tax_to_dist, na.rm = TRUE),
    final_tax_to_tif = sum(final_tax_to_tif, na.rm = TRUE),
    tax_amt_exe = sum(tax_amt_exe, na.rm = TRUE), # revenue lost due to exemptions
    tax_amt_pre_exe = sum(tax_amt_pre_exe, na.rm = TRUE), # total rev before all exemptions
    tax_amt_post_exe = sum(tax_amt_post_exe, na.rm = TRUE), # total rev after all exemptions
    rpm_tif_to_cps = sum(rpm_tif_to_cps, na.rm = TRUE), # not used
    rpm_tif_to_rpm = sum(rpm_tif_to_rpm, na.rm=TRUE), # not used
    rpm_tif_to_dist = sum(rpm_tif_to_dist, na.rm=TRUE), # not used
    tif_share = mean(tif_share, na.rm=TRUE),
  ) %>%
  
  mutate(total_bill = final_tax_to_dist + final_tax_to_tif)
head(taxcodes_by_majorclass_current)


taxcodes_by_majorclass_current %>% write_csv("./Output/2_Summed_Bills_by_Taxcode_and_MajorClass.csv")
```


# Composite Tax Rates

```{r eval=FALSE}
TC_bills_current <- read_csv("./Output/2_Summed_Bills_by_Taxcode_and_Class.csv") %>% 
  mutate(tax_code = as.character(tax_code),
         class = as.character(class))

class_dict$class_code <- as.character(class_dict$class_code)
 

taxcodes_current <- left_join(TC_bills_current, muni_tax_codes, 
                      by = c("tax_code" = "tax_code_num")) 


taxcodes_current %>% 
  left_join(muni_agency_names, by = "agency_num") %>%
  left_join(nicknames, by = c("agency_name" = "agency_name")) %>%
  #filter(!agency_num %in% cross_county_lines) %>%
 # group_by(clean_name, agency_name) %>%
  summarize(MuniLevy = sum(final_tax_to_dist, na.rm = TRUE), # amount billed by munis with current exemptions in place
            nonTIF_EAV_post_exemps = sum(final_tax_to_dist/(tax_code_rate/100), na.rm = TRUE),
            TIF_increment_EAV = sum(final_tax_to_tif/(tax_code_rate/100), na.rm=TRUE),  
            Exempt_EAV = sum(tax_amt_exe/(tax_code_rate/100), na.rm=TRUE), 
            Total_EAV = sum((tax_amt_exe+final_tax_to_dist+final_tax_to_tif)/(tax_code_rate/100), na.rm = TRUE)) %>%

  mutate(tax_rate_current = MuniLevy/nonTIF_EAV_post_exemps,
         nonTIF_EAV_pre_exemps = nonTIF_EAV_post_exemps + Exempt_EAV,
         taxrate_new = MuniLevy/nonTIF_EAV_pre_exemps,
         taxrate_change = tax_rate_current-taxrate_new) %>% 
 # select(clean_name, taxrate_change, tax_rate_current, taxrate_new, everything()) %>% 
  arrange(desc(tax_rate_current))


Current_Taxrates_perTC <- taxcodes_current %>% 
  left_join(muni_agency_names, by = "agency_num") %>%
  left_join(nicknames, by = c("agency_name" = "agency_name")) %>%
  #filter(!agency_num %in% cross_county_lines) %>%
  group_by(clean_name, agency_name, tax_code, pins_in_class) %>%
  summarize(
    MuniLevy = sum(final_tax_to_dist, na.rm = TRUE), # amount billed by munis with current exemptions in place
    nonTIF_EAV_post_exemps = sum(final_tax_to_dist/(tax_code_rate/100), na.rm = TRUE),
    TIF_increment_EAV = sum(final_tax_to_tif/(tax_code_rate/100), na.rm=TRUE),  
    Exempt_EAV = sum(tax_amt_exe/(tax_code_rate/100), na.rm=TRUE), 
    Total_EAV = sum((tax_amt_exe+final_tax_to_dist+final_tax_to_tif)/(tax_code_rate/100), na.rm = TRUE)) %>%
  
  mutate(tax_rate_current = MuniLevy/nonTIF_EAV_post_exemps,
         nonTIF_EAV_pre_exemps = nonTIF_EAV_post_exemps + Exempt_EAV,
         taxrate_new = MuniLevy/nonTIF_EAV_pre_exemps,
         taxrate_change = tax_rate_current-taxrate_new) %>% 
  select(clean_name, taxrate_change, tax_rate_current, taxrate_new, tax_code, everything()) %>% 
  arrange(desc(tax_rate_current))

Current_Taxrates_perTC

# write_csv(Current_Taxrates_perTC, "./Output/2_taxcode_taxrates.csv")


```
