==> old_versions/4b_Taxbill_Hypotheticals_SimilarProperties.Rmd <==
---
title: "4b_Similar Properties"
author: "Alea Wilbur"
date: "`r Sys.Date()`"
output:
  html_document:
    code_folding: hide
    code_download: yes
---

```{r setup, warning=FALSE, message=FALSE, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, warning = FALSE, message = FALSE)


library(tidyverse)
library(data.table)
library(gstat)
library(ptaxsim)
library(kableExtra)

```

# Finding Similar Properties and their Taxbills

```{r}
library(glue)
ptaxsim_db_conn <- DBI::dbConnect(RSQLite::SQLite(), "./ptaxsim.db/ptaxsim-2021.0.4.db")

# has EAV values, extensions by agency_num
agency_dt <- DBI::dbGetQuery(
  ptaxsim_db_conn,
  "SELECT *
  FROM agency
  WHERE year = 2021
  "
)

# grabs all unique muni names & numbs
# don't forget Cicero
muni_agency_names <- DBI::dbGetQuery(
  ptaxsim_db_conn,
  "SELECT DISTINCT agency_num, agency_name, minor_type
  FROM agency_info
  WHERE minor_type = 'MUNI'
  OR agency_num = '020060000'  

  "
)


# list of all taxcodes in municipalities. 
# This does NOT include unincorporated tax codes!!
doltchi_tax_codes <- DBI::dbGetQuery(
  ptaxsim_db_conn,
  glue_sql("
  SELECT*
  FROM tax_code
  WHERE (agency_num = '030210000' OR agency_num = '030310000')
  AND year = 2021
  ",
  .con = ptaxsim_db_conn
  )
) %>% mutate(tax_code = as.character(tax_code_num))

```

```{r eval = FALSE}
taxbills_current <- read_csv("./Output/1_Get_All_Pins-CookPinTaxbills_2021_Actual.csv")
# 22,453,875 tax bills in 2021 in municipalities (incorporated areas). 
# 22,972,902 tax bills in all of Cook County in 2021 (incorporated and unincorporated)
# DOES INCLUDE unincorporated tax bills based on how we pulled the data in Step 1.



# 1,825,816 billed properties with 14-digit PINs  for incorporated areras
# 1,864,594 billed properties in Cook county (incorporated and unincorporated)
# 
pin14_bills_current <- taxbills_current %>%
  group_by(tax_code, class, pin) %>%
  
  mutate(total_bill = final_tax_to_dist + final_tax_to_tif) %>% # from each taxing agency
  
  summarize(total_billed = sum(total_bill, na.rm = TRUE), # total on someone's property tax bill
            av = first(av),
            eav = first(eav),
            pin_count_in_parcel = n(),
            final_tax_to_dist = sum(final_tax_to_dist, na.rm = TRUE),
            final_tax_to_tif = sum(final_tax_to_tif, na.rm = TRUE),
            tax_amt_exe = sum(tax_amt_exe, na.rm = TRUE), # revenue lost due to exemptions
            tax_amt_pre_exe = sum(tax_amt_pre_exe, na.rm = TRUE), # total rev before all exemptions
            tax_amt_post_exe = sum(tax_amt_post_exe, na.rm = TRUE), # total rev after all exemptions
            rpm_tif_to_cps = sum(rpm_tif_to_cps, na.rm = TRUE), # not used
            rpm_tif_to_rpm = sum(rpm_tif_to_rpm, na.rm=TRUE), # not used
            rpm_tif_to_dist = sum(rpm_tif_to_dist, na.rm=TRUE), # not used
            tif_share = mean(tif_share, na.rm=TRUE), # not used
  )  %>% 
  mutate(propclass_1dig = str_sub(class, 1, 1))

head(pin14_bills_current)



DoltonChicago <- pin14_bills_current %>% 
  filter(tax_code %in% doltchi_tax_codes$tax_code_num) %>%
  filter(class != "0")


DoltonChicago <- DoltonChicago %>% 
  mutate(tax_code = as.character(tax_code))%>%
  left_join(doltchi_tax_codes) %>% 
  group_by(agency_num, class) %>% arrange(av)

write_csv(DoltonChicago, "./Output/Cholton_taxbills.csv")
```

```{r}
nicknames <- readxl::read_excel("./Necessary_Files/muni_shortnames.xlsx")

class_dict <- read_csv("./Necessary_Files/class_dict.csv")

taxcode_taxrates <- read_csv("./Output/2_taxcode_taxrates.csv")


DoltonChicago <- read_csv("./Output/Cholton_taxbills.csv") %>% 
  arrange(av) %>%     
  left_join(taxcode_taxrates) %>%
    mutate(propclass_1dig = str_sub(class, 1, 1))



```


> Max composite tax rate in a tax code in Chicago is 9.1% and minimum composite tax rate is 6.7%. Chicago has 665 tax codes. 

> Max composite tax rate in a tax code in Dolton is 27.9% and minimum composite tax rate is 22.9%. Dolton has 13 tax codes. 



```{r}
# DoltonChicago %>% 
#   group_by(agency_num, tax_code) %>% 
#   summarize(max_comprate = max(tax_rate_current),
#             min_comprate = min(tax_rate_current)) %>% arrange(-max_comprate)

DoltonChicago %>% group_by(agency_num) %>% 
  summarize(max_comprate = max(tax_rate_current, na.rm=TRUE),
            mean_comprate = mean(tax_rate_current, na.rm=TRUE),
            min_comprate = min(tax_rate_current, na.rm=TRUE)) %>% arrange(-mean_comprate)

DoltonChicago %>% filter(agency_num == "030310000") %>%
  group_by(agency_num, class) %>% 
  summarize(max_comprate = max(tax_rate_current, na.rm=TRUE),
            mean_comprate = mean(tax_rate_current, na.rm=TRUE),
            min_comprate = min(tax_rate_current, na.rm=TRUE),
            pin_count = n(),
            av = mean(av),
            eav = mean(eav),
            tax_amt_post_exe = mean(tax_amt_post_exe),
            tax_amt_exe = mean(tax_amt_exe),
            tax_amt_pre_exe = mean(tax_amt_pre_exe)) %>% 
  arrange(-pin_count) %>% head() %>% kbl()

DoltonChicago %>% filter(agency_num == "030210000") %>%
  group_by(agency_num, class) %>% 
  summarize(max_comprate = max(tax_rate_current, na.rm=TRUE),
            mean_comprate = mean(tax_rate_current, na.rm=TRUE),
            min_comprate = min(tax_rate_current, na.rm=TRUE),
            pin_count = n(),
            av = mean(av),
            eav = mean(eav),
            tax_amt_post_exe = mean(tax_amt_post_exe),
            tax_amt_exe = mean(tax_amt_exe),
            tax_amt_pre_exe = mean(tax_amt_pre_exe)) %>% 
  arrange(-pin_count) %>% head() %>% kbl()
``` 



```{r}

DoltonChicago %>% 
  filter(class == "203") %>% 
  arrange(av)%>%
  group_by(agency_num, class) %>%
  summarize(medianbill = median(total_billed),
            meanbill = mean(total_billed),
            medianAV = median(av),
            meanAV = mean(av),
            )%>% 
 # pivot_longer(medianbill:meanAV, names_to = "Stats", values_to = "Values")  %>%
  kbl(caption= "Chicago and Dolton, Class 203, Measures of the Middle", digits=0, booktabs = T) %>%   
  kable_styling(full_width = T)



DoltonChicago %>% 
  filter(class == "205") %>% 
  group_by(agency_num, class) %>%
  summarize(medianbill = median(total_billed),
            meanbill = mean(total_billed),
            medianAV = median(av),
            meanAV = mean(av),
            )%>% 
 # pivot_longer(medianbill:meanAV, names_to = "Stats", values_to = "Values")  %>%
  kbl(caption= "Chicago and Dolton, Class 205, Measures of the Middle", digits=0, booktabs = T) %>%   
  kable_styling(full_width = T)

DoltonChicago %>% 
  filter(class == "211") %>% 
  group_by(agency_num, class) %>%
  summarize(medianbill = median(total_billed),
            meanbill = mean(total_billed),
            medianAV = median(av),
            meanAV = mean(av),
            )%>% 
 # pivot_longer(medianbill:meanAV, names_to = "Stats", values_to = "Values")  %>%
  kbl(caption= "Chicago and Dolton, Class 211, Measures of the Middle", digits=0, booktabs = T) %>%   
  kable_styling(full_width = T)

DoltonChicago %>% 
  filter(class == "234") %>% 
  group_by(agency_num, class) %>%
  summarize(medianbill = median(total_billed),
            meanbill = mean(total_billed),
            medianAV = median(av),
            meanAV = mean(av),
            )%>% 
 # pivot_longer(medianbill:meanAV, names_to = "Stats", values_to = "Values")  %>%
  kbl(caption= "Chicago and Dolton, Class 234, Measures of the Middle", digits=0, booktabs = T) %>%   
  kable_styling(full_width = T)


```

## Using similarily Assessed Value properties 

Chicago has 655 tax codes in its borders and Dolton has 13 tax codes in its borders. 



### Class 203

One story residence, any age, 1,000 to 1,800 sq. ft.


```{r}
DoltonChicago %>% 
  filter(agency_num == "030210000" ) %>% #& class == "203" & between(av,9950,10050)) %>% 
  group_by(tax_code) %>% 
  summarize(count = n(),            
            avg_current_comprate = mean(tax_rate_current, na.rm=TRUE)
  ) %>% arrange(-avg_current_comprate) %>% head()

DoltonChicago %>% 
  filter(agency_num == "030310000") %>% # & class == "203" & between(av, 9950, 10050)) %>%
  group_by(tax_code) %>%
  summarize(count = n(),
            avg_current_comprate = mean(tax_rate_current, na.rm=TRUE)
  ) %>% arrange(-avg_current_comprate) %>% head()
```


If holding the levy constant and acknowledging the change in tax rates that would occur from having additional taxable EAV within the taxing jurisdictions....

For property class 203 PINs with assessed values between \$9,000 and \$11,000, the average change in tax bill would be \$75 more in Chicago and \$205 more in Dolton compared to their current tax bills (\$1475 and $4312 respectively). 

The "average" property tax payer would think they are saving \$624 in Chicago and $2708 in Dolton due to exemptions. This number appears on their taxbill and calculated by the full EAV * current tax rate and does NOT consider the change in tax rate that would occur if levies are held constant and all EAV became taxable. 


```{r}
Chi3 <- DoltonChicago %>% 
  filter(agency_num == "030210000" & class == "203" & between(av,9000,11000)) %>%
    summarize(              
    comp_taxrate = mean(tax_code_rate, na.rm=TRUE),
    bill_current = mean(tax_amt_post_exe, na.rm=TRUE),
    bill_hyp = mean(eav*taxrate_new, na.rm=TRUE),
    bill_change = bill_hyp - bill_current,
    tax_amt_post_exe = mean(tax_amt_post_exe),
    tax_amt_exe = mean(tax_amt_exe),
    tax_amt_pre_exe = mean(tax_amt_pre_exe),
    pin_count = n(),
    av = mean(av), 
    eav = mean(eav)
  ) %>% 
  pivot_longer(cols = comp_taxrate:eav, names_to = "Stats", values_to = "Values")

Dol3 <- DoltonChicago %>% 
  filter(agency_num == "030310000" & class == "203" & between(av, 9000, 11000)) %>%
    summarize(              
    comp_taxrate = mean(tax_code_rate, na.rm=TRUE),
    bill_current = mean(tax_amt_post_exe, na.rm=TRUE),
    bill_hyp = mean(eav*taxrate_new, na.rm=TRUE),
    bill_change = bill_hyp - bill_current,
    tax_amt_post_exe = mean(tax_amt_post_exe),
    tax_amt_exe = mean(tax_amt_exe, na.rm=TRUE),
    tax_amt_pre_exe = mean(tax_amt_pre_exe, na.rm=TRUE),
    pin_count = n(),
    av = mean(av), 
    eav = mean(eav)
  ) %>% 
    pivot_longer(cols = comp_taxrate:eav, names_to = "Stats", values_to = "Values")

both_dt <- cbind(Chi3, Dol3)


kbl(both_dt, booktabs = T, digits = 0, 
    caption = "Property Class 203 Comparison, AV ~ $10,000 (9000-11000 range)") %>%
  kable_styling(full_width = T)%>%
add_header_above(c("Chicago" = 2, "Dolton" = 2))
```


> Changing the range of PINs included in the calculation alters the "Median Property Statistic" 

If holding the levy constant and acknowledging the change in tax rates that would occur from having additional taxable EAV within the taxing jurisdictions....

For property class 203 PINs with assessed values between \$8,000 and \$12,000, the average change in tax bill would be \$98 more in Chicago and \$55 more in Dolton compared to their current tax bills (\$1513 and $4342 respectively). 

```{r}
Chi3 <- DoltonChicago %>% 
  filter(agency_num == "030210000" & class == "203" & between(av,8000,12000)) %>%
    summarize(              
    comp_taxrate = mean(tax_code_rate, na.rm=TRUE),
    bill_current = mean(tax_amt_post_exe, na.rm=TRUE),
    bill_hyp = mean(eav*taxrate_new, na.rm=TRUE),
    bill_change = bill_hyp - bill_current,
    tax_amt_post_exe = mean(tax_amt_post_exe),
    tax_amt_exe = mean(tax_amt_exe, na.rm=TRUE),
    tax_amt_pre_exe = mean(tax_amt_pre_exe, na.rm=TRUE),
    pin_count = n(),
    av = mean(av), 
    eav = mean(eav)
  ) %>% 
  pivot_longer(cols = comp_taxrate:eav, names_to = "Stats", values_to = "Values")

# Chicago only has 2 pins that are similar to Dolton's median pin (which had a lot of matches)
Dol3 <- DoltonChicago %>% 
  filter(agency_num == "030310000" & class == "203" & between(av, 8000, 12000)) %>%
    summarize(              
    comp_taxrate = mean(tax_code_rate, na.rm=TRUE),
    bill_current = mean(tax_amt_post_exe, na.rm=TRUE),
    bill_hyp = mean(eav*taxrate_new, na.rm=TRUE),
    bill_change = bill_hyp - bill_current,
    tax_amt_post_exe = mean(tax_amt_post_exe),
    tax_amt_exe = mean(tax_amt_exe, na.rm=TRUE),
    tax_amt_pre_exe = mean(tax_amt_pre_exe, na.rm=TRUE),
    pin_count = n(),
    av = mean(av), 
    eav = mean(eav)
  ) %>% 
    pivot_longer(cols = comp_taxrate:eav, names_to = "Stats", values_to = "Values")

both_dt <- cbind(Chi3, Dol3)


kbl(both_dt, booktabs = T, digits = 0, 
    caption = "Property Major Class 2 Comparison, AV ~ $10,000 (AV range $8000-$12000)") %>%
  kable_styling(full_width = T)%>%
add_header_above(c("Chicago Class 203, 8K-12K Property Stats" = 2, "Dolton Class 203, 8L-12K AV Property Stats" = 2))
```



### Major Class 2

Average and median tax bills and assessed values are calculated below for ALL property class types within the the broader "Residential" property class type (property classes that have the first digit "2", or Major Class Type 2)

The median AV  is used to select a range of pins (based on their AV) to calculate the average current bill, hypothetical bill, and hypothetical change in tax bill for a "median property". 

This is done because some properties receive multiple exemptions while others receive none. Using the literal median pin can skew the summary statistics of that specific pin receives no exemptions or multiple exemptions. The average for the range of "median PINs" is created to smooth out the variation within the observations. 

```{r}
DoltonChicago %>% 
  filter(propclass_1dig == "2") %>% 
  arrange(av) %>%
  group_by(agency_num) %>%
  summarize(medianbill = median(total_billed),
            meanbill = mean(total_billed),
            medianAV = median(av),
            meanAV = mean(av),
            pin_count = n()
            )%>% 
 # pivot_longer(medianbill:meanAV, names_to = "Stats", values_to = "Values")  %>%
  kbl(caption= "Chicago and Dolton, Major Class 2, Measures of the Middle", digits=0, booktabs = T) %>%   
  kable_styling(full_width = T)

# 
# DoltonChicago %>% 
#   filter(propclass_1dig == "3") %>% 
#   arrange(av) %>%
#   group_by(agency_num) %>%
#   summarize(medianbill = median(total_billed),
#             meanbill = mean(total_billed),
#             medianAV = median(av),
#             meanAV = mean(av),
#             pin_count = n()
#             )%>% 
#  # pivot_longer(medianbill:meanAV, names_to = "Stats", values_to = "Values")  %>%
#   kbl(caption= "Chicago and Dolton, Major Class 3, Measures of the Middle", digits=0, booktabs = T) %>%   
#   kable_styling(full_width = T)
```

If holding the levy constant and acknowledging the change in tax rates that would occur from having additional taxable EAV within the taxing jurisdictions, Chicago's hypothetical tax bill would be \$1570 (an \$14 increase from \$1556) and Dolton's would be \$4610 (a \$30 increase from $4580).   

On average, residents would think they "saved" \$501 in Chicago and \$2588 in Dolton (based on tax_amt_exe which also shows up on their tax bill based on the "naive" pre-tax exemption tax bill amount on the tax bill). The amount saved per person will depend on how many exemptions they qualified for in the first place. This is a rough average for all types of exemptions and includes those that received no exemptions and multiiple exemptions. 

Values were calculated by selecting pins with AVs between \$9000 and $11,000 and then calculating the average current bill, change in bill, and other statistics seen in the table. Pin count tells you the number of pins that were included in the AV range used for the "median property."

```{r}
Chi3 <- DoltonChicago %>% 
  filter(agency_num == "030210000" & propclass_1dig == "2" & between(av,9000,11000)) %>%
    summarize(              
    comp_taxrate = mean(tax_code_rate, na.rm=TRUE),
    bill_current = mean(tax_amt_post_exe, na.rm=TRUE),
    bill_hyp = mean(eav*taxrate_new, na.rm=TRUE),
    bill_change = bill_hyp - bill_current,
    tax_amt_post_exe = mean(tax_amt_post_exe),
    tax_amt_exe = mean(tax_amt_exe, na.rm=TRUE),
    tax_amt_pre_exe = mean(tax_amt_pre_exe, na.rm=TRUE),
    pin_count = n(),
    av = mean(av), 
    eav = mean(eav)
  ) %>% 
  pivot_longer(cols = comp_taxrate:eav, names_to = "Stats", values_to = "Values")

# Chicago only has 2 pins that are similar to Dolton's median pin (which had a lot of matches)
Dol3 <- DoltonChicago %>% 
  filter(agency_num == "030310000" & propclass_1dig == "2" & between(av, 9000, 11000)) %>%
    summarize(              
    comp_taxrate = mean(tax_code_rate, na.rm=TRUE),
    bill_current = mean(tax_amt_post_exe, na.rm=TRUE),
    bill_hyp = mean(eav*taxrate_new, na.rm=TRUE),
    bill_change = bill_hyp - bill_current,
    tax_amt_post_exe = mean(tax_amt_post_exe),
    tax_amt_exe = mean(tax_amt_exe, na.rm=TRUE),
    tax_amt_pre_exe = mean(tax_amt_pre_exe, na.rm=TRUE),
    pin_count = n(),
    av = mean(av), 
    eav = mean(eav)
  ) %>% 
    pivot_longer(cols = comp_taxrate:eav, names_to = "Stats", values_to = "Values")

both_dt <- cbind(Chi3, Dol3)


kbl(both_dt, booktabs = T, digits = 0, 
    caption = "Property Major Class 2 Comparison, AV ~ $10,000 (AV range $9000-$11000)") %>%
  kable_styling(full_width = T)%>%
add_header_above(c("Chicago" = 2, "Dolton" = 2))
```


```{r}
Chi3 <- DoltonChicago %>% 
  filter(agency_num == "030210000" & propclass_1dig == "2" & between(av,8000,12000)) %>%
    summarize(              
    comp_taxrate = mean(tax_code_rate, na.rm=TRUE),
    bill_current = mean(tax_amt_post_exe, na.rm=TRUE),
    bill_hyp = mean(eav*taxrate_new, na.rm=TRUE),
    bill_change = bill_hyp - bill_current,
    tax_amt_post_exe = mean(tax_amt_post_exe),
    tax_amt_exe = mean(tax_amt_exe, na.rm=TRUE),
    tax_amt_pre_exe = mean(tax_amt_pre_exe, na.rm=TRUE),
    pin_count = n(),
    av = mean(av), 
    eav = mean(eav)
  ) %>% 
  pivot_longer(cols = comp_taxrate:eav, names_to = "Stats", values_to = "Values")

# Chicago only has 2 pins that are similar to Dolton's median pin (which had a lot of matches)
Dol3 <- DoltonChicago %>% 
  filter(agency_num == "030310000" & propclass_1dig == "2" & between(av, 8000, 12000)) %>%
    summarize(              
    comp_taxrate = mean(tax_code_rate, na.rm=TRUE),
    bill_current = mean(tax_amt_post_exe, na.rm=TRUE),
    bill_hyp = mean(eav*taxrate_new, na.rm=TRUE),
    bill_change = bill_hyp - bill_current,
    tax_amt_post_exe = mean(tax_amt_post_exe),
    tax_amt_exe = mean(tax_amt_exe, na.rm=TRUE),
    tax_amt_pre_exe = mean(tax_amt_pre_exe, na.rm=TRUE),
    pin_count = n(),
    av = mean(av), 
    eav = mean(eav)
  ) %>% 
    pivot_longer(cols = comp_taxrate:eav, names_to = "Stats", values_to = "Values")

both_dt <- cbind(Chi3, Dol3)


kbl(both_dt, booktabs = T, digits = 0, 
    caption = "Property Major Class 2 Comparison, AV ~ $10,000 (AV range $8000-$12000)") %>%
  kable_styling(full_width = T)%>%
add_header_above(c("Chicago" = 2, "Dolton" = 2))
```
If holding the levy constant and acknowledging the change in tax rates that would occur from having additional taxable EAV within the taxing jurisdictions, Chicago's hypothetical tax bill would be \$1592 (an \$13 increase from \$1579) and Dolton's would be \$4598 (a \$99 DECREASE from $4596).   

Values were calculated by selecting pins with AVs between \$8000 and $12,000 and then calculating the average current bill, change in bill, and other statistics seen in the table. Pin count tells you the number of pins that were included in the AV range used for the "median property."


> Increasing the number of pins included in the measurement of "average bill" and "average bill change" completely changed the results for Dolton. Using the "median" value must be done super carefully. 

> If we "removed" only homeowners exemptions or only senior exemptions, then the median statistic would be more reliable... potentially. 



### Class 205

Two or more story residence, over 62 years, up to 2,200 sq. ft


```{r}
Chi3 <- DoltonChicago %>% 
  filter(agency_num == "030210000" & class == "205" & between(av,8000,12000)) %>%
    summarize(              
    comp_taxrate = mean(tax_code_rate, na.rm=TRUE),
    bill_current = mean(tax_amt_post_exe, na.rm=TRUE),
    bill_hyp = mean(eav*taxrate_new, na.rm=TRUE),
    bill_change = bill_hyp - bill_current,
    tax_amt_post_exe = mean(tax_amt_post_exe),
    tax_amt_exe = mean(tax_amt_exe, na.rm=TRUE),
    tax_amt_pre_exe = mean(tax_amt_pre_exe, na.rm=TRUE),
    pin_count = n(),
    av = mean(av), 
    eav = mean(eav)
  ) %>% 
  pivot_longer(cols = comp_taxrate:eav, names_to = "Stats", values_to = "Values")


Dol3 <- DoltonChicago %>% 
  filter(agency_num == "030310000" & class == "205" & between(av, 8000, 12000)) %>%
    summarize(              
    comp_taxrate = mean(tax_code_rate, na.rm=TRUE),
    bill_current = mean(tax_amt_post_exe, na.rm=TRUE),
    bill_hyp = mean(eav*taxrate_new, na.rm=TRUE),
    bill_change = bill_hyp - bill_current,
    tax_amt_post_exe = mean(tax_amt_post_exe),
    tax_amt_exe = mean(tax_amt_exe, na.rm=TRUE),
    tax_amt_pre_exe = mean(tax_amt_pre_exe, na.rm=TRUE),
    pin_count = n(),
    av = mean(av), 
    eav = mean(eav)
  )  %>% 
    pivot_longer(cols = comp_taxrate:eav, names_to = "Stats", values_to = "Values")

both_dt <- cbind(Chi3, Dol3)


kbl(both_dt, booktabs = T, digits = 0, 
    caption = "Property Class 205 Comparison, AV ~ $10,000 (8000 to 12000 AV range)") %>%
  kable_styling(full_width = T)%>%
add_header_above(c("Chicago" = 2, "Dolton" = 2))
```


### Class 211

Two to six residential apartments, any age.


```{r}
Chi3 <- DoltonChicago %>% 
  filter(agency_num == "030210000" & class == "211" & between(av,17000,19000)) %>%
    summarize(              
    comp_taxrate = mean(tax_code_rate, na.rm=TRUE),
    bill_current = mean(tax_amt_post_exe, na.rm=TRUE),
    bill_hyp = mean(eav*taxrate_new, na.rm=TRUE),
    bill_change = bill_hyp - bill_current,
    tax_amt_post_exe = mean(tax_amt_post_exe),
    tax_amt_exe = mean(tax_amt_exe, na.rm=TRUE),
    tax_amt_pre_exe = mean(tax_amt_pre_exe, na.rm=TRUE),
    pin_count = n(),
    av = mean(av), 
    eav = mean(eav)
  ) %>% 
  pivot_longer(cols = comp_taxrate:eav, names_to = "Stats", values_to = "Values")


Dol3 <- DoltonChicago %>% 
  filter(agency_num == "030310000" & class == "211" & between(av, 17000, 19000)) %>%
    summarize(              
    comp_taxrate = mean(tax_code_rate, na.rm=TRUE),
    bill_current = mean(tax_amt_post_exe, na.rm=TRUE),
    bill_hyp = mean(eav*taxrate_new, na.rm=TRUE),
    bill_change = bill_hyp - bill_current,
    tax_amt_post_exe = mean(tax_amt_post_exe),
    tax_amt_exe = mean(tax_amt_exe, na.rm=TRUE),
    tax_amt_pre_exe = mean(tax_amt_pre_exe, na.rm=TRUE),
    pin_count = n(),
    av = mean(av), 
    eav = mean(eav)
  ) %>% 
    pivot_longer(cols = comp_taxrate:eav, names_to = "Stats", values_to = "Values")

both_dt <- cbind(Chi3, Dol3)


kbl(both_dt, booktabs = T, digits = 0, 
    caption = "Property Class 211 Comparison, AV ~ $18,000") %>%
  kable_styling(full_width = T)%>%
add_header_above(c("Chicago" = 2, "Dolton" = 2))
```



### Class 234

Split level residence, with a lower level below grade, all ages, all sizes


```{r}
Chi3 <- DoltonChicago %>% 
  filter(agency_num == "030210000" & class == "234" & between(av,12000,13000)) %>%
    summarize(              
    comp_taxrate = mean(tax_code_rate, na.rm=TRUE),
    bill_current = mean(tax_amt_post_exe, na.rm=TRUE),
    bill_hyp = mean(eav*taxrate_new, na.rm=TRUE),
    bill_change = bill_hyp - bill_current,
    tax_amt_post_exe = mean(tax_amt_post_exe),
    tax_amt_exe = mean(tax_amt_exe, na.rm=TRUE),
    tax_amt_pre_exe = mean(tax_amt_pre_exe, na.rm=TRUE),
    pin_count = n(),
    av = mean(av), 
    eav = mean(eav)
  ) %>% 
  pivot_longer(cols = comp_taxrate:eav, names_to = "Stats", values_to = "Values")


Dol3 <- DoltonChicago %>% 
  filter(agency_num == "030310000" & class == "234" & between(av, 12450, 12750)) %>%
    summarize(              
    comp_taxrate = mean(tax_code_rate, na.rm=TRUE),
    bill_current = mean(tax_amt_post_exe, na.rm=TRUE),
    bill_hyp = mean(eav*taxrate_new, na.rm=TRUE),
    bill_change = bill_hyp - bill_current,
    tax_amt_post_exe = mean(tax_amt_post_exe),
    tax_amt_exe = mean(tax_amt_exe, na.rm=TRUE),
    tax_amt_pre_exe = mean(tax_amt_pre_exe, na.rm=TRUE),
    pin_count = n(),
    av = mean(av), 
    eav = mean(eav)) %>% 
    pivot_longer(cols = comp_taxrate:eav, names_to = "Stats", values_to = "Values")

both_dt <- cbind(Chi3, Dol3)


kbl(both_dt, booktabs = T, digits = 0, 
    caption = "Property Class 234 Comparison, AV ~ $12,500") %>%
  kable_styling(full_width = T)%>%
add_header_above(c("Chicago" = 2, "Dolton" = 2))
```


# Other ways to compare locations (that I don't like as much)

## Median Chicago AV vs Median Dolton AV

```{r}


## Chicago #

Chi1 <- DoltonChicago %>% 
  arrange(av)%>%
  filter(agency_num == "030210000" & class == "203" & between(av,18000,22000)) %>%
  summarize(
    comp_taxrate = mean(tax_code_rate),
    hypothetical_taxrate = mean(taxrate_new, na.rm=TRUE),
    bill_current = mean(tax_amt_post_exe),
    bill_hyp = mean(eav*taxrate_new, na.rm=TRUE),
    bill_change = bill_hyp - bill_current,
    tax_amt_post_exe = mean(tax_amt_post_exe),
    tax_amt_exe = mean(tax_amt_exe),
    tax_amt_pre_exe = mean(tax_amt_pre_exe),
    pin_count = n(),
    av = mean(av), 
    eav = mean(eav)
  ) %>% 
  pivot_longer(cols = comp_taxrate:eav, names_to = "Stats", values_to = "Values")
  # kbl(caption= "Chicago, Class 203, Tax Bill ~$3150", digits=0, booktabs = T)


## Dolton ## 

Dol1<- DoltonChicago %>% 
  arrange(av) %>%
  filter(agency_num == "030310000" & class == "203" & between(av,8000,12000)) %>%
  summarize(
    comp_taxrate = mean(tax_code_rate),
    hypothetical_taxrate = mean(taxrate_new, na.rm=TRUE),

    bill_current = mean(tax_amt_post_exe),
    bill_hyp = mean(eav*taxrate_new, na.rm=TRUE),
    bill_change = bill_hyp - bill_current,
    tax_amt_post_exe = mean(tax_amt_post_exe),
    tax_amt_exe = mean(tax_amt_exe),
    tax_amt_pre_exe = mean(tax_amt_pre_exe),
    pin_count = n(),
    av = mean(av), 
    eav = mean(eav)
  ) %>% 
  pivot_longer(cols = comp_taxrate:eav, names_to = "Stats", values_to = "Values")


both_dt <- cbind(Chi1, Dol1)


kbl(both_dt, booktabs = T, digits = 0, caption = "Property Class 203 Comparison, Chicago Median AV vs Dolton Median AV") %>%
  kable_styling(full_width = T) %>%
add_header_above(c("Chicago Median AV 203 Property" = 2, "Dolton Median AV 203 Property" = 2))


```


## Median Chicago Taxbill vs Median Dolton Taxbill


```{r}


## Chicago #

Chi1 <- DoltonChicago %>% 
  arrange(av)%>%
  filter(agency_num == "030210000" & class == "203" & between(total_billed,3100,3300)) %>%
  summarize(
    comp_taxrate = mean(tax_code_rate),
    bill_current = mean(tax_amt_post_exe),
    bill_hyp = mean(eav*taxrate_new, na.rm=TRUE),
    bill_change = bill_hyp - bill_current,
    tax_amt_post_exe = mean(tax_amt_post_exe),
    tax_amt_exe = mean(tax_amt_exe),
    tax_amt_pre_exe = mean(tax_amt_pre_exe),
    pin_count = n(),
    av = mean(av), 
    eav = mean(eav)
  ) %>% 
  pivot_longer(cols = comp_taxrate:eav, names_to = "Stats", values_to = "Values")
  # kbl(caption= "Chicago, Class 203, Tax Bill ~$3150", digits=0, booktabs = T)


## Dolton ## 

Dol1<- DoltonChicago %>% 
  arrange(av) %>%
  filter(agency_num == "030310000" & class == "203" & between(total_billed,3100,3300)) %>%
  summarize(
    comp_taxrate = mean(tax_code_rate),
    bill_current = mean(tax_amt_post_exe),
    bill_hyp = mean(eav*taxrate_new, na.rm=TRUE),
    bill_change = bill_hyp - bill_current,
    tax_amt_post_exe = mean(tax_amt_post_exe),
    tax_amt_exe = mean(tax_amt_exe),
    tax_amt_pre_exe = mean(tax_amt_pre_exe),
    pin_count = n(),
    av = mean(av), 
    eav = mean(eav)
  ) %>% 
  pivot_longer(cols = comp_taxrate:eav, names_to = "Stats", values_to = "Values") #%>% 
  # kbl(caption= "Dolton, Class 203, Tax Bill ~ $3150", digits=0, booktabs = T)


# av = 
#exempt eav = 


both_dt <- cbind(Chi1, Dol1)


kbl(both_dt, booktabs = T, digits = 0, caption = "Property Class 203 Comparison, Tax Bill ~ $3,150 (Chicago's median current bill)") %>%
  kable_styling(full_width = T)%>%
add_header_above(c("Chicago" = 2, "Dolton" = 2))




Chi1a <- DoltonChicago %>% filter(agency_num == "030210000" & class == "203" & between(total_billed,4100,4200)) %>%
  arrange(av)%>%
  summarize(
    comp_taxrate = mean(tax_code_rate),
    bill_current = mean(tax_amt_post_exe),
    bill_hyp = mean(eav*taxrate_new, na.rm=TRUE),
    bill_change = bill_hyp - bill_current,
    tax_amt_post_exe = mean(tax_amt_post_exe),
    tax_amt_exe = mean(tax_amt_exe),
    tax_amt_pre_exe = mean(tax_amt_pre_exe),
    pin_count = n(),
    av = mean(av), 
    eav = mean(eav)
                ) %>% 
  pivot_longer(comp_taxrate:eav, names_to = "Stats", values_to = "Values") 
# kbl(caption= "Chicago, Class 203, Tax Bill ~$3150", digits=0, booktabs = T)


## Chicago ## 

# Chicago only has 2 pins that are similar to Dolton's median pin (which had a lot of matches)
#DoltonChicago %>% filter(agency_num == "030310000" & class == "203" & between(total_billed,3140,3160))

Dol1a <- DoltonChicago %>% filter(agency_num == "030310000" & class == "203" & between(total_billed,4100,4200)) %>%
  summarize(
    comp_taxrate = mean(tax_code_rate),
    bill_current = mean(tax_amt_post_exe),
    bill_hyp = mean(eav*taxrate_new, na.rm=TRUE),
    bill_change = bill_hyp - bill_current,
    tax_amt_post_exe = mean(tax_amt_post_exe),
    tax_amt_exe = mean(tax_amt_exe),
    tax_amt_pre_exe = mean(tax_amt_pre_exe),
    pin_count = n(),
    av = mean(av), 
    eav = mean(eav)) %>% 
  pivot_longer(comp_taxrate:eav, names_to = "Stats", values_to = "Values") #%>% 
# kbl(caption= "Dolton, Class 203, Tax Bill ~ $3150", digits=0, booktabs = T)


# av = 
#exempt eav = 


both_dt <- cbind(Chi1a, Dol1a)


kbl(both_dt, booktabs = T, digits = 0, caption = "Property Class 203 Comparison, Tax Bill ~ $4150 (Dolton's median current bill)") %>%
  kable_styling(full_width = T) %>%
add_header_above(c("Chicago" = 2, "Dolton" = 2))


both_dt <- cbind(Chi1, Dol1a)


kbl(both_dt, booktabs = T, digits = 0, caption = "Property Class 203 Comparison, Median Dolton vs Median Chicago") %>%
  kable_styling(full_width = T)%>%
add_header_above(c("Chicago Median 203 Property" = 2, "Dolton Median 203 Property" = 2))
```


## Using a current tax bill ~$3500:

If holding the levy constant and acknowledging the change in tax rates that would occur from having additional taxable EAV within the taxing jurisdictions, Chicago's hypothetical tax bill would be \$3644 (an \$87 increase from \$3557) and Dolton's would be \$3692 (a \$136 increase from $3555).

```{r}
DoltonChicago %>% 
  filter(agency_num == "030210000" & class == "203" & between(tax_amt_post_exe, 3500, 3600)) %>%
    summarize(              
    comp_taxrate = mean(tax_code_rate),
    bill_current = mean(tax_amt_post_exe),
    bill_hyp = mean(eav*taxrate_new, na.rm=TRUE),
    bill_change = bill_hyp - bill_current,
    tax_amt_post_exe = mean(tax_amt_post_exe),
    tax_amt_exe = mean(tax_amt_exe),
    tax_amt_pre_exe = mean(tax_amt_pre_exe),
    pin_count = n(),
    av = mean(av), 
    eav = mean(eav))  %>% 
    pivot_longer(cols = comp_taxrate:eav, names_to = "Stats", values_to = "Values")%>%
  kbl(caption= "Chicago, Class 203, Current Bill ~ $3,550", digits=0)

# Chicago only has 2 pins that are similar to Dolton's median pin (which had a lot of matches)
DoltonChicago %>% 
  filter(agency_num == "030310000" & class == "203" & between(tax_amt_post_exe,3500,3600)) %>%
    summarize(              
    comp_taxrate = mean(tax_code_rate),
    bill_current = mean(tax_amt_post_exe),
    bill_hyp = mean(eav*taxrate_new, na.rm=TRUE),
    bill_change = bill_hyp - bill_current,
    tax_amt_post_exe = mean(tax_amt_post_exe),
    tax_amt_exe = mean(tax_amt_exe),
    tax_amt_pre_exe = mean(tax_amt_pre_exe),
    pin_count = n(),
    av = mean(av), 
    eav = mean(eav)) %>% 
    pivot_longer(cols = comp_taxrate:eav, names_to = "Stats", values_to = "Values")%>%
  kbl(caption= "Dolton, Class 203, Current Bill ~ $3,550", digits=0)
```





## Using a current tax bill ~$3700:

Value was chosen randomly as roughly between the median property AV in Dolton and Chicago. 

For post-exemption bills around \$3700: Chicago has $26,448 AV and \$79,420 in EAV. Chicago's Composite tax rate for the properties examined is 6.7%. Dolton has an AV of \$7,631 and \$22,914 in EAV. Dolton's average composite tax rate for the properties examined was 24.2%.

If holding the levy constant and acknowledging the change in tax rates that would occur from having additional taxable EAV within the taxing jurisdictions, Chicago's hypothetical tax bill would be \$4503 (an \$802 increase from \$3700) and Dolton's would be \$3479 (a \$220 DECREASE from $3700). 

```{r}
## Chicago ## 
DoltonChicago %>% 
  filter(agency_num == "030210000" & class == "203" & between(tax_amt_post_exe, 3695, 3705)) %>%
  summarize(
    comp_taxrate = mean(tax_code_rate),
    bill_current = mean(tax_amt_post_exe),
    bill_hyp = mean(eav*taxrate_new, na.rm=TRUE),
    bill_change = bill_hyp - bill_current,

    tax_amt_exe = mean(tax_amt_exe),
    tax_amt_pre_exe = mean(tax_amt_pre_exe),
    pin_count = n(),
    av = mean(av),
    eav = mean(eav)
  )  %>% 
    pivot_longer(cols = comp_taxrate:eav, names_to = "Stats", values_to = "Values")%>%
  kbl(caption= "Chicago NOT ENOGUH PINS, Class 203, Current Bill ~$3,700", digits=0)

# Chicago only has 2 pins that are similar to Dolton's median pin (which had a lot of matches)

## Dolton ## 
DoltonChicago %>% 
  filter(agency_num == "030310000" & class == "203" & between(tax_amt_post_exe,3695,3705)) %>%
  summarize(
    comp_taxrate = mean(tax_code_rate),
    bill_current = mean(tax_amt_post_exe),
    bill_hyp = mean(eav*taxrate_new, na.rm=TRUE),  
    bill_change = bill_hyp - bill_current,
    tax_amt_pre_exe = mean(tax_amt_pre_exe), # the naive amount that appears on peoples taxbills
    pin_count = n(),
    av = mean(av),
    eav = mean(eav)
  ) %>% 
    pivot_longer(cols = comp_taxrate:eav, names_to = "Stats", values_to = "Values")%>%
  kbl(caption= "Dolton NOT ENOUGH PINS, Class 203, Current Bill ~ $3,700" , digits=0)
```



> Expanded taxbill range:


```{r}
## Chicago ## 
DoltonChicago %>% 
  filter(agency_num == "030210000" & class == "203" & between(tax_amt_post_exe, 3600, 3800)) %>%
  summarize(
    comp_taxrate = mean(tax_code_rate),
    bill_current = mean(tax_amt_post_exe),
    bill_hyp = mean(eav*taxrate_new, na.rm=TRUE),
    bill_change = bill_hyp - bill_current,

    tax_amt_exe = mean(tax_amt_exe),
    tax_amt_pre_exe = mean(tax_amt_pre_exe),
    pin_count = n(),
    av = mean(av),
    eav = mean(eav)
  )  %>% 
    pivot_longer(cols = comp_taxrate:eav, names_to = "Stats", values_to = "Values")%>%
  kbl(caption= "Chicago, Class 203, Current Bill ~$3,700", digits=0)

# Chicago only has 2 pins that are similar to Dolton's median pin (which had a lot of matches)

## Dolton ## 
DoltonChicago %>% 
  filter(agency_num == "030310000" & class == "203" & between(tax_amt_post_exe,3600,3800)) %>%
  summarize(
    comp_taxrate = mean(tax_code_rate),
    bill_current = mean(tax_amt_post_exe),
    bill_hyp = mean(eav*taxrate_new, na.rm=TRUE),  
    bill_change = bill_hyp - bill_current,
    tax_amt_pre_exe = mean(tax_amt_pre_exe), # the naive amount that appears on peoples taxbills
    pin_count = n(),
    av = mean(av),
    eav = mean(eav)
  ) %>% 
    pivot_longer(cols = comp_taxrate:eav, names_to = "Stats", values_to = "Values")%>%
  kbl(caption= "Dolton, Class 203, Current Bill ~ $3,700", digits=0)
```



## Using pre-exemption taxbill amount of $3700:

For pre exemption tax bills around \$3700 in Chicago, average AV is \$18,378, EAV is \$55,184. Average current taxbill is $2901.


Chicago's Composite tax rate for the properties examined is 6.7%.
Dolton's average composite tax rate for the properties examiend was 24.2%.

For pre exemption tax bills around \$3700 in Dolton, average AV is \$5,235, EAV is \$15,719. Average current taxbill is $3044.


If holding the levy constant and acknowledging the change in tax rates that would occur from having additional taxable EAV within the taxing jurisdictions, Chicago's hypothetical tax bill would be \$2990 (an \$89 increase from \$2901) and Dolton's would be \$2377 (a \$667 DECREASE from $3044). 

Uses a tax bill range of 3600 to 3800 to increase "median pins" used for summary stats. 

```{r}
## Chicago ## 
DoltonChicago %>% 
  filter(agency_num == "030210000" & class == "203" & between(tax_amt_pre_exe, 3600, 3800)) %>%
  summarize(
    comp_taxrate = mean(tax_code_rate),
    bill_current = mean(tax_amt_post_exe),
    bill_hyp = mean(eav*taxrate_new, na.rm=TRUE),
    bill_change = bill_hyp - bill_current,
    tax_amt_post_exe = mean(tax_amt_post_exe),
    tax_amt_exe = mean(tax_amt_exe),
    tax_amt_pre_exe = mean(tax_amt_pre_exe),
    pin_count = n(),
    av = mean(av),
    eav = mean(eav)
  )  %>% 
    pivot_longer(cols = comp_taxrate:eav, names_to = "Stats", values_to = "Values")%>%
  kbl(caption= "Chicago, Class 203, Naive Pre-Exemp Bill ~ $3,700", digits=0)


## Dolton ##
DoltonChicago %>% 
  filter(agency_num == "030310000" & class == "203" & between(tax_amt_pre_exe,3600,3800)) %>%
  summarize(
    comp_taxrate = mean(tax_code_rate),
    bill_current = mean(tax_amt_post_exe),
    bill_hyp = mean(eav*taxrate_new, na.rm=TRUE),
    bill_change = bill_hyp - bill_current,
    tax_amt_exe = mean(tax_amt_exe),
    tax_amt_pre_exe = mean(tax_amt_pre_exe),
    pin_count = n(),
    av = mean(av),
    eav = mean(eav)
  )  %>% 
    pivot_longer(cols = comp_taxrate:eav, names_to = "Stats", values_to = "Values")%>%
  kbl(caption= "Dolton, Class 203, Naive Pre-Exemp Bill ~ $3,700", digits=0)
```
