==> old_versions/5b_GHE_Scenarios.qmd <==
---
title: "Exemption Progressivity & GHE Scenarios"
author: "Alea Wilbur"
date: "`r Sys.Date()`"
format: 
  html:
    df-print: paged
    code-fold: true
    code-download: true
    toc: true
    toc-location: left
    fig-cap-location: top
---



```{r setup, warning = FALSE, message = FALSE}
knitr::opts_chunk$set(echo = TRUE, warning = FALSE, message = FALSE)


library(tidyverse)
library(DBI)
library(data.table)
library(ggspatial)
library(gstat)
library(here)
library(httr)
library(jsonlite)
library(ptaxsim)
library(sf)
library(stars)
library(glue)
library(ggpattern)

# Create the DB connection with the default name expected by PTAXSIM functions
ptaxsim_db_conn <- DBI::dbConnect(RSQLite::SQLite(), "./ptaxsim.db/ptaxsim-2021.0.4.db")


options(digits=4, scipen = 999)


# link to the API output as a JSON file
muni_shp <- read_sf("https://gis.cookcountyil.gov/traditional/rest/services/politicalBoundary/MapServer/2/query?outFields=*&where=1%3D1&f=geojson")

cook_shp <- read_sf("https://gis.cookcountyil.gov/traditional/rest/services/plss/MapServer/1/query?outFields=*&where=1%3D1&f=geojson")

township_shp <- read_sf("https://gis.cookcountyil.gov/traditional/rest/services/politicalBoundary/MapServer/3/query?outFields=*&where=1%3D1&f=geojson")

#muni_shp <- read_json("muni_shp.json")
nicknames <- readxl::read_excel("./Necessary_Files/muni_shortnames.xlsx")

class_dict <- read_csv("./Necessary_Files/class_dict_expanded.csv") %>% 
  mutate(class_code = as.character(class_code))



# `agency_dt` has all taxing agencies (but not TIFs) that existed each year and includes their total taxable base (cty_cook_eav), their levy, taxing rate, binary variables for if a municipality is home rule or not, as well as many other variables. tax_bill() uses this table for the taxable EAV that is used to  calculate the tax rates in the tax bills. For simulations, you must alter the taxable EAV or levy or other variables and then tell tax_bill() function to use the modified agency data table for simulated tax bills.
# 



# has EAV values, extensions by agency_num
agency_dt <- DBI::dbGetQuery(
  ptaxsim_db_conn,
  "SELECT *
  FROM agency
  WHERE year = 2021
  "
)


# cook_agency_names <- DBI::dbGetQuery(
#   ptaxsim_db_conn,
#   "SELECT DISTINCT agency_num, agency_name
#   FROM agency_info
#   "
# )
# 
#  
# 
# 
# # has all tax codes and the taxing agency that taxes them. Tax code rates and agency rates. 
# cook_tax_codes <- DBI::dbGetQuery(
#   ptaxsim_db_conn,
#   glue_sql("
#   SELECT*
#   FROM tax_code
#   WHERE agency_num IN ({cook_agency_names$agency_num*})
#   AND year = 2021
#   ",
#   .con = ptaxsim_db_conn
#   )
# )


muni_agency_names <- DBI::dbGetQuery(
  ptaxsim_db_conn,
  "SELECT DISTINCT agency_num, agency_name, minor_type
  FROM agency_info
  WHERE minor_type = 'MUNI'
  OR agency_num = '020060000'  

  "
)

muni_tax_codes <- DBI::dbGetQuery(
  ptaxsim_db_conn,
  glue_sql("
  SELECT*
  FROM tax_code
  WHERE agency_num IN ({muni_agency_names$agency_num*})
  AND year = 2021
  ",
  .con = ptaxsim_db_conn
  )
)  
  
tax_codes <- DBI::dbGetQuery(
  ptaxsim_db_conn,
  glue_sql("
  SELECT DISTINCT tax_code_num, tax_code_rate
  FROM tax_code
  WHERE year = 2021  
  ",
  .con = ptaxsim_db_conn
  )
)

## All tax codes. 
## tax codes within municipalities have additional info 
tc_muninames <- tax_codes %>% 
  left_join(muni_tax_codes) %>%
  left_join(muni_agency_names) %>% 
  select(-agency_rate) %>% 
  left_join(nicknames) %>% 
  select(-c(minor_type, short_name, agency_number))


# Agency number and agency name for all TIFs
TIF_agencies <- DBI::dbGetQuery(
  ptaxsim_db_conn,
  "SELECT DISTINCT agency_num, agency_name, major_type, minor_type
  FROM agency_info
  WHERE minor_type = 'TIF'
  "
)

unique_tif_taxcodes <- DBI::dbGetQuery(
  ptaxsim_db_conn, 
  glue_sql("
  SELECT DISTINCT tax_code_num
  FROM tax_code
  WHERE agency_num IN ({TIF_agencies$agency_num*})
  AND year = 2021
  ",
  .con = ptaxsim_db_conn
  )
)


tif_distrib <- DBI::dbGetQuery(
  ptaxsim_db_conn, 
  glue_sql("
  SELECT *
  FROM tif_distribution
  WHERE tax_code_num IN ({muni_tax_codes$tax_code_num*})
  AND year = 2021
  ",
  .con = ptaxsim_db_conn
  )
) %>% mutate(tax_code_num = as.character(tax_code_num))



cross_county_lines <- c("030440000", "030585000", "030890000", "030320000", "031280000","030080000", "030560000", "031120000", "030280000", "030340000","030150000","030050000", "030180000","030500000","031210000")


cross_county_lines <- muni_agency_names %>% 
  filter(agency_num %in% cross_county_lines) %>% 
  left_join(nicknames, by = "agency_name")
```

# Changing General Homeowner Exemption - Scenarios

Create a PIN input with modified exemption amounts, then recalculate the base by taking the difference between the real and hypothetical exemptions.

Then recalculate the tax base for each district. If increasing the exemption, the base should decrease because there is less taxable EAV.

```{r current-taxsystem, eval=FALSE, include =FALSE}

### NOT NEEDED ### 

## Used to check the "current year" in the code chunk below
## Where the current exemption is reduced by 10,000 
## Like our very main scenario where we dropped exemptions 


tax_codes <- DBI::dbGetQuery(
  ptaxsim_db_conn,
  glue_sql("
  SELECT DISTINCT tax_code_num, tax_code_rate
  FROM tax_code
  WHERE year = 2021
  ",
  .con = ptaxsim_db_conn
  )
) %>% mutate(tax_code_num = as.numeric(tax_code_num))



# Load pins
t_pins <- DBI::dbGetQuery(
  ptaxsim_db_conn,
  "
  SELECT DISTINCT pin
  FROM pin
  WHERE year = 2021
  "
)

t_pins <- t_pins$pin
t_years <- 2021

rates <- NULL
exe_steps <- c(0)

for(m in exe_steps){
  
# Set exemption value
  t_pin_dt_new_exe <- lookup_pin(t_years, t_pins)
 # t_pin_dt_new_exe[, tax_code := lookup_tax_code(year, pin)]
  
  # new taxcode level sums of exemption
  
   t_pin_dt_new_exe <- t_pin_dt_new_exe[
    ,.(exe_homeowner = ifelse(m*exe_homeowner!=0 > eav, eav, 
                              ifelse(m*exe_homeowner!=0<eav, m, exe_homeowner)))
    ]
   
   
  t_tc_sum_new_exe <- t_pin_dt_new_exe[
    ,.(exe_homeowner = ifelse(m*exe_homeowner!=0 > eav, eav, 
                              ifelse(m*exe_homeowner!=0<eav, m, exe_homeowner)))
    ][
    ,.(exe_homeowner = sum(m*exe_homeowner!=0)),
    by = .(year, tax_code)]
  [
    # if exe_homeowner > 0, then add the alternate GHE amount to the current GHE amount
    # calculates the amount of EAV to be added/subtracted from taxable base of taxing agencies
    , .(exe_total = sum(((m * exe_homeowner!=0)))),
    by = .(year, tax_code)
  ]
  
  # Calculate bases by adding agency total and the exemption amount
  # more complex than it needs to be since we only are working with 1 year of data
  t_agency_dt_new_exe <- lookup_agency(t_years, t_pin_dt_new_exe$tax_code)
  t_agency_dt_new_exe[
    t_tc_sum_new_exe,
    on = .(year, tax_code),
    agency_total_eav := agency_total_eav - exe_total
  ][, total_final_rate := total_ext/cty_cook_eav]
  
  # Recalculate tax bills
  t_pin_dt_new_exe <- t_pin_dt_new_exe[
   # exe_homeowner!=0, exe_homeowner := exe_homeowner+m
    exe_homeowner!=0, exe_homeowner := ifelse(exe_homeowner + m > eav , eav, exe_homeowner + m)

  ][
    , c("tax_code") := NULL       # deletes tax_code column
  ]
  
  bills <- tax_bill(
    year_vec = t_years,
    pin_vec = t_pins,
    agency_dt = t_agency_dt_new_exe,
    pin_dt = t_pin_dt_new_exe,
    simplify = FALSE
  )
  
# Clear up memory
  rm(t_pin_dt_new_exe, t_tc_sum_new_exe)
  
    # Combine bills, aggregate to pin by year
  current_rates <- bills %>%
    dplyr::group_by(pin, tax_code) %>%
    mutate(total_bill = final_tax_to_dist + final_tax_to_tif) %>% # from each taxing agency
    
    summarize(
      total_billed = sum(total_bill, na.rm = TRUE), # total on someone's property tax bill
      av = first(av),
      eav = first(eav),
      taxing_agency_count = n(), # number of taxing agencies that tax the pin
      final_tax_to_dist = sum(final_tax_to_dist, na.rm = TRUE), # portion of all levies paid by the pin
      final_tax_to_tif = sum(final_tax_to_tif, na.rm = TRUE), 
      tax_amt_exe = sum(tax_amt_exe, na.rm = TRUE),           # revenue "lost" due to exemptions
      tax_amt_pre_exe = sum(tax_amt_pre_exe, na.rm = TRUE),   # total rev before all exemptions
      tax_amt_post_exe = sum(tax_amt_post_exe, na.rm = TRUE), # total rev after all exemptions
      # rpm_tif_to_cps = sum(rpm_tif_to_cps, na.rm = TRUE),     # not used
      # rpm_tif_to_rpm = sum(rpm_tif_to_rpm, na.rm=TRUE),       # not used
      # rpm_tif_to_dist = sum(rpm_tif_to_dist, na.rm=TRUE),     # not used
      # tif_share = mean(tif_share, na.rm=TRUE),                # not used
    ) %>%
    mutate(tax_code = as.numeric(tax_code)) %>% 
    left_join(tax_codes, by = c("tax_code" = "tax_code_num") ) %>% # add current, real tax code level tax rates
    mutate(exemption_level = m) %>%
    dplyr::group_by(tax_code, exemption_level) %>%
    dplyr::summarize(
      tc_levy = sum(final_tax_to_dist, na.rm = TRUE), # amount billed by munis with current exemptions in place
      nonTIF_EAV_post_exemps = sum(final_tax_to_dist/(tax_code_rate/100), na.rm = TRUE),
      TIF_increment_EAV = sum(final_tax_to_tif/(tax_code_rate/100), na.rm=TRUE),  
      Exempt_EAV = sum(tax_amt_exe/(tax_code_rate/100), na.rm=TRUE), 
      Total_EAV = sum((tax_amt_exe+final_tax_to_dist+final_tax_to_tif)/(tax_code_rate/100), na.rm = TRUE),
    ) %>%
  
    mutate(
      tax_rate_current = tc_levy/nonTIF_EAV_post_exemps,
      nonTIF_EAV_pre_exemps = nonTIF_EAV_post_exemps + Exempt_EAV,
      taxrate_new = tc_levy/nonTIF_EAV_pre_exemps, # tax rate if there were no exemptions
      taxrate_change = tax_rate_current-taxrate_new,
    ) %>% 
    select(tax_code, taxrate_change, tax_rate_current, taxrate_new, everything()) %>% 
    arrange(desc(tax_code))
  
   # rm(bills)
}

tc_current_rates <- rates

bills <- tax_bill(2021, pin_vec = t_pins, simplify = FALSE)


munilevel_current_rates <- bills %>%
    dplyr::group_by(pin, tax_code) %>%
    mutate(total_bill = final_tax_to_dist + final_tax_to_tif) %>%      # from each taxing agency
    summarize(
      total_billed = sum(total_bill, na.rm = TRUE),                    # total on someone's property tax bill
      av = first(av),
      eav = first(eav),
      taxing_agency_count = n(),                                       # number of taxing agencies that tax the pin
      final_tax_to_dist = sum(final_tax_to_dist, na.rm = TRUE),        # portion of all levies paid by the pin
      final_tax_to_tif = sum(final_tax_to_tif, na.rm = TRUE), 
      tax_amt_exe = sum(tax_amt_exe, na.rm = TRUE),           # revenue lost due to exemptions
      tax_amt_pre_exe = sum(tax_amt_pre_exe, na.rm = TRUE),   # total rev before all exemptions
      tax_amt_post_exe = sum(tax_amt_post_exe, na.rm = TRUE), # total rev after all exemptions
      ) %>%
    # mutate(tax_code = as.numeric(tax_code)) %>% 
    left_join(tc_muninames, by = c("tax_code" = "tax_code_num") ) %>%
  # add current, real tax code level tax rates
    dplyr::group_by(clean_name, agency_num) %>%
    dplyr::summarize(
      muni_levy = sum(final_tax_to_dist, na.rm = TRUE), # amount billed by munis with current exemptions in place
      nonTIF_EAV_post_exemps = sum(final_tax_to_dist/(tax_code_rate/100), na.rm = TRUE),
      TIF_increment_EAV = sum(final_tax_to_tif/(tax_code_rate/100), na.rm=TRUE),  
      Exempt_EAV = sum(tax_amt_exe/(tax_code_rate/100), na.rm=TRUE), 
      Total_EAV = sum((tax_amt_exe+final_tax_to_dist+final_tax_to_tif)/(tax_code_rate/100), na.rm = TRUE),
      pin_count = n()
    ) %>%
  
    mutate(
      tax_rate_current = muni_levy/nonTIF_EAV_post_exemps,
      nonTIF_EAV_pre_exemps = nonTIF_EAV_post_exemps + Exempt_EAV,
      taxrate_new = muni_levy/nonTIF_EAV_pre_exemps, # tax rate if there were no exemptions at all
      taxrate_change = tax_rate_current-taxrate_new,
    )
  
  
```

```{r scenarios-taxcodelevel, eval=FALSE, include=FALSE}

tax_codes <- DBI::dbGetQuery(
  ptaxsim_db_conn,
  glue_sql("
  SELECT DISTINCT tax_code_num, tax_code_rate
  FROM tax_code
  WHERE year = 2021
  ",
  .con = ptaxsim_db_conn
  )
) %>% mutate(tax_code_num = as.numeric(tax_code_num))



# Load pins
t_pins <- DBI::dbGetQuery(
  ptaxsim_db_conn,
  "
  SELECT DISTINCT pin
  FROM pin
  "
)

t_pins <- t_pins$pin
t_years <- 2021

rates <- NULL
exe_steps <- c(-10000, -5000, 0, 5000, 10000, 15000, 20000, 30000, 40000)

start_time <- Sys.time()
for(m in exe_steps){
  
# Set exemption value
  t_pin_dt_new_exe <- lookup_pin(t_years, t_pins)
  t_pin_dt_new_exe[, tax_code := lookup_tax_code(year, pin)]
  
  # new taxcode level sums of exemption
  t_tc_sum_new_exe <- t_pin_dt_new_exe[
    # if exe_homeowner > 0, then add the alternate GHE amount to the current GHE amount
    # calculates the amount of EAV to be added/subtracted from taxable base of taxing agencies
    , .(exe_total = sum(((m * exe_homeowner!=0)))),
    by = .(year, tax_code)
  ]
  
  # Calculate bases by adding agency total and the exemption amount
  t_agency_dt_new_exe <- lookup_agency(t_years, t_pin_dt_new_exe$tax_code)
  t_agency_dt_new_exe[
    t_tc_sum_new_exe,
    on = .(year, tax_code),
    agency_total_eav := agency_total_eav - exe_total
  ]
  
  # Recalculate tax bills
  t_pin_dt_new_exe <- t_pin_dt_new_exe[
    exe_homeowner!=0, exe_homeowner := exe_homeowner+m
  ][
    ### ADD IF ELSE STATEMENT 
    
    
    , c("tax_code") := NULL       # deletes tax_code column
  ]
  
  bills <- tax_bill(
    year_vec = t_years,
    pin_vec = t_pins,
    agency_dt = t_agency_dt_new_exe,
    pin_dt = t_pin_dt_new_exe,
    simplify = FALSE
  )
  
# Clear up memory
  rm(t_pin_dt_new_exe, t_tc_sum_new_exe)
  
    # Combine bills, aggregate to pin by year
  rates2 <- bills %>%
    dplyr::group_by(pin, tax_code) %>%
    
    mutate(total_bill = final_tax_to_dist + final_tax_to_tif) %>% # from each taxing agency
    
    summarize(
      total_billed = sum(total_bill, na.rm = TRUE), # total on someone's property tax bill
      av = first(av),
      eav = first(eav),
      taxing_agency_count = n(), # number of taxing agencies that tax the pin
      final_tax_to_dist = sum(final_tax_to_dist, na.rm = TRUE), # portion of all levies paid by the pin
      final_tax_to_tif = sum(final_tax_to_tif, na.rm = TRUE), 
      tax_amt_exe = sum(tax_amt_exe, na.rm = TRUE),           # revenue lost due to exemptions
      tax_amt_pre_exe = sum(tax_amt_pre_exe, na.rm = TRUE),   # total rev before all exemptions
      tax_amt_post_exe = sum(tax_amt_post_exe, na.rm = TRUE)#, # total rev after all exemptions
      # rpm_tif_to_cps = sum(rpm_tif_to_cps, na.rm = TRUE),     # not used
      # rpm_tif_to_rpm = sum(rpm_tif_to_rpm, na.rm=TRUE),       # not used
      # rpm_tif_to_dist = sum(rpm_tif_to_dist, na.rm=TRUE),     # not used
      # tif_share = mean(tif_share, na.rm=TRUE),                # not used
    ) %>%
   # mutate(tax_code = as.numeric(tax_code)) %>% 
    mutate(exemption_level = m) %>%
   # left_join(tax_codes, by = c("tax_code" = "tax_code_num") ) %>% # add current, real tax code level tax rates
    dplyr::group_by(tax_code, exemption_level) %>%
    dplyr::summarize(
      total_billed = sum(total_billed, na.rm = TRUE), # total on someone's property tax bill
      av = sum(av, na.rm=TRUE),
      eav = sum(eav, na.rm = TRUE),
      taxing_agency_count = first(taxing_agency_count), # number of taxing agencies that tax the pin
      final_tax_to_dist = sum(final_tax_to_dist, na.rm = TRUE), # portion of all levies paid by the pin
      final_tax_to_tif = sum(final_tax_to_tif, na.rm = TRUE), 
      tax_amt_exe = sum(tax_amt_exe, na.rm = TRUE),           # revenue lost due to exemptions
      tax_amt_pre_exe = sum(tax_amt_pre_exe, na.rm = TRUE),   # total rev before all exemptions
      tax_amt_post_exe = sum(tax_amt_post_exe, na.rm = TRUE) ) %>% # total rev after all exemptions
# 
#     dplyr::summarize(
#       tc_levy = sum(final_tax_to_dist, na.rm = TRUE), # amount billed by munis with current exemptions in place
#       nonTIF_EAV_post_exemps = sum(final_tax_to_dist/(tax_code_rate/100), na.rm = TRUE),
#       TIF_increment_EAV = sum(final_tax_to_tif/(tax_code_rate/100), na.rm=TRUE),  
#       Exempt_EAV = sum(tax_amt_exe/(tax_code_rate/100), na.rm=TRUE), 
#       Total_EAV = sum((tax_amt_exe+final_tax_to_dist+final_tax_to_tif)/(tax_code_rate/100), na.rm = TRUE),
#     ) %>%
  
    # mutate(
    #   tc_levy = sum(final_tax_to_dist, na.rm = TRUE), # amount billed by munis with current exemptions in place
    #   
    #   tax_rate_current = tc_levy/nonTIF_EAV_post_exemps,
    #   nonTIF_EAV_pre_exemps = nonTIF_EAV_post_exemps + Exempt_EAV,
    #   taxrate_new = tc_levy/nonTIF_EAV_pre_exemps,
    #   taxrate_change = tax_rate_current-taxrate_new,
    # ) %>% 
     select(tax_code, everything())

  
  if(is.data.frame(rates)){rates <- rbind(rates2, rates)}else{rates <- rates2}
  rm(rates2, bills)
}
end_time <- Sys.time()
end_time - start_time

write_csv(rates, "C:/Users/aleaw/OneDrive/Documents/PhD Fall 2021 - Spring 2022/Merriman RA/ptax/Output/5b_scenario_rates3.csv")
```

<!--- line 521 - finish thought!  --->


```{r scenarios-munilevel, eval=FALSE, include=FALSE}

rates <- NULL
exe_steps <- c(-10000, -5000, 0, 5000, 10000, 15000, 20000, 30000, 40000)

start_time <- Sys.time()

for(m in exe_steps){
  
# Set exemption value
  t_pin_dt_new_exe <- lookup_pin(t_years, t_pins)      # creates "current" tax system pin data with current exemptions
  
  t_pin_dt_new_exe[, tax_code := lookup_tax_code(year, pin)]  # adds tax code variable to pin & exemption data 
  
  
  

  # new taxcode level sums of exemption
  
 # summarizes pin data table that has the tax code variable by year and tax code
 # creates a new variable exe_total
  t_tc_sum_new_exe <- t_pin_dt_new_exe[
    # if exe_homeowner > 0, then add the alternate GHE amount to the current GHE amount
    # calculates the amount of EAV to be added/subtracted from taxable base of taxing agencies
    , .(exe_change = sum(((m * exe_homeowner!=0)))),
    by = .(year, tax_code)
  ]
  
##### COME BACK TO THIS POINT ###### 
  
# # Calculate bases by adding agency total and the exemption amount # # 
  
  # new agency data table <- # same input as normal agency table
  t_agency_dt_new_exe <- lookup_agency(t_years, t_pin_dt_new_exe$tax_code)
  
  ## updates taxable base EAV in agency data table
  t_agency_dt_new_exe[
    t_tc_sum_new_exe,
    on = .(year, tax_code),
    agency_total_eav := agency_total_eav - exe_change
  ]
  
  # # Recalculate tax bills
  # t_pin_dt_new_exe <- t_pin_dt_new_exe[
  #   exe_homeowner!=0, exe_homeowner := exe_homeowner + m
  # ][
  #   exe_homeowner > eav, exe_homeowner:= eav
  # ][
  #   , c("tax_code") := NULL       # deletes tax_code column
  # ]

# Recalculate tax bills
t_pin_dt_new_exe <- t_pin_dt_new_exe[
 exe_homeowner!= 0, exe_homeowner := ifelse(exe_homeowner + m > eav , eav, exe_homeowner + m)
 ][
    , c("tax_code") := NULL       # deletes tax_code column from pin table that has exemption data
  ]

## re-sum all homeowner exemptions by agency * update taxable EAV in agency_dt 
  
  bills <- tax_bill(
    year_vec = t_years,
    pin_vec = t_pins,
    agency_dt = t_agency_dt_new_exe,
    pin_dt = t_pin_dt_new_exe,
    simplify = FALSE
  )
  
# Clear up memory
  rm(t_pin_dt_new_exe, t_tc_sum_new_exe)
  
    # Combine bills, aggregate to pin by year
  rates2 <- bills %>%
    dplyr::group_by(pin, tax_code) %>%
    
    mutate(total_bill = final_tax_to_dist + final_tax_to_tif) %>% # from each taxing agency
    
    summarize(
      total_billed = sum(total_bill, na.rm = TRUE), # total on someone's property tax bill
      av = first(av),
      eav = first(eav),
      taxing_agency_count = n(), # number of taxing agencies that tax the pin
      final_tax_to_dist = sum(final_tax_to_dist, na.rm = TRUE), # portion of all levies paid by the pin
      final_tax_to_tif = sum(final_tax_to_tif, na.rm = TRUE), 
      tax_amt_exe = sum(tax_amt_exe, na.rm = TRUE),           # revenue lost due to exemptions
      tax_amt_pre_exe = sum(tax_amt_pre_exe, na.rm = TRUE),   # total rev before all exemptions
      tax_amt_post_exe = sum(tax_amt_post_exe, na.rm = TRUE)#, # total rev after all exemptions
) %>%
   # mutate(tax_code = as.numeric(tax_code)) %>% 
    mutate(exemption_level = m) %>%
   # left_join(tax_codes, by = c("tax_code" = "tax_code_num") ) %>% # add current, real tax code level tax rates
    left_join(tc_muninames, by = c("tax_code" = "tax_code_num")) %>%
    dplyr::group_by(clean_name, agency_num, exemption_level) %>%
    dplyr::summarize(
      total_billed = sum(total_billed, na.rm = TRUE), # total on someone's property tax bill
      av = sum(av, na.rm=TRUE),
      eav = sum(eav, na.rm = TRUE),
      
      nonTIF_EAV_post_exemps = sum(final_tax_to_dist/(tax_code_rate/100), na.rm = TRUE),
      TIF_increment_EAV = sum(final_tax_to_tif/(tax_code_rate/100), na.rm=TRUE),
      Exempt_EAV = sum(tax_amt_exe/(tax_code_rate/100), na.rm=TRUE),
      Total_EAV = sum((tax_amt_exe+final_tax_to_dist+final_tax_to_tif)/(tax_code_rate/100), na.rm = TRUE),
      final_tax_to_dist = sum(final_tax_to_dist, na.rm = TRUE), 
      final_tax_to_tif = sum(final_tax_to_tif, na.rm = TRUE), 
      tax_amt_exe = sum(tax_amt_exe, na.rm = TRUE),    
      tax_amt_pre_exe = sum(tax_amt_pre_exe, na.rm = TRUE),   
      tax_amt_post_exe = sum(tax_amt_post_exe, na.rm = TRUE),
      pin_count = n()

    ) %>%
  
    mutate(
      tax_rate_current = final_tax_to_dist/nonTIF_EAV_post_exemps,
      nonTIF_EAV_pre_exemps = nonTIF_EAV_post_exemps + Exempt_EAV,
      taxrate_new = final_tax_to_dist/nonTIF_EAV_pre_exemps,
      taxrate_change = tax_rate_current-taxrate_new,
    ) %>%
     select(clean_name, everything())

  
  
  if(is.data.frame(rates)){rates <- rbind(rates2, rates)}else{rates <- rates2}
  rm(rates2, bills)
}
end_time <- Sys.time()
end_time - start_time


write_csv(rates, "C:/Users/aleaw/OneDrive/Documents/PhD Fall 2021 - Spring 2022/Merriman RA/ptax/Output/5b_scenario_rates_munilevel_scenarios2.csv")



```


```{r read-in-rates, eval=FALSE, include=FALSE}
scenario_rates <- read_csv("C:/Users/aleaw/OneDrive/Documents/PhD Fall 2021 - Spring 2022/Merriman RA/ptax/Output/5b_scenario_rates3.csv") %>%
 # read_csv("C:/Users/aleaw/OneDrive/Documents/PhD Fall 2021 - Spring 2022/Merriman RA/ptax/Output/5b_scenario_rates.csv") %>%
  mutate(tax_code = as.character(tax_code)) %>% 
  left_join(tc_muninames, by = c("tax_code" = "tax_code_num")) %>%
  select(clean_name, tax_code, everything())

 # scenario_rates %>% write_csv("C:/Users/aleaw/OneDrive/Documents/PhD Fall 2021 - Spring 2022/Merriman RA/ptax/Output/5b_scenario_rates_labeled2.csv")

scenario_rates

scenario_rates_munilevel_scenarios <- read_csv("C:/Users/aleaw/OneDrive/Documents/PhD Fall 2021 - Spring 2022/Merriman RA/ptax/Output/5b_scenario_rates_munilevel_scenarios.csv")

scenario_rates_munilevel_scenarios

```


## Quartiles and Progressivity

```{r}
class_dict <- read_csv("./Necessary_Files/class_dict_singlefamcodes.csv") %>% 
  mutate(class_code = as.character(class_code)) # change variable type to character so the join works.

nicknames <- readxl::read_xlsx("./Necessary_Files/muni_shortnames.xlsx")

#pin_data2 <- read_csv("./Output/4C_joined_PINs_bills_and_exemptions.csv")

muni_taxrates <- read_csv("./Output/4C_muni_taxrates.csv")

#pin_data2 <- pin_data2 %>% left_join(class_dict)

muni_TC_fullyCook <- muni_tax_codes %>%
  filter(!agency_num %in% cross_county_lines)


joined_pins <- read_csv("./Output/4C_joined_PINs_bills_and_exemptions.csv") %>%
  mutate(tax_code_num = as.character(tax_code_num)) %>%  
  left_join(tc_muninames) %>% left_join(class_dict)

# all pins in munis fully within cook county that are some form of single-family, detached home
singfam_pins <- joined_pins %>% 
  filter(tax_code %in% muni_TC_fullyCook$tax_code_num) %>% # excludes county line crossing munis
  filter(Option2 == "Single-Family")
```

## Cook County Quartiles

Cook County quartiles are calculated fromg single family properties assessed value in 2021. 


```{r}
q = c(.25, .5, .75)


cook_quartiles <- singfam_pins %>%
  filter(Option2 == "Single-Family") %>%
  filter(tax_code %in% muni_TC_fullyCook$tax_code_num) %>% # excludes county line crossing munis
  arrange(av) %>%
  summarize(count_pins = n(), 
            min = min(av),
            quant25 = round(quantile(av, probs = q[1])), 
            quant50 = round(quantile(av, probs = q[2])),
            quant75 = round(quantile(av, probs = q[3])),
            max = max(av))
cook_quartiles

```




## Scenarios & tax rates

Similar to File 5_Exemption_Scenarios.rmd. 

- Calculate Class 2 Burden --> Calculate the amount of taxable EAV in the Municipality (for each scenario) and multiply it by the new composite tax rate (for each scenario).   


- Burden Share  = Taxable EAV within Property Class * Composite tax rate   

- Composite Tax Rate = (Municipal Levy / Taxable EAV )    


```{r}
## Bring in tax bills and exemption data for 2021 PINs ##
# 
# joined_pins <- read_csv("./Output/4C_joined_PINs_bills_and_exemptions.csv") %>%
#   mutate(tax_code_num = as.character(tax_code_num)) %>%  
#   left_join(tc_muninames) %>% left_join(class_dict)

MuniLevy <- joined_pins %>% 
  group_by(clean_name, agency_num) %>%
  
  summarize(MuniLevy = sum(final_tax_to_dist, na.rm = TRUE), # amount billed by munis with current exemptions in place
            current_nonTIF_EAV_post_exemps = sum(final_tax_to_dist/(tax_code_rate/100), na.rm = TRUE),
            current_TIF_increment_EAV = sum(final_tax_to_tif/(tax_code_rate/100), na.rm=TRUE),  
            current_Exempt_EAV = sum(tax_amt_exe/(tax_code_rate/100), na.rm=TRUE), 
            Total_EAV = sum((tax_amt_exe+final_tax_to_dist+final_tax_to_tif)/(tax_code_rate/100), na.rm = TRUE)) %>% 
  mutate(cur_muni_comp_rate = MuniLevy / current_nonTIF_EAV_post_exemps)

MuniLevy
```

```{r}
joined_pins <- joined_pins %>% 
  mutate(exe_neg10 = 0,
         exe_0 = ifelse(eav < 10000 & exe_homeowner!=0, eav, 
                             ifelse(eav>10000 & exe_homeowner!=0, 10000, 0 )),  #would be if there is no change in exemptions
         exe_plus10 = ifelse(eav < 20000 & exe_homeowner!=0, eav, 
                             ifelse(eav>20000 & exe_homeowner!=0, 20000, 0 )),
         exe_plus20 = ifelse(eav < 30000 & exe_homeowner!=0, eav, 
                             ifelse(eav>30000 & exe_homeowner!=0, 30000, 0 ) ),
         exe_plus30 = ifelse(eav < 40000 & exe_homeowner!=0, eav, 
                             ifelse(eav>40000 & exe_homeowner!=0, 40000, 0) ),
         exe_plus40 = ifelse(eav < 50000 & exe_homeowner!=0, eav, 
                             ifelse(eav>50000 & exe_homeowner!=0, 50000, 0) ) )

scenario_calcs <- joined_pins %>%    
  group_by(clean_name) %>%

    summarize(MuniLevy = sum(final_tax_to_dist, na.rm = TRUE), # amount billed by munis with current exemptions in place
            current_nonTIF_EAV_post_exemps = sum(final_tax_to_dist/(tax_code_rate/100), na.rm = TRUE),
            current_TIF_increment_EAV = sum(final_tax_to_tif/(tax_code_rate/100), na.rm=TRUE),  
            current_Exempt_EAV = sum(tax_amt_exe/(tax_code_rate/100), na.rm=TRUE), 
            current_GHE = sum(exe_homeowner, na.rm=TRUE),
            Total_EAV = sum((tax_amt_exe+final_tax_to_dist+final_tax_to_tif)/(tax_code_rate/100), na.rm = TRUE),
            exe_neg10 = sum(exe_neg10),
            exe_0 = sum(exe_0), # no change, for comparison
            exe_plus10 = sum(exe_plus10),
            exe_plus20 = sum(exe_plus20),
            exe_plus30 = sum(exe_plus30),
            exe_plus40 = sum(exe_plus40)) %>%

  # remove all GHE (up to 10,000 EAV added back to base per PIN), 
  # add exe_homeowner back to taxable base
  mutate(neg10_taxable_eav = Total_EAV - current_TIF_increment_EAV - current_Exempt_EAV + current_GHE,    # adds GHE exempt EAV back to taxable base and decreases tax rates
         plus10_taxable_eav = Total_EAV - current_TIF_increment_EAV - current_Exempt_EAV + current_GHE - exe_plus10, # will increase tax rates
         plus20_taxable_eav = Total_EAV - current_TIF_increment_EAV - current_Exempt_EAV + current_GHE - exe_plus20,
         plus30_taxable_eav = Total_EAV - current_TIF_increment_EAV - current_Exempt_EAV + current_GHE - exe_plus30,
         plus40_taxable_eav = Total_EAV - current_TIF_increment_EAV - current_Exempt_EAV + current_GHE - exe_plus40,
         scenario_noexemptions_taxable_eav = Total_EAV - current_TIF_increment_EAV) %>%
  
  mutate(tr_neg10 = MuniLevy / neg10_taxable_eav,
         tr_nochange = MuniLevy / current_nonTIF_EAV_post_exemps,
         tr_plus10 = MuniLevy / plus10_taxable_eav,
         tr_plus20 = MuniLevy / plus20_taxable_eav,
         tr_plus30 = MuniLevy / plus30_taxable_eav,
         tr_plus40 = MuniLevy / plus40_taxable_eav, 
         tax_rate_current = MuniLevy/current_nonTIF_EAV_post_exemps,
         taxrate_noexemps = MuniLevy /(Total_EAV - current_TIF_increment_EAV  ),
         taxrate_noTIFs = MuniLevy / (Total_EAV - current_Exempt_EAV),
         taxrate_noTIFs_orExemps = MuniLevy / Total_EAV) %>%
  select(clean_name, MuniLevy, tr_neg10:taxrate_noTIFs_orExemps, everything())

write_csv(scenario_calcs, "./Output/5b_scenario_calcs.csv")


scenario_taxrates <- scenario_calcs %>%  select(clean_name, MuniLevy, tr_neg10:taxrate_noTIFs_orExemps) 

scenario_taxrates

```

## Class 2 Burden Shift

```{r}

C2_taxableEAV <- joined_pins %>%   
  filter(class >= 200 & class <= 300) %>% 
  #left_join(scenario_taxrates, by = c("tax_code" = "tax_code_num")) %>%
  group_by(clean_name) %>%

    summarize(
      C2_av = sum(av),
      C2_eav_original = sum(equalized_AV), 
      C2_DistrictRev = sum(final_tax_to_dist, na.rm=TRUE),
      C2_current_nonTIF_EAV_post_exemps = sum(final_tax_to_dist/(tax_code_rate/100), na.rm = TRUE),
      C2_current_TIF_increment_EAV = sum(final_tax_to_tif/(tax_code_rate/100), na.rm=TRUE),  
      C2_current_Exempt_EAV = sum(tax_amt_exe/(tax_code_rate/100), na.rm=TRUE), 
      C2_current_GHE = sum(exe_homeowner, na.rm=TRUE),
      C2_Total_EAV = sum((tax_amt_exe+final_tax_to_dist+final_tax_to_tif)/(tax_code_rate/100), na.rm = TRUE),
      C2_exe_neg10 = sum(exe_neg10),
      C2_exe_0 = sum(exe_0), # no change, for comparison
      C2_exe_plus10 = sum(exe_plus10),
      C2_exe_plus20 = sum(exe_plus20),
      C2_exe_plus30 = sum(exe_plus30),
      C2_exe_plus40 = sum(exe_plus40),
      C2_PC_permuni = n())  %>% 
  left_join(MuniLevy, by = "clean_name") %>%
  mutate(C2_EAV_pct = C2_eav_original / Total_EAV)



C2_burden_shift <- C2_taxableEAV %>%
  left_join(scenario_taxrates) %>%
  mutate(C2_neg10_taxableEAV = C2_Total_EAV - C2_current_TIF_increment_EAV - C2_current_Exempt_EAV + C2_current_GHE - C2_exe_neg10,
         C2_nochange = C2_Total_EAV - C2_current_TIF_increment_EAV - C2_current_Exempt_EAV,
         C2_plus10_taxableEAV = C2_Total_EAV - C2_current_TIF_increment_EAV - C2_current_Exempt_EAV + C2_current_GHE - C2_exe_plus10,
         C2_plus20_taxableEAV = C2_Total_EAV - C2_current_TIF_increment_EAV - C2_current_Exempt_EAV + C2_current_GHE - C2_exe_plus20,
        C2_plus30_taxableEAV = C2_Total_EAV - C2_current_TIF_increment_EAV - C2_current_Exempt_EAV + C2_current_GHE - C2_exe_plus30,
        C2_plus40_taxableEAV = C2_Total_EAV - C2_current_TIF_increment_EAV - C2_current_Exempt_EAV + C2_current_GHE - C2_exe_plus40
         ) %>%
  mutate(burden_C2_neg10 = (C2_neg10_taxableEAV * tr_neg10)/ MuniLevy,
         burden_C2_nochange = C2_nochange * tax_rate_current  / MuniLevy,
         burden_C2_plus10 = (C2_plus10_taxableEAV * tr_plus10) / MuniLevy,
         burden_C2_plus20 = C2_plus20_taxableEAV * tr_plus20/ MuniLevy,
         burden_C2_plus30 = C2_plus30_taxableEAV * tr_plus30/ MuniLevy,
         burden_C2_plus40 = C2_plus40_taxableEAV * tr_plus40/ MuniLevy,
         
        burden_C2_noexemps = ( (C2_Total_EAV - C2_current_TIF_increment_EAV)*taxrate_noexemps ) / MuniLevy) %>%
  select(clean_name, C2_EAV_pct, burden_C2_neg10:burden_C2_plus40, everything())

C2_burden_shift

write_csv(C2_burden_shift, "./Output/5b_Class2_burdenshift.csv")
```


#### Scenario Tax rate graphs 

```{r}
scenarios_long <- scenario_taxrates  %>% 
  select(clean_name, MuniLevy, tr_neg10:taxrate_noTIFs_orExemps) %>%
  pivot_longer(cols = c(tr_neg10:taxrate_noTIFs_orExemps), names_to = "GHE_Amount")


scenario_taxrates %>% 
  select(clean_name, MuniLevy, tr_neg10:taxrate_noTIFs_orExemps) %>%
  pivot_longer(cols = c(tr_neg10:taxrate_noTIFs_orExemps), names_to = "GHE_Amount") %>%
  ggplot() + 
  geom_col(aes(x=value, y = GHE_Amount))

scenario_taxrates %>% 
  filter(clean_name %in% c("Chicago", "Dolton", "Glencoe")) %>%
  select(clean_name, MuniLevy, tr_neg10:taxrate_noTIFs_orExemps) %>%
  pivot_longer(cols = c(tr_neg10:taxrate_noTIFs_orExemps), names_to = "GHE_Amount") %>%
  ggplot() + 
  geom_col(aes(x=value, y = GHE_Amount, fill = clean_name), position = "dodge")  + 
  labs(x = "Municipality Composite Tax Rate", y = "Exemption Scenarios")
```


# Ranked Properties and Muni Quartiles

```{r}
q = c(.25, .5, .75)

## ranks properties that are considered single family homes in order of AV for each Muni
muni_quartiles <- joined_pins %>%
  filter(Option2 == "Single-Family") %>% 
  filter(tax_code %in% muni_TC_fullyCook$tax_code_num) %>%
  group_by(agency_name, clean_name) %>%
  arrange(av) %>%
  summarize(count_pins = n(), 
            min = min(av),
            quant25 = round(quantile(av, probs = q[1])), 
            quant50 = round(quantile(av, probs = q[2])),
            quant75 = round(quantile(av, probs = q[3])),
            max = max(av)
           ) %>% 
  arrange( desc( quant50))
muni_quartiles
```


```{r}
## create rank variable for properties that fall within the quartiles +/- $500 range
munis_ranked <- joined_pins  %>%
  inner_join(muni_quartiles, by = c("agency_name", "clean_name")) %>% 
  mutate(rank = case_when(
    av > (quant25-500) & (av<quant25+500) ~ "q25",
    av > (quant50-500) & (av<quant50+500) ~ "q50",
    av > (quant75-500) & (av<quant75+500) ~ "q75")
    ) %>%
  select(clean_name, rank, av, pin, class, everything()) %>%
  left_join(nicknames)




munis_billchange <-  munis_ranked %>% 
  group_by(clean_name, rank) %>%
  left_join(scenario_taxrates) %>%
  arrange(av) %>%
 # group_by(agency_name, has_HO_exemp) %>% 
  mutate(#taxable_eav = final_tax_to_dist / tax_code_rate,
    # current bill = current tax rate * portion of levy billed
    
    
   # ## Made negative tax bills!! ## #
         
         bill_neg10 = tr_neg10*(equalized_AV-all_exemptions+ exe_homeowner -exe_neg10),

         bill_current = cur_comp_TC_rate/100*(equalized_AV-all_exemptions),
         bill_plus10 =  tr_plus10*(equalized_AV-all_exemptions+ exe_homeowner -exe_plus10),
         bill_plus20 = tr_plus20*(equalized_AV-all_exemptions+ exe_homeowner -exe_plus20),
         bill_plus30 = tr_plus20*(equalized_AV-all_exemptions+ exe_homeowner -exe_plus30),
         bill_plus40 = tr_plus20*(equalized_AV-all_exemptions+ exe_homeowner -exe_plus40),
    
    # ## make bills $1 if they had been negative.    
         bill_neg10 = ifelse(bill_neg10 < 1, 1, bill_neg10),
         bill_current = ifelse(bill_current < 1, 1, bill_current),
         bill_plus10 = ifelse(bill_plus10 < 1, 1, bill_plus10),
         bill_plus20 = ifelse(bill_plus20 < 1, 1, bill_plus20),
         bill_plus30 = ifelse(bill_plus30 < 1, 1, bill_plus30),
         bill_plus40 = ifelse(bill_plus40 < 1, 1, bill_plus40),
         
## Prevent tax bills from having negative values  (if exemptions > eav of home)
         # bill_neg10 = ifelse(tr_neg10*(equalized_AV-all_exemptions+ exe_homeowner -exe_neg10) > 1,
         #                              tr_neg10*(equalized_AV-all_exemptions+ exe_homeowner -exe_neg10), 1),
         # 
         # bill_current = ifelse(cur_comp_TC_rate/100*(equalized_AV-all_exemptions) > 1,
         #                       cur_comp_TC_rate/100*(equalized_AV-all_exemptions), 1),
         # 
         # bill_plus10 =  ifelse(tr_plus10*(equalized_AV-all_exemptions+ exe_homeowner -exe_plus10) > 1,
         #                       tr_plus10*(equalized_AV-all_exemptions+ exe_homeowner -exe_plus10),1),
         # 
         # bill_plus20 = ifelse(tr_plus20*(equalized_AV-all_exemptions+ exe_homeowner -exe_plus20) > 1,
         #                      tr_plus20*(equalized_AV-all_exemptions+ exe_homeowner -exe_plus20), 1),
         # 
         # bill_plus30 = ifelse(tr_plus30*(equalized_AV-all_exemptions+ exe_homeowner -exe_plus30) >1, 
         #                      tr_plus30*(equalized_AV-all_exemptions+ exe_homeowner -exe_plus30), 1),
         #                      
         #                      
         # bill_plus40 = ifelse(tr_plus40*(equalized_AV-all_exemptions+ exe_homeowner -exe_plus40) > 1,
         #                      tr_plus40*(equalized_AV-all_exemptions+ exe_homeowner -exe_plus40), 1)
         )%>%
  mutate(
         zerodol_bills_ghe0 = ifelse(bill_neg10 < 5, 1, 0),
         zerodol_bills_current = ifelse(bill_current < 5, 1, 0),
         zerodol_bills_ghe20 = ifelse(bill_plus10 < 5, 1, 0),
         zerodol_bills_ghe30 = ifelse(bill_plus20 < 5, 1, 0),
         zerodol_bills_ghe40 = ifelse(bill_plus30 < 5, 1, 0),
         zerodol_bills_ghe50 = ifelse(bill_plus40 < 5, 1, 0),
  ) %>%
  group_by(clean_name, rank, has_HO_exemp) %>% 
  summarize(median_AV = round(median(av)),
            median_EAV = round(median(eav)),
            mean_bill_neg10 = round(mean(bill_neg10, na.rm=TRUE)),
            mean_bill_cur = round(mean(bill_current, na.rm=TRUE)),
            mean_bill_plus10 = round(mean(bill_plus10, na.rm=TRUE)),
            mean_bill_plus20 = round(mean(bill_plus20, na.rm=TRUE)),
            mean_bill_plus30 = round(mean(bill_plus30, na.rm=TRUE)),
            mean_bill_plus40 = round(mean(bill_plus40, na.rm=TRUE)),
            
            # current perceived_savings = median(tax_amt_exe),
            tr_neg10 = round(mean(tr_neg10*100), digits = 2), 
            cur_comp_TC_rate = round(mean(cur_comp_TC_rate), digits = 2),
            tr_plus10 = round(mean(tr_plus10*100), digits = 2),
            tr_plus20 = round(mean(tr_plus20*100), digits = 2),
            tr_plus30 = round(mean(tr_plus30*100), digits = 2),
            tr_plus40 = round(mean(tr_plus40*100), digits = 2),
            pincount=n(),
            zerodol_bills_ghe0 = sum(zerodol_bills_ghe0),           
            zerodol_bills_current = sum(zerodol_bills_current),
            zerodol_bills_ghe20 = sum(zerodol_bills_ghe20),
            zerodol_bills_ghe30 = sum(zerodol_bills_ghe30),
            zerodol_bills_ghe40 = sum(zerodol_bills_ghe40),
            zerodol_bills_ghe50 = sum(zerodol_bills_ghe50),


  ) %>%
  mutate(clean_name = ifelse(clean_name == "Hazelcrest", "Hazel Crest", clean_name)) %>%
  arrange(has_HO_exemp, rank)


munis_billchange <- munis_billchange %>% left_join(muni_quartiles)
munis_billchange

#write_csv(munis_billchange, "./Output/5b_muni_billchange_scenarios.csv")
```


## Case Studies in Red - 25 v 75 Percentile Homes

```{r}
#| label: fig-currenttaxsystem
#| fig-cap: "Current GHE Amount (up to 10,000 EAV exempt per property)"
#| 
ratios<- munis_billchange %>% 
  filter(has_HO_exemp == 1 &  !is.na(rank)) %>% # claimed exemption in 2021
  mutate(currbill_to_AV_25 = ifelse(rank == "q25", mean_bill_cur/median_AV, NA)) %>%
  mutate(currbill_to_AV_75 = ifelse(rank == "q75", mean_bill_cur/median_AV, NA)) %>% 
  group_by(clean_name) %>%
  summarize(currbill_to_AV_25 = max(currbill_to_AV_25, na.rm=TRUE),
            currbill_to_AV_75 = max(currbill_to_AV_75, na.rm=TRUE)) %>%
  mutate(muni_ratio_25to75 = currbill_to_AV_25/currbill_to_AV_75) %>% filter(muni_ratio_25to75 > 0.01)



ggplot(data = ratios, aes(y = currbill_to_AV_25, x = currbill_to_AV_75, label = clean_name)) + 
  geom_abline(intercept = 0, slope = 1) +
  geom_point(data = ratios, aes(alpha = .5)) + 
  ggrepel::geom_label_repel(data = (ratios %>% filter(clean_name %in% c("Park Forest", "Chicago", "Winnetka", "Glencoe", "Riverdale", "Dolton", "Markham", "Chicago Heights", "Hazel Crest", "Phoenix"))), aes(y = currbill_to_AV_25, x = currbill_to_AV_75), size = 2, max.overlaps = Inf, point.padding = 0, # additional padding around each point
    min.segment.length = 0, # draw all line segments
    )+
    geom_point(data = (ratios %>% filter(clean_name %in% c("Park Forest", "Chicago", "Winnetka", "Glencoe", "Riverdale", "Dolton", "Markham", "Chicago Heights", "Hazel Crest", "Phoenix"))), aes(y = currbill_to_AV_25, x = currbill_to_AV_75, color = "red"), size = 2)+

  theme_classic() +
    scale_y_continuous(limits = c(0, .6), expand = c(0,0))+
   scale_x_continuous(limits = c(0, .6), expand = c(0,0))+
  theme(legend.position = "none") +
  labs(title="25th & 75th Percentile AV Properties (Class 2 Only): Ratio of Tax Bill to AV", y="25th Percentile AV", x="75th Percentile AV", subtitle = "GHE = $10,000 EAV")
```

```{r casestudies-inred}
#| layout-ncol: 3
#| layout-nrow: 2
#| column: page

ratios<- munis_billchange %>% 
  filter(has_HO_exemp == 1 &  !is.na(rank)) %>% # claimed exemption in 2021
  mutate(currbill_to_AV_25 = ifelse(rank == "q25", mean_bill_neg10/median_AV, NA)) %>%
  mutate(currbill_to_AV_75 = ifelse(rank == "q75", mean_bill_neg10/median_AV, NA)) %>% 
  group_by(clean_name) %>%
  summarize(GHE_0_bill_to_AV_25 = max(currbill_to_AV_25, na.rm=TRUE),
            GHE_0_bill_to_AV_75 = max(currbill_to_AV_75, na.rm=TRUE)) %>%
  mutate(muni_ratio_25to75 = GHE_0_bill_to_AV_25/GHE_0_bill_to_AV_75) 


ggplot(data = ratios, aes(y = GHE_0_bill_to_AV_25, x = GHE_0_bill_to_AV_75, label = clean_name)) + 
  geom_abline(intercept = 0, slope = 1) +
  geom_point(aes(alpha = .5)) + 
    ggrepel::geom_label_repel(data = (ratios %>% filter(clean_name %in% c("Park Forest", "Chicago", "Winnetka", "Glencoe", "Riverdale", "Dolton", "Markham", "Chicago Heights", "Hazel Crest", "Phoenix"))), aes(y = GHE_0_bill_to_AV_25, x = GHE_0_bill_to_AV_75 + 0.1, label = clean_name, size = 2))+
    geom_point(data = (ratios %>% filter(clean_name %in% c("Park Forest", "Chicago", "Winnetka", "Glencoe", "Riverdale", "Dolton", "Markham", "Chicago Heights", "Hazel Crest", "Phoenix"))), aes(y = GHE_0_bill_to_AV_25, x = GHE_0_bill_to_AV_75, color = "red"), size = 3)+
  theme_classic() + 
    scale_y_continuous(limits = c(0, .6), expand = c(0,0))+
  scale_x_continuous(limits = c(0, .6), expand = c(0,0))+
  theme(legend.position = "none") +
  labs(title = "GHE = 0 EAV", x = "", y = "")

ratios<- munis_billchange %>% 
  filter(has_HO_exemp == 1 &  !is.na(rank)) %>% # claimed exemption in 2021
  mutate(currbill_to_AV_25 = ifelse(rank == "q25", mean_bill_cur/median_AV, NA)) %>%
  mutate(currbill_to_AV_75 = ifelse(rank == "q75", mean_bill_cur/median_AV, NA)) %>% 
  group_by(clean_name) %>%
  summarize(currbill_to_AV_25 = max(currbill_to_AV_25, na.rm=TRUE),
            currbill_to_AV_75 = max(currbill_to_AV_75, na.rm=TRUE)) %>%
  mutate(muni_ratio_25to75 = currbill_to_AV_25/currbill_to_AV_75) %>% filter(muni_ratio_25to75 > 0.01)



ggplot(data = ratios, aes(y = currbill_to_AV_25, x = currbill_to_AV_75, label = clean_name)) + 
  geom_abline(intercept = 0, slope = 1) +
  geom_point(data = ratios, aes(alpha = .5)) + 
  ggrepel::geom_label_repel(data = (ratios %>% filter(clean_name %in% c("Park Forest", "Chicago", "Winnetka", "Glencoe", "Riverdale", "Dolton", "Markham", "Chicago Heights", "Hazel Crest", "Phoenix"))), aes(y = currbill_to_AV_25, x = currbill_to_AV_75), size = 3, max.overlaps = Inf, point.padding = 0, # additional padding around each point
    min.segment.length = 0, # draw all line segments
    )+
    geom_point(data = (ratios %>% filter(clean_name %in% c("Park Forest", "Chicago", "Winnetka", "Glencoe", "Riverdale", "Dolton", "Markham", "Chicago Heights", "Hazel Crest", "Phoenix"))), aes(y = currbill_to_AV_25, x = currbill_to_AV_75, color = "red"), size = 2)+

  theme_classic() +
    scale_y_continuous(limits = c(0, .6), expand = c(0,0))+
   scale_x_continuous(limits = c(0, .6), expand = c(0,0))+
  theme(legend.position = "none") +
  labs(title = "Current GHE Amount (up to 10,000 EAV exempt per property)", x="", y="")


new_ratios<- munis_billchange %>% 
  filter(has_HO_exemp == 1 &  !is.na(rank)) %>% # claimed exemption in 2021
  mutate(newbill_to_AV_25 = ifelse(rank == "q25", mean_bill_plus10/median_AV, NA)) %>%
  mutate(newbill_to_AV_75 = ifelse(rank == "q75", mean_bill_plus10/median_AV, NA)) %>% 
  group_by(clean_name) %>%
  summarize(newbill_to_AV_25 = max(newbill_to_AV_25, na.rm=TRUE),
            newbill_to_AV_75 = max(newbill_to_AV_75, na.rm=TRUE)) %>%
  mutate(muni_ratio_25to75 = newbill_to_AV_25/newbill_to_AV_75) %>% filter(muni_ratio_25to75 > 0.01)


ggplot(data = new_ratios, aes(y = newbill_to_AV_25, x = newbill_to_AV_75, label = clean_name)) + 
  geom_abline(intercept = 0, slope = 1) +
  geom_point(data = new_ratios, aes(alpha = .5)) + 
  ggrepel::geom_label_repel(data = (new_ratios %>% filter(clean_name %in% c("Park Forest", "Chicago", "Winnetka", "Glencoe", "Riverdale", "Dolton", "Markham", "Chicago Heights", "Hazel Crest", "Phoenix"))), aes(y = newbill_to_AV_25, x = newbill_to_AV_75), size = 2, max.overlaps = Inf, point.padding = 0, # additional padding around each point
    min.segment.length = 0, # draw all line segments
    )+
    geom_point(data = (new_ratios %>% filter(clean_name %in% c("Park Forest", "Chicago", "Winnetka", "Glencoe", "Riverdale", "Dolton", "Markham", "Chicago Heights", "Hazel Crest", "Phoenix"))), aes(y = newbill_to_AV_25, x = newbill_to_AV_75, color = "red"), size = 2)+
  theme_classic() + 
  scale_y_continuous(limits = c(0, .6), expand = c(0,0))+
  scale_x_continuous(limits = c(0, .6), expand = c(0,0))+
  theme(legend.position = "none") + 
  labs(title = "Increase GHE by 10,000 EAV (for up to 20,000 EAV exempt)", x= "", y="")


new_ratios<- munis_billchange %>% 
  filter(has_HO_exemp == 1 &  !is.na(rank)) %>% # claimed exemption in 2021
  mutate(newbill_to_AV_25 = ifelse(rank == "q25", mean_bill_plus20/median_AV, NA)) %>%
  mutate(newbill_to_AV_75 = ifelse(rank == "q75", mean_bill_plus20/median_AV, NA)) %>% 
  group_by(clean_name) %>%
  summarize(newbill_to_AV_25 = max(newbill_to_AV_25, na.rm=TRUE),
            newbill_to_AV_75 = max(newbill_to_AV_75, na.rm=TRUE)) %>%
  mutate(muni_ratio_25to75 = newbill_to_AV_25/newbill_to_AV_75) %>% filter(muni_ratio_25to75 > 0.01)


ggplot(data = new_ratios, aes(y = newbill_to_AV_25, x = newbill_to_AV_75, label = clean_name)) + 
  geom_abline(intercept = 0, slope = 1) +
  geom_point(aes(alpha = .5)) + 
  ggrepel::geom_label_repel(data = (new_ratios %>% filter(clean_name %in% c("Park Forest", "Chicago", "Winnetka", "Glencoe", "Riverdale", "Dolton", "Markham", "Chicago Heights", "Hazel Crest", "Phoenix"))), aes(y = newbill_to_AV_25, x = newbill_to_AV_75), size = 2, max.overlaps = Inf, point.padding = 0, # additional padding around each point
    min.segment.length = 0, # draw all line segments
    )+
    geom_point(data = (new_ratios %>% filter(clean_name %in% c("Park Forest", "Chicago", "Winnetka", "Glencoe", "Riverdale", "Dolton", "Markham", "Chicago Heights", "Hazel Crest", "Phoenix"))), aes(y = newbill_to_AV_25, x = newbill_to_AV_75, color = "red"), size = 2)+
  theme_classic() + 
    scale_y_continuous(limits = c(0, .6), expand = c(0,0))+

  scale_x_continuous(limits = c(0, .6), expand = c(0,0))+
  theme(legend.position = "none") + 
  labs(title = "Increase GHE by 20,000 EAV (for up to 30,000 EAV exempt)",
                                        x = "", y = "")

new_ratios<- munis_billchange %>% 
  filter(has_HO_exemp == 1 &  !is.na(rank)) %>% # claimed exemption in 2021
  mutate(newbill_to_AV_25 = ifelse(rank == "q25", mean_bill_plus30/median_AV, NA)) %>%
  mutate(newbill_to_AV_75 = ifelse(rank == "q75", mean_bill_plus30/median_AV, NA)) %>% 
  group_by(clean_name) %>%
  summarize(newbill_to_AV_25 = max(newbill_to_AV_25, na.rm=TRUE),
            newbill_to_AV_75 = max(newbill_to_AV_75, na.rm=TRUE)) %>%
  mutate(muni_ratio_25to75 = newbill_to_AV_25/newbill_to_AV_75) %>% filter(muni_ratio_25to75 > 0.01)


ggplot(data = new_ratios, aes(y = newbill_to_AV_25, x = newbill_to_AV_75, label = clean_name)) + 
  geom_abline(intercept = 0, slope = 1) +
  geom_point(aes(alpha = .5)) + 
  ggrepel::geom_label_repel(data = (new_ratios %>% filter(clean_name %in% c("Park Forest", "Chicago", "Winnetka", "Glencoe", "Riverdale", "Dolton", "Markham", "Chicago Heights", "Hazel Crest", "Phoenix"))), aes(y = newbill_to_AV_25, x = newbill_to_AV_75), size = 2, max.overlaps = Inf, point.padding = 0, # additional padding around each point
    min.segment.length = 0, # draw all line segments
    )+
    geom_point(data = (new_ratios %>% filter(clean_name %in% c("Park Forest", "Chicago", "Winnetka", "Glencoe", "Riverdale", "Dolton", "Markham", "Chicago Heights", "Hazel Crest", "Phoenix"))), aes(y = newbill_to_AV_25, x = newbill_to_AV_75, color = "red"), size = 2)+
  theme_classic() + 
    scale_y_continuous(limits = c(0, .6), expand = c(0,0))+

  scale_x_continuous(limits = c(0, .6), expand = c(0,0))+
  theme(legend.position = "none") + 
  labs(title = "Increase GHE by 30,000 EAV (for up to 40,000 EAV exempt) ", 
                                     x= "", y = "")


new_ratios<- munis_billchange %>% 
  filter(has_HO_exemp == 1 &  !is.na(rank)) %>% # claimed exemption in 2021
  mutate(newbill_to_AV_25 = ifelse(rank == "q25", mean_bill_plus40/median_AV, NA)) %>%
  mutate(newbill_to_AV_75 = ifelse(rank == "q75", mean_bill_plus40/median_AV, NA)) %>% 
  group_by(clean_name) %>%
  summarize(newbill_to_AV_25 = max(newbill_to_AV_25, na.rm=TRUE),
            newbill_to_AV_75 = max(newbill_to_AV_75, na.rm=TRUE)) %>%
  mutate(muni_ratio_25to75 = newbill_to_AV_25/newbill_to_AV_75) %>% filter(muni_ratio_25to75 > 0.01)


ggplot(data = new_ratios, aes(y = newbill_to_AV_25, x = newbill_to_AV_75, label = clean_name)) + 
  geom_abline(intercept = 0, slope = 1) +
  geom_point(aes(alpha = .5)) + 
  ggrepel::geom_label_repel(data = (new_ratios %>% filter(clean_name %in% c("Park Forest", "Chicago", "Winnetka", "Glencoe", "Riverdale", "Dolton", "Markham", "Chicago Heights", "Hazel Crest", "Phoenix"))), aes(y = newbill_to_AV_25, x = newbill_to_AV_75), size = 2, max.overlaps = Inf, point.padding = 0, # additional padding around each point
    min.segment.length = 0, # draw all line segments
    )+
    geom_point(data = (new_ratios %>% filter(clean_name %in% c("Park Forest", "Chicago", "Winnetka", "Glencoe", "Riverdale", "Dolton", "Markham", "Chicago Heights", "Hazel Crest", "Phoenix"))), aes(y = newbill_to_AV_25, x = newbill_to_AV_75, color = "red"), size = 2)+
  theme_classic() + 
    scale_y_continuous(limits = c(0, .6), expand = c(0,0))+

  scale_x_continuous(limits = c(0, .6), expand = c(0,0))+
  theme(legend.position = "none") + labs(title = "Increase GHE by 40,000 EAV (for up to 50,000 EAV exempt) ", 
                                      x = "", y = "")
```





## 25 v 75 Percentile Homes


```{r}
#| layout-ncol: 3
#| layout-nrow: 2
#| column: page


ratios<- munis_billchange %>% 
  filter(has_HO_exemp == 1 &  !is.na(rank)) %>% # claimed exemption in 2021
  mutate(currbill_to_AV_25 = ifelse(rank == "q25", mean_bill_neg10/median_AV, NA)) %>%
  mutate(currbill_to_AV_75 = ifelse(rank == "q75", mean_bill_neg10/median_AV, NA)) %>% 
  group_by(clean_name) %>%
  summarize(GHE_0_bill_to_AV_25 = max(currbill_to_AV_25, na.rm=TRUE),
            GHE_0_bill_to_AV_75 = max(currbill_to_AV_75, na.rm=TRUE)) %>%
  mutate(muni_ratio_25to75 = GHE_0_bill_to_AV_25/GHE_0_bill_to_AV_75)


ggplot(data = ratios, aes(y = GHE_0_bill_to_AV_25, x = GHE_0_bill_to_AV_75, label = clean_name)) + 
  geom_abline(intercept = 0, slope = 1) +
  geom_point(aes(alpha = .5)) + 
  geom_text(nudge_x = .03, nudge_y=0.01, size = 3, check_overlap = TRUE)+ 
  theme_classic() + 
    scale_y_continuous(limits = c(0, .6))+

  scale_x_continuous(limits = c(0, .6))+
  theme(legend.position = "none") +
  labs(title = "Reduced GHE Amount by 10,000 EAV (0 EAV exempt from GHE)",
       subtitle = "Other exemptions still in place")


ratios<- munis_billchange %>% 
  filter(has_HO_exemp == 1 &  !is.na(rank)) %>% # claimed exemption in 2021
  mutate(currbill_to_AV_25 = ifelse(rank == "q25", mean_bill_cur/median_AV, NA)) %>%
  mutate(currbill_to_AV_75 = ifelse(rank == "q75", mean_bill_cur/median_AV, NA)) %>% 
  group_by(clean_name) %>%
  summarize(currbill_to_AV_25 = max(currbill_to_AV_25, na.rm=TRUE),
            currbill_to_AV_75 = max(currbill_to_AV_75, na.rm=TRUE)) %>%
  mutate(muni_ratio_25to75 = currbill_to_AV_25/currbill_to_AV_75)


ggplot(data = ratios, aes(y = currbill_to_AV_25, x = currbill_to_AV_75, label = clean_name)) + 
  geom_abline(intercept = 0, slope = 1) +
  geom_point(data = ratios, aes(alpha = .5)) + 
  geom_point(data = (ratios %>% filter(clean_name %in% c("Park Forest", "Chicago", "Winnetka", "Riverdale","Dolton"))), aes(y = currbill_to_AV_25, x = currbill_to_AV_75, color = "red"), size = 3)+
  ggrepel::geom_text_repel(data = (ratios %>% filter(clean_name %in% c("Park Forest", "Chicago", "Winnetka", "Riverdale", "Dolton"))), aes(y = currbill_to_AV_25, x = currbill_to_AV_75, label = clean_name), size = 3)+
  theme_classic() +
    scale_y_continuous(limits = c(0, .6))+


   scale_x_continuous(limits = c(0, .6))+
  theme(legend.position = "none") +
  labs(title = "Current GHE Amount (up to 10,000 EAV exempt per property)", x="", y="")

new_ratios<- munis_billchange %>% 
  filter(has_HO_exemp == 1 &  !is.na(rank)) %>% # claimed exemption in 2021
  mutate(newbill_to_AV_25 = ifelse(rank == "q25", mean_bill_plus10/median_AV, NA)) %>%
  mutate(newbill_to_AV_75 = ifelse(rank == "q75", mean_bill_plus10/median_AV, NA)) %>% 
  group_by(clean_name) %>%
  summarize(newbill_to_AV_25 = max(newbill_to_AV_25, na.rm=TRUE),
            newbill_to_AV_75 = max(newbill_to_AV_75, na.rm=TRUE)) %>%
  mutate(muni_ratio_25to75 = newbill_to_AV_25/newbill_to_AV_75)


ggplot(data = new_ratios, aes(y = newbill_to_AV_25, x = newbill_to_AV_75, label = clean_name)) + 
  geom_abline(intercept = 0, slope = 1) +
  geom_point(aes(alpha = .5)) + 
  geom_text(nudge_x = .03, nudge_y=0.01, size = 3, check_overlap = TRUE)+ 
  theme_classic() + 
  scale_y_continuous(limits = c(0, .6))+
  scale_x_continuous(limits = c(0, .6))+
  theme(legend.position = "none") + 
  labs(title = "Increase GHE by 10,000 EAV (for up to 20,000 EAV exempt) ", 
                                   y = "25th percentile of homes, taxbill:AV",
                                         x= "75th percentile of homes, taxbill:AV",
                                         caption =  "Uses single-family properties. Exempt EAV up to 20K per property")


new_ratios<- munis_billchange %>% 
  filter(has_HO_exemp == 1 &  !is.na(rank)) %>% # claimed exemption in 2021
  mutate(newbill_to_AV_25 = ifelse(rank == "q25", mean_bill_plus20/median_AV, NA)) %>%
  mutate(newbill_to_AV_75 = ifelse(rank == "q75", mean_bill_plus20/median_AV, NA)) %>% 
  group_by(clean_name) %>%
  summarize(newbill_to_AV_25 = max(newbill_to_AV_25, na.rm=TRUE),
            newbill_to_AV_75 = max(newbill_to_AV_75, na.rm=TRUE)) %>%
  mutate(muni_ratio_25to75 = newbill_to_AV_25/newbill_to_AV_75)


ggplot(data = new_ratios, aes(y = newbill_to_AV_25, x = newbill_to_AV_75, label = clean_name)) + 
  geom_abline(intercept = 0, slope = 1) +
  geom_point(aes(alpha = .5)) + 
  geom_text(nudge_x = .03, nudge_y=0.01, size = 3, check_overlap = TRUE)+ 
  theme_classic() + 
    scale_y_continuous(limits = c(0, .6))+

  scale_x_continuous(limits = c(0, .6), expand = c(0,0))+
  theme(legend.position = "none") + 
  labs(title = "Increase GHE by 20,000 EAV (for up to 30,000 EAV exempt) ",
                                         y = "25th percentile of homes, taxbill:AV",
                                         x= "75th percentile of homes, taxbill:AV",
                                         caption =  "Uses single-family properties. Exempt EAV up to 30K per property")

new_ratios<- munis_billchange %>% 
  filter(has_HO_exemp == 1 &  !is.na(rank)) %>% # claimed exemption in 2021
  mutate(newbill_to_AV_25 = ifelse(rank == "q25", mean_bill_plus30/median_AV, NA)) %>%
  mutate(newbill_to_AV_75 = ifelse(rank == "q75", mean_bill_plus30/median_AV, NA)) %>% 
  group_by(clean_name) %>%
  summarize(newbill_to_AV_25 = max(newbill_to_AV_25, na.rm=TRUE),
            newbill_to_AV_75 = max(newbill_to_AV_75, na.rm=TRUE)) %>%
  mutate(muni_ratio_25to75 = newbill_to_AV_25/newbill_to_AV_75)


ggplot(data = new_ratios, aes(y = newbill_to_AV_25, x = newbill_to_AV_75, label = clean_name)) + 
  geom_abline(intercept = 0, slope = 1) +
  geom_point(aes(alpha = .5)) + 
  geom_text(nudge_x = .03, nudge_y=0.01, size = 3, check_overlap = TRUE)+ 
  theme_classic() + 
    scale_y_continuous(limits = c(0, .6))+

  scale_x_continuous(limits = c(0, .6))+
  theme(legend.position = "none") + 
  labs(title = "Increase GHE by 30,000 EAV (for up to 40,000 EAV exempt) ", 
                                         y = "25th percentile of homes, taxbill:AV",
                                         x= "75th percentile of homes, taxbill:AV",
                                         caption =  "Uses single-family properties. Exempt EAV up to 40K per property")


new_ratios<- munis_billchange %>% 
  filter(has_HO_exemp == 1 &  !is.na(rank)) %>% # claimed exemption in 2021
  mutate(newbill_to_AV_25 = ifelse(rank == "q25", mean_bill_plus40/median_AV, NA)) %>%
  mutate(newbill_to_AV_75 = ifelse(rank == "q75", mean_bill_plus40/median_AV, NA)) %>% 
  group_by(clean_name) %>%
  summarize(newbill_to_AV_25 = max(newbill_to_AV_25, na.rm=TRUE),
            newbill_to_AV_75 = max(newbill_to_AV_75, na.rm=TRUE)) %>%
  mutate(muni_ratio_25to75 = newbill_to_AV_25/newbill_to_AV_75)


ggplot(data = new_ratios, aes(y = newbill_to_AV_25, x = newbill_to_AV_75, label = clean_name)) + 
  geom_abline(intercept = 0, slope = 1) +
  geom_point(aes(alpha = .5)) + 
  geom_text(nudge_x = .03, nudge_y=0.01, size = 3, check_overlap = TRUE)+ 
  theme_classic() + 
    scale_y_continuous(limits = c(0, .6))+

  scale_x_continuous(limits = c(0, .6))+
  theme(legend.position = "none") + labs(title = "Increase GHE by 40,000 EAV (for up to 50,000 EAV exempt) ", 
                                         y = "25th percentile of homes, taxbill:AV",
                                         x= "75th percentile of homes, taxbill:AV",
                                         caption =  "Uses single-family properties. Exempt EAV up to 50K per property")
```












## 90/10 deciles

```{r}
q = c(.1, .5, .9)

## ranks properties that are considered single family homes in order of AV for each Muni
muni_quartiles <- joined_pins %>%
  filter(Option2 == "Single-Family") %>% 
  filter(tax_code %in% muni_TC_fullyCook$tax_code_num) %>%
  group_by(agency_name, clean_name) %>%
  arrange(av) %>%
  summarize(count_pins = n(), 
            min = min(av),
            quant10 = round(quantile(av, probs = q[1])), 
            quant50 = round(quantile(av, probs = q[2])),
            quant90 = round(quantile(av, probs = q[3])),
            max = max(av)
           ) %>% 
  arrange( desc( quant50))
muni_quartiles
```


```{r}
## create rank variable for properties that fall within the quartiles +/- $500 range
munis_ranked <- joined_pins  %>%
  inner_join(muni_quartiles, by = c("agency_name", "clean_name")) %>% 
  mutate(rank = case_when(
    av > (quant10-500) & (av<quant10+500) ~ "q10",
    av > (quant50-500) & (av<quant50+500) ~ "q50",
    av > (quant90-500) & (av<quant90+500) ~ "q90")
    ) %>%
  select(clean_name, rank, av, pin, class, everything()) %>%
  left_join(nicknames)




munis_billchange <-  munis_ranked %>% 
  group_by(clean_name, rank) %>%
  left_join(scenario_taxrates) %>%
  arrange(av) %>%
 # group_by(agency_name, has_HO_exemp) %>% 
  mutate(#taxable_eav = final_tax_to_dist / tax_code_rate,
    # current bill = current tax rate * portion of levy billed
         bill_neg10 = tr_neg10*(equalized_AV-all_exemptions+ exe_homeowner -exe_neg10),

         bill_current = cur_comp_TC_rate/100*(equalized_AV-all_exemptions),
         bill_plus10 =  tr_plus10*(equalized_AV-all_exemptions+ exe_homeowner -exe_plus10),
         bill_plus20 = tr_plus20*(equalized_AV-all_exemptions+ exe_homeowner -exe_plus20),
         bill_plus30 = tr_plus30*(equalized_AV-all_exemptions+ exe_homeowner -exe_plus30),
         bill_plus40 = tr_plus40*(equalized_AV-all_exemptions+ exe_homeowner -exe_plus40),

    # ## make bills $1 if they had been negative.    
         bill_neg10 = ifelse(bill_neg10 < 1, 1, bill_neg10),
         bill_current = ifelse(bill_current < 1, 1, bill_current),
         bill_plus10 = ifelse(bill_plus10 < 1, 1, bill_plus10),
         bill_plus20 = ifelse(bill_plus20 < 1, 1, bill_plus20),
         bill_plus30 = ifelse(bill_plus30 < 1, 1, bill_plus30),
         bill_plus40 = ifelse(bill_plus40 < 1, 1, bill_plus40)) %>%

  mutate(
         zerodol_bills_ghe0 = ifelse(bill_neg10 < 5, 1, 0),
         zerodol_bills_current = ifelse(bill_current < 5, 1, 0),
         zerodol_bills_ghe20 = ifelse(bill_plus10 < 5, 1, 0),
         zerodol_bills_ghe30 = ifelse(bill_plus20 < 5, 1, 0),
         zerodol_bills_ghe40 = ifelse(bill_plus30 < 5, 1, 0),
         zerodol_bills_ghe50 = ifelse(bill_plus40 < 5, 1, 0),
  ) %>%
  group_by(clean_name, rank, has_HO_exemp) %>% 
  summarize(median_AV = round(median(av)),
            median_EAV = round(median(eav)),
            mean_bill_neg10 = round(mean(bill_neg10, na.rm=TRUE)),
            mean_bill_cur = round(mean(bill_current, na.rm=TRUE)),
            mean_bill_plus10 = round(mean(bill_plus10, na.rm=TRUE)),
            mean_bill_plus20 = round(mean(bill_plus20, na.rm=TRUE)),
            mean_bill_plus30 = round(mean(bill_plus30, na.rm=TRUE)),
            mean_bill_plus40 = round(mean(bill_plus40, na.rm=TRUE)),
            
            # current perceived_savings = median(tax_amt_exe),
            tr_neg10 = round(mean(tr_neg10*100), digits = 2), 
            cur_comp_TC_rate = round(mean(cur_comp_TC_rate), digits = 2),
            tr_plus10 = round(mean(tr_plus10*100), digits = 2),
            tr_plus20 = round(mean(tr_plus20*100), digits = 2),
            tr_plus30 = round(mean(tr_plus30*100), digits = 2),
            tr_plus40 = round(mean(tr_plus40*100), digits = 2),
            pincount=n()
  ) %>%
  arrange(has_HO_exemp, rank)


munis_billchange <- munis_billchange %>% left_join(muni_quartiles)
munis_billchange
```

```{r}
#| label: progressivity-90to10
#| layout-ncol: 3
#| layout-nrow: 2
#| column: page

ratios<- munis_billchange %>% 
  filter(has_HO_exemp == 1 &  !is.na(rank)) %>% # claimed exemption in 2021
  mutate(currbill_to_AV_10 = ifelse(rank == "q10", mean_bill_neg10/median_AV, NA)) %>%
  mutate(currbill_to_AV_90 = ifelse(rank == "q90", mean_bill_neg10/median_AV, NA)) %>% 
  group_by(clean_name) %>%
  summarize(GHE_0_bill_to_AV_10 = max(currbill_to_AV_10, na.rm=TRUE),
            GHE_0_bill_to_AV_90 = max(currbill_to_AV_90, na.rm=TRUE)) %>%
  mutate(muni_ratio_10to90 = GHE_0_bill_to_AV_10/GHE_0_bill_to_AV_90)


ggplot(data = ratios, aes(y = GHE_0_bill_to_AV_10, x = GHE_0_bill_to_AV_90, label = clean_name)) + 
  geom_abline(intercept = 0, slope = 1) +
  geom_point(aes(alpha = .5)) + 
  geom_text(nudge_x = .03, nudge_y=0.01, size = 3, check_overlap = TRUE)+ 
  theme_classic() + 
    scale_y_continuous(limits = c(0, .6))+

  scale_x_continuous(limits = c(0, .6))+
  theme(legend.position = "none") +
  labs(title = "Reduced GHE Amount by 10,000 EAV (0 EAV exempt from GHE)",
       subtitle = "Other exemptions still in place")


ratios<- munis_billchange %>% 
  filter(has_HO_exemp == 1 &  !is.na(rank)) %>% # claimed exemption in 2021
  mutate(currbill_to_AV_10 = ifelse(rank == "q10", mean_bill_cur/median_AV, NA)) %>%
  mutate(currbill_to_AV_90 = ifelse(rank == "q90", mean_bill_cur/median_AV, NA)) %>% 
  group_by(clean_name) %>%
  summarize(currbill_to_AV_10 = max(currbill_to_AV_10, na.rm=TRUE),
            currbill_to_AV_90 = max(currbill_to_AV_90, na.rm=TRUE)) %>%
  mutate(muni_ratio_10to90 = currbill_to_AV_10/currbill_to_AV_90)


ggplot(data = ratios, aes(y = currbill_to_AV_10, x = currbill_to_AV_90, label = clean_name)) + 
  geom_abline(intercept = 0, slope = 1) +
  geom_point(aes(alpha = .5)) + 
  geom_text(nudge_x = .03, nudge_y=0.01, size = 3, check_overlap = TRUE)+ 
  theme_classic() + 
    scale_y_continuous(limits = c(0, .6))+

  scale_x_continuous(limits = c(0, .6))+
  theme(legend.position = "none") +
  labs(title = "Current GHE Amount (up to 10,000 EAV exempt per property)")


new_ratios<- munis_billchange %>% 
  filter(has_HO_exemp == 1 &  !is.na(rank)) %>% # claimed exemption in 2021
  mutate(newbill_to_AV_10 = ifelse(rank == "q10", mean_bill_plus10/median_AV, NA)) %>%
  mutate(newbill_to_AV_90 = ifelse(rank == "q90", mean_bill_plus10/median_AV, NA)) %>% 
  group_by(clean_name) %>%
  summarize(newbill_to_AV_10 = max(newbill_to_AV_10, na.rm=TRUE),
            newbill_to_AV_90 = max(newbill_to_AV_90, na.rm=TRUE)) %>%
  mutate(muni_ratio_10to90 = newbill_to_AV_10/newbill_to_AV_90)


ggplot(data = new_ratios, aes(y = newbill_to_AV_10, x = newbill_to_AV_90, label = clean_name)) + 
  geom_abline(intercept = 0, slope = 1) +
  geom_point(aes(alpha = .5)) + 
  geom_text(nudge_x = .03, nudge_y=0.01, size = 3, check_overlap = TRUE)+ 
  theme_classic() + 
  scale_y_continuous(limits = c(0, .6))+
  scale_x_continuous(limits = c(0, .6))+
  theme(legend.position = "none") + labs(title = "Increase GHE by 10,000 EAV (for up to 20,000 EAV exempt) ", 
                                         y = "10th percentile of homes, taxbill:AV",
                                         x= "90th percentile of homes, taxbill:AV",
                                         caption =  "Uses single-family properties. Exempt EAV up to 20K per property")


new_ratios<- munis_billchange %>% 
  filter(has_HO_exemp == 1 &  !is.na(rank)) %>% # claimed exemption in 2021
  mutate(newbill_to_AV_10 = ifelse(rank == "q10", mean_bill_plus20/median_AV, NA)) %>%
  mutate(newbill_to_AV_90 = ifelse(rank == "q90", mean_bill_plus20/median_AV, NA)) %>% 
  group_by(clean_name) %>%
  summarize(newbill_to_AV_10 = max(newbill_to_AV_10, na.rm=TRUE),
            newbill_to_AV_90 = max(newbill_to_AV_90, na.rm=TRUE)) %>%
  mutate(muni_ratio_10to90 = newbill_to_AV_10/newbill_to_AV_90)


ggplot(data = new_ratios, aes(y = newbill_to_AV_10, x = newbill_to_AV_90, label = clean_name)) + 
  geom_abline(intercept = 0, slope = 1) +
  geom_point(aes(alpha = .5)) + 
  geom_text(nudge_x = .03, nudge_y=0.01, size = 3, check_overlap = TRUE)+ 
  theme_classic() + 
    scale_y_continuous(limits = c(0, .6))+

  scale_x_continuous(limits = c(0, .6))+
  theme(legend.position = "none") + labs(title = "Increase GHE by 20,000 EAV (for up to 30,000 EAV exempt) ", 
                                         y = "10th percentile of homes, taxbill:AV",
                                         x= "90th percentile of homes, taxbill:AV",
                                         caption =  "Uses single-family properties. Exempt EAV up to 30K per property")

new_ratios<- munis_billchange %>% 
  filter(has_HO_exemp == 1 &  !is.na(rank)) %>% # claimed exemption in 2021
  mutate(newbill_to_AV_10 = ifelse(rank == "q10", mean_bill_plus30/median_AV, NA)) %>%
  mutate(newbill_to_AV_90 = ifelse(rank == "q90", mean_bill_plus30/median_AV, NA)) %>% 
  group_by(clean_name) %>%
  summarize(newbill_to_AV_10 = max(newbill_to_AV_10, na.rm=TRUE),
            newbill_to_AV_90 = max(newbill_to_AV_90, na.rm=TRUE)) %>%
  mutate(muni_ratio_10to90 = newbill_to_AV_10/newbill_to_AV_90)


ggplot(data = new_ratios, aes(y = newbill_to_AV_10, x = newbill_to_AV_90, label = clean_name)) + 
  geom_abline(intercept = 0, slope = 1) +
  geom_point(aes(alpha = .5)) + 
  geom_text(nudge_x = .03, nudge_y=0.01, size = 3, check_overlap = TRUE)+ 
  theme_classic() + 
    scale_y_continuous(limits = c(0, .6))+

  scale_x_continuous(limits = c(0, .6))+
  theme(legend.position = "none") + labs(title = "Increase GHE by 30,000 EAV (for up to 40,000 EAV exempt) ", 
                                         y = "10th percentile of homes, taxbill:AV",
                                         x= "90th percentile of homes, taxbill:AV",
                                         caption =  "Uses single-family properties. Exempt EAV up to 40K per property")


new_ratios<- munis_billchange %>% 
  filter(has_HO_exemp == 1 &  !is.na(rank)) %>% # claimed exemption in 2021
  mutate(newbill_to_AV_10 = ifelse(rank == "q10", mean_bill_plus40/median_AV, NA)) %>%
  mutate(newbill_to_AV_90 = ifelse(rank == "q90", mean_bill_plus40/median_AV, NA)) %>% 
  group_by(clean_name) %>%
  summarize(newbill_to_AV_10 = max(newbill_to_AV_10, na.rm=TRUE),
            newbill_to_AV_90 = max(newbill_to_AV_90, na.rm=TRUE)) %>%
  mutate(muni_ratio_10to90 = newbill_to_AV_10/newbill_to_AV_90)


ggplot(data = new_ratios, aes(y = newbill_to_AV_10, x = newbill_to_AV_90, label = clean_name)) + 
  geom_abline(intercept = 0, slope = 1) +
  geom_point(aes(alpha = .5)) + 
  geom_text(nudge_x = .03, nudge_y=0.01, size = 3, check_overlap = TRUE)+ 
  theme_classic() + 
    scale_y_continuous(limits = c(0, .6))+

  scale_x_continuous(limits = c(0, .6))+
  theme(legend.position = "none") + labs(title = "Increase GHE by 40,000 EAV (for up to 50,000 EAV exempt) ", 
                                         y = "10th percentile of homes, taxbill:AV",
                                         x= "90th percentile of homes, taxbill:AV",
                                         caption =  "Uses single-family properties. Exempt EAV up to 50K per property")
```






### 90/10 deciles - Bigger AV range

Increased range of homes included in quantile to avoid having municipalities drop out if there were no tax bills. Not sure if it addresses the problem. 

```{r}


q = c(.1, .5, .9)

## ranks properties that are considered single family homes in order of AV for each Muni
muni_quartiles <- joined_pins %>%
  filter(Option2 == "Single-Family") %>% 
  filter(tax_code %in% muni_TC_fullyCook$tax_code_num) %>%
  group_by(agency_name, clean_name) %>%
  arrange(av) %>%
  summarize(count_pins = n(), 
            min = min(av),
            quant10 = round(quantile(av, probs = q[1])), 
            quant50 = round(quantile(av, probs = q[2])),
            quant90 = round(quantile(av, probs = q[3])),
            max = max(av)
           ) %>% 
  arrange( desc( quant50))
muni_quartiles
```


```{r}
## create rank variable for properties that fall within the quartiles +/- $500 range
munis_ranked <- joined_pins  %>%
  inner_join(muni_quartiles, by = c("agency_name", "clean_name")) %>% 
  mutate(rank = case_when(
    av > (quant10-1000) & (av<quant10+1000) ~ "q10",
    av > (quant50-1000) & (av<quant50+1000) ~ "q50",
    av > (quant90-1000) & (av<quant90+1000) ~ "q90")
    ) %>%
  select(clean_name, rank, av, pin, class, everything()) %>%
  left_join(nicknames)




munis_billchange <-  munis_ranked %>% 
  group_by(clean_name, rank) %>%
  left_join(scenario_taxrates) %>%
  arrange(av) %>%
  mutate(#taxable_eav = final_tax_to_dist / tax_code_rate,
    # current bill = current tax rate * portion of levy billed
         bill_neg10 = tr_neg10*(equalized_AV-all_exemptions+ exe_homeowner -exe_neg10),

         bill_current = cur_comp_TC_rate/100*(equalized_AV-all_exemptions),
         bill_plus10 =  tr_plus10*(equalized_AV-all_exemptions+ exe_homeowner -exe_plus10),
         bill_plus20 = tr_plus20*(equalized_AV-all_exemptions+ exe_homeowner -exe_plus20),
         bill_plus30 = tr_plus30*(equalized_AV-all_exemptions+ exe_homeowner -exe_plus30),
         bill_plus40 = tr_plus40*(equalized_AV-all_exemptions+ exe_homeowner -exe_plus40),

    # ## make bills $1 if they had been negative.    
         bill_neg10 = ifelse(bill_neg10 < 1, 1, bill_neg10),
         bill_current = ifelse(bill_current < 1, 1, bill_current),
         bill_plus10 = ifelse(bill_plus10 < 1, 1, bill_plus10),
         bill_plus20 = ifelse(bill_plus20 < 1, 1, bill_plus20),
         bill_plus30 = ifelse(bill_plus30 < 1, 1, bill_plus30),
         bill_plus40 = ifelse(bill_plus40 < 1, 1, bill_plus40)) %>%

  mutate(
         zerodol_bills_ghe0 = ifelse(bill_neg10 < 5, 1, 0),
         zerodol_bills_current = ifelse(bill_current < 5, 1, 0),
         zerodol_bills_ghe20 = ifelse(bill_plus10 < 5, 1, 0),
         zerodol_bills_ghe30 = ifelse(bill_plus20 < 5, 1, 0),
         zerodol_bills_ghe40 = ifelse(bill_plus30 < 5, 1, 0),
         zerodol_bills_ghe50 = ifelse(bill_plus40 < 5, 1, 0),
  ) %>%
  group_by(clean_name, rank, has_HO_exemp) %>% 
  summarize(median_AV = round(median(av)),
            median_EAV = round(median(eav)),
            mean_bill_neg10 = round(mean(bill_neg10, na.rm=TRUE)),
            mean_bill_cur = round(mean(bill_current, na.rm=TRUE)),
            mean_bill_plus10 = round(mean(bill_plus10, na.rm=TRUE)),
            mean_bill_plus20 = round(mean(bill_plus20, na.rm=TRUE)),
            mean_bill_plus30 = round(mean(bill_plus30, na.rm=TRUE)),
            mean_bill_plus40 = round(mean(bill_plus40, na.rm=TRUE)),
            
            # current perceived_savings = median(tax_amt_exe),
            tr_neg10 = round(mean(tr_neg10), digits = 2), 
            cur_comp_TC_rate = round(mean(cur_comp_TC_rate), digits = 2),
            tr_plus10 = round(mean(tr_plus10*100), digits = 2),
            tr_plus20 = round(mean(tr_plus20*100), digits = 2),
            tr_plus30 = round(mean(tr_plus30*100), digits = 2),
            tr_plus40 = round(mean(tr_plus40*100), digits = 2),
            pincount=n()
  ) %>%
  arrange(has_HO_exemp, rank)


munis_billchange <- munis_billchange %>% left_join(muni_quartiles)
munis_billchange
```

```{r}
#| label: progressivity-90to10-expanded
#| layout-ncol: 3
#| layout-nrow: 2
#| column: page


ratios<- munis_billchange %>% 
  filter(has_HO_exemp == 1 &  !is.na(rank)) %>% # claimed exemption in 2021
  mutate(currbill_to_AV_10 = ifelse(rank == "q10", mean_bill_neg10/median_AV, NA)) %>%
  mutate(currbill_to_AV_90 = ifelse(rank == "q90", mean_bill_neg10/median_AV, NA)) %>% 
  group_by(clean_name) %>%
  summarize(GHE_0_bill_to_AV_10 = max(currbill_to_AV_10, na.rm=TRUE),
            GHE_0_bill_to_AV_90 = max(currbill_to_AV_90, na.rm=TRUE)) %>%
  mutate(muni_ratio_10to90 = GHE_0_bill_to_AV_10/GHE_0_bill_to_AV_90)


ggplot(data = ratios, aes(y = GHE_0_bill_to_AV_10, x = GHE_0_bill_to_AV_90, label = clean_name)) + 
  geom_abline(intercept = 0, slope = 1) +
  geom_point(aes(alpha = .5)) + 
  geom_text(nudge_x = .03, nudge_y=0.01, size = 3, check_overlap = TRUE)+ 
  theme_classic() + 
    scale_y_continuous(limits = c(0, .6))+

  scale_x_continuous(limits = c(0, .6))+
  theme(legend.position = "none") +
  labs(title = "Reduced GHE Amount by 10,000 EAV (0 EAV exempt from GHE)",
       subtitle = "Other exemptions still in place")


ratios<- munis_billchange %>% 
  filter(has_HO_exemp == 1 &  !is.na(rank)) %>% # claimed exemption in 2021
  mutate(currbill_to_AV_10 = ifelse(rank == "q10", mean_bill_cur/median_AV, NA)) %>%
  mutate(currbill_to_AV_90 = ifelse(rank == "q90", mean_bill_cur/median_AV, NA)) %>% 
  group_by(clean_name) %>%
  summarize(currbill_to_AV_10 = max(currbill_to_AV_10, na.rm=TRUE),
            currbill_to_AV_90 = max(currbill_to_AV_90, na.rm=TRUE)) %>%
  mutate(muni_ratio_10to90 = currbill_to_AV_10/currbill_to_AV_90)


ggplot(data = ratios, aes(y = currbill_to_AV_10, x = currbill_to_AV_90, label = clean_name)) + 
  geom_abline(intercept = 0, slope = 1) +
  geom_point(aes(alpha = .5)) + 
  geom_text(nudge_x = .03, nudge_y=0.01, size = 3, check_overlap = TRUE)+ 
  theme_classic() + 
    scale_y_continuous(limits = c(0, .6))+

  scale_x_continuous(limits = c(0, .6))+
  theme(legend.position = "none") +
  labs(title = "Current GHE Amount (up to 10,000 EAV exempt per property)")
new_ratios<- munis_billchange %>% 
  filter(has_HO_exemp == 1 &  !is.na(rank)) %>% # claimed exemption in 2021
  mutate(newbill_to_AV_10 = ifelse(rank == "q10", mean_bill_plus10/median_AV, NA)) %>%
  mutate(newbill_to_AV_90 = ifelse(rank == "q90", mean_bill_plus10/median_AV, NA)) %>% 
  group_by(clean_name) %>%
  summarize(newbill_to_AV_10 = max(newbill_to_AV_10, na.rm=TRUE),
            newbill_to_AV_90 = max(newbill_to_AV_90, na.rm=TRUE)) %>%
  mutate(muni_ratio_10to90 = newbill_to_AV_10/newbill_to_AV_90)


ggplot(data = new_ratios, aes(y = newbill_to_AV_10, x = newbill_to_AV_90, label = clean_name)) + 
  geom_abline(intercept = 0, slope = 1) +
  geom_point(aes(alpha = .5)) + 
  geom_text(nudge_x = .03, nudge_y=0.01, size = 3, check_overlap = TRUE)+ 
  theme_classic() + 
  scale_y_continuous(limits = c(0, .6))+
  scale_x_continuous(limits = c(0, .6))+
  theme(legend.position = "none") + labs(title = "Increase GHE by 10,000 EAV (for up to 20,000 EAV exempt) ", 
                                         y = "10th percentile of homes, taxbill:AV",
                                         x= "90th percentile of homes, taxbill:AV",
                                         caption =  "Uses single-family properties. Exempt EAV up to 20K per property")


new_ratios<- munis_billchange %>% 
  filter(has_HO_exemp == 1 &  !is.na(rank)) %>% # claimed exemption in 2021
  mutate(newbill_to_AV_10 = ifelse(rank == "q10", mean_bill_plus20/median_AV, NA)) %>%
  mutate(newbill_to_AV_90 = ifelse(rank == "q90", mean_bill_plus20/median_AV, NA)) %>% 
  group_by(clean_name) %>%
  summarize(newbill_to_AV_10 = max(newbill_to_AV_10, na.rm=TRUE),
            newbill_to_AV_90 = max(newbill_to_AV_90, na.rm=TRUE)) %>%
  mutate(muni_ratio_10to90 = newbill_to_AV_10/newbill_to_AV_90)


ggplot(data = new_ratios, aes(y = newbill_to_AV_10, x = newbill_to_AV_90, label = clean_name)) + 
  geom_abline(intercept = 0, slope = 1) +
  geom_point(aes(alpha = .5)) + 
  geom_text(nudge_x = .03, nudge_y=0.01, size = 3, check_overlap = TRUE)+ 
  theme_classic() + 
    scale_y_continuous(limits = c(0, .6))+

  scale_x_continuous(limits = c(0, .6))+
  theme(legend.position = "none") + labs(title = "Increase GHE by 20,000 EAV (for up to 30,000 EAV exempt) ", 
                                         y = "10th percentile of homes, taxbill:AV",
                                         x= "90th percentile of homes, taxbill:AV",
                                         caption =  "Uses single-family properties. Exempt EAV up to 30K per property")

new_ratios<- munis_billchange %>% 
  filter(has_HO_exemp == 1 &  !is.na(rank)) %>% # claimed exemption in 2021
  mutate(newbill_to_AV_10 = ifelse(rank == "q10", mean_bill_plus30/median_AV, NA)) %>%
  mutate(newbill_to_AV_90 = ifelse(rank == "q90", mean_bill_plus30/median_AV, NA)) %>% 
  group_by(clean_name) %>%
  summarize(newbill_to_AV_10 = max(newbill_to_AV_10, na.rm=TRUE),
            newbill_to_AV_90 = max(newbill_to_AV_90, na.rm=TRUE)) %>%
  mutate(muni_ratio_10to90 = newbill_to_AV_10/newbill_to_AV_90)


ggplot(data = new_ratios, aes(y = newbill_to_AV_10, x = newbill_to_AV_90, label = clean_name)) + 
  geom_abline(intercept = 0, slope = 1) +
  geom_point(aes(alpha = .5)) + 
  geom_text(nudge_x = .03, nudge_y=0.01, size = 3, check_overlap = TRUE)+ 
  theme_classic() + 
    scale_y_continuous(limits = c(0, .6))+

  scale_x_continuous(limits = c(0, .6))+
  theme(legend.position = "none") + labs(title = "Increase GHE by 30,000 EAV (for up to 40,000 EAV exempt) ", 
                                         y = "10th percentile of homes, taxbill:AV",
                                         x= "90th percentile of homes, taxbill:AV",
                                         caption =  "Uses single-family properties. Exempt EAV up to 40K per property")


new_ratios<- munis_billchange %>% 
  filter(has_HO_exemp == 1 &  !is.na(rank)) %>% # claimed exemption in 2021
  mutate(newbill_to_AV_10 = ifelse(rank == "q10", mean_bill_plus40/median_AV, NA)) %>%
  mutate(newbill_to_AV_90 = ifelse(rank == "q90", mean_bill_plus40/median_AV, NA)) %>% 
  group_by(clean_name) %>%
  summarize(newbill_to_AV_10 = max(newbill_to_AV_10, na.rm=TRUE),
            newbill_to_AV_90 = max(newbill_to_AV_90, na.rm=TRUE)) %>%
  mutate(muni_ratio_10to90 = newbill_to_AV_10/newbill_to_AV_90)


ggplot(data = new_ratios, aes(y = newbill_to_AV_10, x = newbill_to_AV_90, label = clean_name)) + 
  geom_abline(intercept = 0, slope = 1) +
  geom_point(aes(alpha = .5)) + 
  geom_text(nudge_x = .03, nudge_y=0.01, size = 3, check_overlap = TRUE)+ 
  theme_classic() + 
    scale_y_continuous(limits = c(0, .6))+

  scale_x_continuous(limits = c(0, .6))+
  theme(legend.position = "none") + labs(title = "Increase GHE by 40,000 EAV (for up to 50,000 EAV exempt) ", 
                                         y = "10th percentile of homes, taxbill:AV",
                                         x= "90th percentile of homes, taxbill:AV",
                                         caption =  "Uses single-family properties. Exempt EAV up to 50K per property")
```


# Combined tables
<!--- Combined tables below --->


```{r}
q = c(.10, .25, .5, .75, .9, .99)

## ranks properties that are considered single family homes in order of AV for each Muni
muni_quartiles <- joined_pins %>%
  filter(Option2 == "Single-Family") %>% 
  filter(tax_code %in% muni_TC_fullyCook$tax_code_num) %>%
  group_by(agency_name, clean_name) %>%
  arrange(av) %>%
  summarize(count_pins = n(), 
            min = min(av),
            quant10 = round(quantile(av, probs = q[1])), 

            quant25 = round(quantile(av, probs = q[2])), 
            quant50 = round(quantile(av, probs = q[3])),
            quant75 = round(quantile(av, probs = q[4])),
            quant90 = round(quantile(av, probs = q[5])),
            quant99 = round(quantile(av, probs = q[6])),

            max = max(av)
           ) %>% 
  arrange( desc( quant50))

muni_quartiles


## create rank variable for properties that fall within the quartiles +/- $500 range
munis_ranked <- joined_pins  %>%
  inner_join(muni_quartiles, by = c("agency_name", "clean_name")) %>% 
  mutate(rank = case_when(
    av > (quant10-500) & (av<quant10+500) ~ "q10",
    av > (quant25-500) & (av<quant25+500) ~ "q25",
    av > (quant50-500) & (av<quant50+500) ~ "q50",
    av > (quant75-500) & (av<quant75+500) ~ "q75",
    av > (quant90-500) & (av<quant90+500) ~ "q90",
    av > (quant99-500) & (av<quant99+500) ~ "q99")

    ) %>%
  select(clean_name, rank, av, pin, class, everything()) %>%
  left_join(nicknames)




munis_billchange <-  munis_ranked %>% 
  group_by(clean_name, rank) %>%
  left_join(scenario_taxrates) %>%
  arrange(av) %>%

    mutate(
   # ## Made negative tax bills!! ## #
         
         bill_neg10 = tr_neg10*(equalized_AV-all_exemptions+ exe_homeowner -exe_neg10),
         bill_current = cur_comp_TC_rate/100*(equalized_AV-all_exemptions),
         bill_plus10 =  tr_plus10*(equalized_AV-all_exemptions+ exe_homeowner -exe_plus10),
         bill_plus20 = tr_plus20*(equalized_AV-all_exemptions+ exe_homeowner -exe_plus20),
         bill_plus30 = tr_plus20*(equalized_AV-all_exemptions+ exe_homeowner -exe_plus30),
         bill_plus40 = tr_plus20*(equalized_AV-all_exemptions+ exe_homeowner -exe_plus40),
    
    # ## make bills $1 if they had been negative.    
         bill_neg10 = ifelse(bill_neg10 < 1, 1, bill_neg10),
         bill_current = ifelse(bill_current < 1, 1, bill_current),
         bill_plus10 = ifelse(bill_plus10 < 1, 1, bill_plus10),
         bill_plus20 = ifelse(bill_plus20 < 1, 1, bill_plus20),
         bill_plus30 = ifelse(bill_plus30 < 1, 1, bill_plus30),
         bill_plus40 = ifelse(bill_plus40 < 1, 1, bill_plus40)
         )%>%
  mutate(
         zerodol_bills_ghe0 = ifelse(bill_neg10 < 5, 1, 0),
         zerodol_bills_current = ifelse(bill_current < 5, 1, 0),
         zerodol_bills_ghe20 = ifelse(bill_plus10 < 5, 1, 0),
         zerodol_bills_ghe30 = ifelse(bill_plus20 < 5, 1, 0),
         zerodol_bills_ghe40 = ifelse(bill_plus30 < 5, 1, 0),
         zerodol_bills_ghe50 = ifelse(bill_plus40 < 5, 1, 0)
  ) %>%
  group_by(clean_name, rank, has_HO_exemp) %>% 
  summarize(median_AV = round(median(av)),
            median_EAV = round(median(eav)),
            mean_bill_neg10 = round(mean(bill_neg10, na.rm=TRUE)),
            mean_bill_cur = round(mean(bill_current, na.rm=TRUE)),
            mean_bill_plus10 = round(mean(bill_plus10, na.rm=TRUE)),
            mean_bill_plus20 = round(mean(bill_plus20, na.rm=TRUE)),
            mean_bill_plus30 = round(mean(bill_plus30, na.rm=TRUE)),
            mean_bill_plus40 = round(mean(bill_plus40, na.rm=TRUE)),
            
            # current perceived_savings = median(tax_amt_exe),
            tr_neg10 = round(mean(tr_neg10*100), digits = 2), 
            cur_comp_TC_rate = round(mean(cur_comp_TC_rate), digits = 2),
            tr_plus10 = round(mean(tr_plus10*100), digits = 2),
            tr_plus20 = round(mean(tr_plus20*100), digits = 2),
            tr_plus30 = round(mean(tr_plus30*100), digits = 2),
            tr_plus40 = round(mean(tr_plus40*100), digits = 2),
            pincount=n(),
            zerodol_bills_ghe0 = sum(zerodol_bills_ghe0),           
            zerodol_bills_current = sum(zerodol_bills_current),
            zerodol_bills_ghe20 = sum(zerodol_bills_ghe20),
            zerodol_bills_ghe30 = sum(zerodol_bills_ghe30),
            zerodol_bills_ghe40 = sum(zerodol_bills_ghe40),
            zerodol_bills_ghe50 = sum(zerodol_bills_ghe50),


  ) %>%
  arrange(has_HO_exemp, rank)


munis_billchange <- munis_billchange %>% left_join(muni_quartiles)

munis_billchange 

write_csv(munis_billchange, "5b_munis_billchange_1025507590.csv")
```
