==> old_versions/6_TIFs.Rmd <==
---
title: "TIFs in Munis"
author: "Alea Wilbur"
date: "`r Sys.Date()`"
output:
  html_document:
    toc: yes
    toc_float: yes
    df_print: paged
---

```{r setup, message = FALSE}
knitr::opts_chunk$set(echo = TRUE, warning = FALSE, message = FALSE, error=TRUE)


library(tidyverse)
library(DBI)
library(data.table)
library(ggspatial)
library(gstat)
library(here)
library(httr)
library(jsonlite)
library(ptaxsim)
library(sf)
library(stars)
library(glue)


#remotes::install_gitlab("ccao-data-science---modeling/packages/ptaxsim")

#renv::install("gitlab::ccao-data-science---modeling/packages/ptaxsim")


# Create the DB connection with the default name expected by PTAXSIM functions
ptaxsim_db_conn <- DBI::dbConnect(RSQLite::SQLite(), "./ptaxsim.db/ptaxsim-2021.0.4.db")

# has all potential property classes for pins
# downloaded from CCAO gitlab website
## I used this to merge additional information to the pins and class data later on.

#options(digits=4, scipen = 999)
options(scipen = 999)


```

# Multiple at a time

## Pulling Muni data tables

Pulling all taxing agencies directly related to Community1...Community6. Includes TIFs, SSAs, Library, and Village of the municiple agency (what we want).

Then we use the agency number for the the Village to pull all unique pins in the municipality, and plug that into tax_bill() for historic data of tax bills within the municipality.

Also create object for unique tax codes, and unique pins to plug into lookup() commands later.


```{r agency-dt}
# has EAV values, extensions by agency_num

agency_dt <- DBI::dbGetQuery(
  ptaxsim_db_conn,
  "SELECT *
  FROM agency
  "
)

agency_names <- DBI::dbGetQuery(
  ptaxsim_db_conn,
  "SELECT DISTINCT agency_num, agency_name, minor_type
  FROM agency_info

  "
  )

tif_distribution <- DBI::dbGetQuery(
  ptaxsim_db_conn,
  "SELECT *
  FROM tif_distribution

  "
  )


TIFs <- DBI::dbGetQuery(
  ptaxsim_db_conn,
  "SELECT DISTINCT *
  FROM agency_info
  WHERE minor_type = 'TIF'
  "
  ) %>%
  mutate(first6 = str_sub(agency_num,1,6),
         first5 = str_sub(agency_num,1,5)
         )


tif_distribution <- left_join(tif_distribution, TIFs)


# TIFs <- DBI::dbGetQuery(
#   ptaxsim_db_conn,
#   "SELECT DISTINCT agency_num, agency_name, minor_type
#   FROM agency_info
#   WHERE minor_type = 'TIF'
#   "
#   ) %>%
#   mutate(first6 = str_sub(agency_num,1,6),
#          first5 = str_sub(agency_num,1,5)
#          ) %>%
#   select(-minor_type)

muni_agency_names <- DBI::dbGetQuery(
  ptaxsim_db_conn,
  "SELECT DISTINCT agency_num, agency_name, minor_type
  FROM agency_info
  WHERE minor_type = 'MUNI'
  OR agency_num = '020060000'
  "
  ) %>% 
  mutate(first6 = str_sub(agency_num,1,6),
         first5 = str_sub(agency_num,1,5)
         ) %>%
  select(-minor_type)



joined <- left_join(TIFs, muni_agency_names, by = "first6")
joined
```



```{r}
 

# 1057 tax codes in these 6 municipalities
muni_tax_codes <- DBI::dbGetQuery(
  ptaxsim_db_conn, 
  glue_sql("
  SELECT *
  FROM tax_code
  WHERE agency_num IN ({muni_agency_nums$agency_num*})
  AND year = 2021
    ",
  .con = ptaxsim_db_conn
  )
)


# for the 6 communities:
# 4242 tax codes in these 6 municipalities
all_tax_codes <- DBI::dbGetQuery(
  ptaxsim_db_conn, 
  glue_sql("
  SELECT *
  FROM tax_code
  WHERE agency_num IN ({agency_names$agency_num*})
  AND year = 2021
    ",
  .con = ptaxsim_db_conn
  )
)

distinct_tax_codes <- DBI::dbGetQuery(
  ptaxsim_db_conn, 
  glue_sql("
  SELECT DISTINCT tax_code_num
  FROM tax_code
  WHERE agency_num IN ({agency_names$agency_num*})
  AND year = 2021
    ",
  .con = ptaxsim_db_conn
  )
)

all_tax_codes <- all_tax_codes %>% left_join(agency_names)



# All taxcodes and the municipality they belong in!!! 
taxcodes_and_agencies <- DBI::dbGetQuery(
  ptaxsim_db_conn, 
  glue_sql("
  SELECT*
  FROM tax_code
  WHERE agency_num IN ({muni_agency_nums$agency_num*})
  AND year = 2021
    ",
  .con = ptaxsim_db_conn
  )
)

```

## Agencies & Levies 

The `agency` data table in ptaxsim has information for every taxing agency, for every year. Includes multiple variables for EAV (cty_cook_eav), Levy (total_levy, total_reduced_levy, total_final_levy), Extension (total_extension), and agency taxing rate (total_prelim_rate, total_final_rate).

agency_dt: has eav that is taxable (after exemptions and levies). cty_cook_eav is the taxable base for the agency.

__cty_cook_eav should be similar to JD's values for EAV outside of TIFs after exemptions!__

- pin_data should have original EAV.
- Tax bills would have amount to TIF and district too.
  
- merge tif_distribituion in. Then do something like ifelse(tax_code_frozen_eav > 0, tax_code_frozen_eav, eav) to get the EAV that is actually taxable for the district. 


total_levy is what the municipality asked for

sum(total_levy) for all taxing agencies with a common eav in 2021  


> Unsure what the total_non_cap_ext, and reduction variables represent.

```{r}

# agency_rates <- all_tax_codes %>% 
#   filter(minor_type %in% c("MUNI", "TOWNSHIP")) %>%
#   group_by(agency_num) %>%
#   summarize(rate = mean(agency_rate))

all_agency_names <- DBI::dbGetQuery(
  ptaxsim_db_conn,
  "SELECT agency_num, agency_name, minor_type
  FROM agency_info
  "
  )


agency_dt <- DBI::dbGetQuery(
  ptaxsim_db_conn,
  "SELECT *
  FROM agency
  WHERE year = 2021
  "
)

# create the variables for first 5 and6 digits to group cities together later 
agency_dt <- agency_dt %>%   
  mutate(first6 = str_sub(agency_num,1,6),
         first5 = str_sub(agency_num,1,5))



all_taxing_agencies <- agency_dt %>%  left_join(all_agency_names)



all_taxing_agencies <- all_taxing_agencies %>%
  left_join(muni_agency_nums, by = c("first5", "first6")) %>% 
  rename(muni_name2 =  agency_name.y,
        muni_num = agency_num.y,
        agency_num = agency_num.x,
        agency_name = agency_name.x)




six_communities <- all_taxing_agencies %>% 
  filter(muni_name %in% muni_agency_nums$muni_name)

six_communities <- left_join(six_communities, agency_rates, 
                              by = c("muni_num"="agency_num"))


# add taxing agency names and agency type to data table that has eav and extension values
six_communities <- six_communities %>%
  mutate(first2 = str_sub(agency_num, 1,2),
         last2 = str_sub(agency_num,8,9),
         in_SSA = ifelse(minor_type == "SSA", 1,0),
         in_chicago = ifelse(str_detect(muni_name, "CHICAGO"),1,0)) %>%
  select(-c(cty_dupage_eav:cty_livingston_eav))


#total_levy = cty_cook*total_final_rate

six_communities %>% 
  mutate(Prelim_weighted_tax_rate = cty_cook_eav*total_prelim_rate,
    weighted_tax_rate = cty_cook_eav*total_final_rate)  %>%
  group_by(muni_name) %>%
  summarize(EAV = max(cty_cook_eav),
            summed_EAV = sum(cty_cook_eav),
    TotalLevy = sum(total_levy),
    WeightedTaxRate = sum(weighted_tax_rate, na.rm=TRUE),
    prelim_weighted = sum(Prelim_weighted_tax_rate, na.rm=TRUE) ) %>%
  mutate(CompositeTaxRate = WeightedTaxRate/summed_EAV)


  
muni_funds <- DBI::dbGetQuery(
  ptaxsim_db_conn,
  glue_sql(
  "SELECT *
  FROM agency_fund
  WHERE year = 2021
  AND agency_num IN ({muni_agency_nums$agency_num*})"
  ,
  .con = ptaxsim_db_conn
  
))

fund_dt <- DBI::dbGetQuery(
  ptaxsim_db_conn,
  "SELECT *
  FROM agency_fund_info"
  )

muni_funds <- left_join(muni_funds, fund_dt)
```

## Lookup_pin() command.

Exemption data is only stored at the individual pin level. Must pull all pins and then aggregate to the level desired.

`area_data` was results from `tax_bill` command which had how much of the taxbill went to each taxing district or the TIF.

`pin_data` created below with the lookup_pin() function has the information on exemptions, eav before exemptions, and the property type. I also create a dummy variable for if the pin is located in a tax_code within a TIF.

```{r}

class_dict$class_code <- as.character(class_dict$class_code)

pin_data <- lookup_pin(2021, muni_pins$pin) %>% left_join(muni_pins, by = c("pin", "class"))

# get class info and taxcode for each pin

#pin_data <- pin_data %>% left_join(muni_pins, by = c("pin", "class"))

#pin_taxcode_sum <- pin_data %>% group_by(tax_code_num) %>% summarize()
 
# pin_data <- pin_data %>% 
#   left_join(muni_tax_codes, by = c("year", "tax_code_num") ) %>% 
#   rename(muni_name = agency_name)

pin_data <- class_dict %>% select(-c(assessment_level:reporting_group, class_desc:max_size)) %>%
  right_join(pin_data, by = c("class_code" = "class"))
 

```

EAV by Property Type for Municipality:

```{r}
eav_by_class_in_TC <- pin_data %>% 
  group_by(tax_code_num, major_class_code, major_class_type) %>% 
  summarise(eav = sum(eav))
eav_by_class_in_TC
```


### Exemptions by Exemption Type and Property Class

```{r}
exemptions_by_class_per_TC <- pin_data %>% 
  group_by(tax_code_num, major_class_code, major_class_type)%>%
  summarize(eav=sum(eav, na.rm=TRUE),
  exe_homeowner = sum(exe_homeowner, na.rm=TRUE),
  exe_senior = sum(exe_senior, na.rm=TRUE),
  exe_freeze = sum(exe_freeze, na.rm=TRUE),
  exe_longtime_homeowner = sum(exe_longtime_homeowner, na.rm=TRUE),
  exe_disabled = sum(exe_disabled, na.rm=TRUE),
  exe_vet_returning = sum(exe_vet_returning, na.rm=TRUE),
  exe_vet_dis = sum(exe_vet_dis_lt50+exe_vet_dis_50_69+exe_vet_dis_ge70, na.rm=TRUE),
  exe_abate = sum(exe_abate, na.rm=TRUE)
  
) #%>% mutate("Total Exemptions in Class" = sum(exe_homeowner + exe_senior + exe_freeze + exe_longtime_homeowner + exe_disabled + exe_vet_returning + exe_vet_dis + exe_abate)) 

exemptions_by_class %>%   
  rename("Major Class Num" = major_class_code,
         "Major Class Name" = major_class_type)

tax_code_exemptions<-pin_data %>% 
  group_by(tax_code_num)%>%
  summarize(
  exe_homeowner = sum(exe_homeowner),
  exe_senior = sum(exe_senior),
  exe_freeze = sum(exe_freeze),
  exe_longtime_homeowner = sum(exe_longtime_homeowner),
  exe_disabled = sum(exe_disabled),
  exe_vet_returning = sum(exe_vet_returning),
  exe_vet_dis = sum(exe_vet_dis_lt50+exe_vet_dis_50_69+exe_vet_dis_ge70),
  exe_abate = sum(exe_abate),
  "Total Exemptions in Class" = sum(exe_homeowner + exe_senior + exe_freeze + exe_longtime_homeowner + 
                       exe_disabled + exe_vet_returning + exe_vet_dis + exe_abate)
  
) 
```



## Burden Shift due to Homeowners Exemptions

```{r}
pin_data %>% 
  mutate(exemptions = sum(exe_homeowner:exe_abate)) %>%
  group_by(muni_name)%>%
  summarize(
  exe_homeowner = sum(exe_homeowner),
  TotalExemptions = sum(exemptions)
  ) %>% 
  left_join(eav_by_class_inMUNI)
```


## Create TIF vectors

```{r}
# Determining the increment / TIF stuff
tif_agency_nums <- agency_names %>% 
  filter(minor_type == "TIF") %>% 
  select(agency_num)

tif_agency_nums <- left_join(tif_agency_nums, agency_names)

tif_tax_codes2021 <- DBI::dbGetQuery(
  ptaxsim_db_conn, 
  glue_sql("
  SELECT *
  FROM tax_code
  WHERE agency_num IN ({tif_agency_nums$agency_num*})
  AND year = 2021
  ",
  .con = ptaxsim_db_conn
  )
)

tif_tax_codes_2021 <- left_join(tif_tax_codes2021, tif_agency_nums) 


tif_pins2021 <- DBI::dbGetQuery(
  ptaxsim_db_conn,
  glue_sql(
  "SELECT DISTINCT year, pin, class
  FROM pin
  WHERE tax_code_num IN ({tif_tax_codes2021$tax_code_num*})
  AND year = 2021
  ",
  .con = ptaxsim_db_conn
))

tif_pins2021 %>%distinct(pin) %>% count()
```

```{r}
# TIF distributions will include all the unique tax codes that make up
# a TIF

# has eav values for each tax code
tif_distrib <- DBI::dbGetQuery(
  ptaxsim_db_conn, 
  glue_sql("
  SELECT *
  FROM tif_distribution
  WHERE agency_num IN ({tif_agency_nums$agency_num*})
  AND year = 2021
  ",
  .con = ptaxsim_db_conn
  )
)

#tif_distribution


# has same number of pins as method used above but way faster. 
tif_pins_vec <- DBI::dbGetQuery(
  ptaxsim_db_conn,
  glue::glue_sql("
    SELECT DISTINCT pin
    FROM pin
    WHERE tax_code_num IN ({unique(tif_distrib$tax_code_num)*})
    AND year = 2021
  ",
    .con = ptaxsim_db_conn
  )
) %>%
  pull(pin)

```

### TIF pins and Lookup_pin() command.

Exemption data is only stored at the individual pin level. Must pull all pins and then aggregate to the level desired.

`area_data` was results from `tax_bill` command which had how much of the taxbill went to each taxing district or the TIF.

`pin_data` created below with the lookup_pin() function has the information on exemptions, eav before exemptions, and the property type. I also create a dummy variable for if the pin is located in a tax_code within a TIF.

```{r}
pin_data <- pin_data %>%
  # create a dummy variable for if the pin is located in a tax_code that is within a tif
  mutate(in_tif = ifelse(pin %in% tif_pins2021$pin, 1,0))
```

Amount of EAV in and outside of TIF areas:

> This is the EAV before exemptions are subtracted. Comes from lookup_pin() command.

Individual properties are summed together within the municipality.

```{r error=TRUE}

EAV_inandout_TIF <- pin_data %>% 
  group_by(muni_name, in_tif) %>% 
  summarise(eav = sum(eav)) %>% 
  pivot_wider(names_from = in_tif, values_from = eav)%>% 
  mutate(total = `0`+`1`) %>%
  rename(
    "EAV outside TIF" = `0`,
    "EAV within TIF" = `1`,
    "Total EAV before exemptions" = total)

EAV_inandout_TIF


EAV_inandout_TIF <- pin_data %>% 
  group_by(muni_name, major_class_code, major_class_type, in_tif) %>% 
  summarise(eav = sum(eav)) %>% 
  pivot_wider(names_from = in_tif, values_from = eav)%>% 
  mutate(total = `0`+`1`) %>%
  rename("Major Class Num" = major_class_code,
         "Major Class Name" = major_class_type,
    "EAV outside TIF" = `0`,
         "EAV within TIF" = `1`,
         "Total EAV before exemptions" = total)

EAV_inandout_TIF

```




### TIF Exemptions by Exemption Type and Property Class

```{r}
exemptions_by_class <- pin_data %>% 
  group_by(muni_name, major_class_code, major_class_type)%>%
  summarize(
  exe_homeowner = sum(exe_homeowner),
  exe_senior = sum(exe_senior),
  exe_freeze = sum(exe_freeze),
  exe_longtime_homeowner = sum(exe_longtime_homeowner),
  exe_disabled = sum(exe_disabled),
  exe_vet_returning = sum(exe_vet_returning),
  exe_vet_dis = sum(exe_vet_dis_lt50+exe_vet_dis_50_69+exe_vet_dis_ge70),
  exe_abate = sum(exe_abate)
  
) %>%
  mutate("Total Exemptions in Class" = sum(exe_homeowner + exe_senior + exe_freeze + exe_longtime_homeowner + 
                       exe_disabled + exe_vet_returning + exe_vet_dis + exe_abate)) 

exemptions_by_class %>%   
  rename("Major Class Num" = major_class_code,
         "Major Class Name" = major_class_type)
```

## Joining pin_data exemptions & tax_bill revenue data

Join by summed pin_data by property type with summed tax_bill property by type?

```{r eval=FALSE}
joined_table <- revenue_by_class %>% 
  left_join(exemptions_by_class)%>%
  mutate(
                            "Tax Revenue(District+TIF)" = scales::dollar(tax_amt_post_exe),
                            TotalEAV = tax_amt_post_exe/composite_tax_rate,
                            "District Revenue" = scales::dollar(final_tax_to_dist),
                            "TIF Revenue" = scales::dollar(final_tax_to_tif),
                            "Lost Revenue from Exempt." = scales::dollar(tax_amt_exe), # same as tax_amt_pre_exe-tax_amt_post_exe
                            "Tax Share" = scales::percent(dist_tax_share),
                            "EAV in TIF" = final_tax_to_tif/composite_tax_rate,
                            # TIF EAV matches JD almost perfectly, Commercial 5A does not match. 
                            "EAV outside TIF" = final_tax_to_dist/composite_tax_rate
                            ) %>% 
  mutate(EAV = `EAV in TIF`+`EAV outside TIF`)
 # select(-c(tax_amt_post_exe:tax_share, final_tax_to_dist:tax_amt_pre_exe,exe_homeowner:TOTAL)) 


#class_exemptions <- full_join(exemptions_by_class,  class_eav, by = "major_class_code") %>%
#  select(major_class_code, major_class_type, eav, final_tax_to_dist, tax_share)
#class_exemptions

joined_table %>% 
  rename("Major Class Num" = major_class_code,
         "Major Class Name" = major_class_type
         ) %>%
  select(-c(tax_amt_pre_exe:pre_minus_post, exe_homeowner:exe_abate, final_tax_to_dist:tax_amt_exe))



```

## TIF Data tables


```{r}


class_dict$class_code <- as.character(class_dict$class_code)


# tif_pins_dt is unique pins and their exemptions. 
# can sum EAV from this data frame. 
eav_by_class_inTIFS <- tif_pins_dt %>% 
  left_join(class_dict, by = c("class"="class_code")) %>%
  filter(year == 2021) %>%
  group_by(muni_name, major_class_code, major_class_type) %>% 
  summarize(eav_inTIF = sum(eav))

eav_by_class_inTIFS

tif_pins_summ <- tif_pins_dt %>%
  group_by(year) %>%
  #Summed TIF EAV amounts 
  summarize(eav_inTIF_pre_exe = sum(eav)) %>%
  left_join(tif_distrib, by = "year"
            ) %>%
  # amount of value taxed by tifs  = all EAV in tax code - Frozen EAV level
  mutate(tif_increment = tax_code_eav - tax_code_frozen_eav,
         tifrev2 = tax_code_rate*tif_increment)


tif_pins_sum <- tif_pins_summ %>% group_by(year) %>% 
  summarize(tax_code_eav_pre_exe = sum(tax_code_eav),
            tif_increment_pre_exe = sum(tif_increment),
            frozen_eav_pre_exe = sum(tax_code_frozen_eav))
tif_pins_sum
```



### TIFs

`tif_share` = the tax_code_distribution_pct / 100 and comes from tif_distribution table. Comes from lookup_tif() command. Calculated behind the scenes. Documentation for their process is in Gitlab R/lookup.R file

Lets find TIF taxing agencies and combine them with the other taxing agencies for the geographic area:

```{r eav-per-taxcode, eval=FALSE}

eav_per_taxcode <- bills %>%
  group_by(tax_code) %>%
  summarize(eav = sum(eav),
            final_tax_to_tif=sum(final_tax_to_tif),
            final_tax_to_dist=sum(final_tax_to_dist))
# 20 tax_codes with eav inside each one
eav_per_taxcode

eav_perclasstype <- bills %>%
  group_by(major_class_code, major_class_type) %>%
  summarize(#eav = sum(eav), Don't sum eav at this stage, adds pin eavs multiple times since pins are not unique
            final_tax_to_tif=sum(final_tax_to_tif),
            final_tax_to_dist=sum(final_tax_to_dist))
## Counts pins multiple times? 
## Not unique pins so eav for a pin is added multiple times
eav_perclasstype


bills %>%
  group_by(tax_code) %>%
  summarize(
           # exe_toal = sum(exe_total),
            pin_count = n())

bills %>%
  group_by(agency_name) %>%
  summarize(
            pin_count = n())
```

## Taxing Agencies

### TIF Taxing Agencies
```{r tif-taxing-agencies}
# slightly more tax_codes than pins
# most pins have one tax code

# all tax codes within the 6 communities
tax_codes <- lookup_tax_code(2021, muni_pins$pin) 

# now take those tax codes and use them to identify if any tifs are associated with them:
# gives you tax_codes and agency_num and agency_name and the tif_share
# agency_minor_type = TIF
tif_taxing_agencies <- lookup_tif(2021, tax_codes) #all tax codes within a tif

tif_taxing_agencies # tif taxcodes and tif taxing agencies




tif_tax_codes <- as.character(tif_taxing_agencies$tax_code)
#tax_codes <- lookup_tax_code(2021, tif_tax_codes)
```

All non tif taxing agencies:

```{r}
taxing_agencies <- lookup_agency(2021, tax_codes) 

taxing_agencies <- taxing_agencies# %>% left_join(muni_tax_codes)
taxing_agencies %>% arrange(tax_code)

taxing_agencies %>% 
  group_by(agency_name)%>%
  summarize(EAV = first(agency_total_eav),
    total_levy = first(agency_total_ext)) %>% 
  arrange(total_levy)
  
# all_taxing_agencies <- full_join(taxing_agencies, tif_taxing_agencies, 
#                              #    by = c("year", "tax_code", "agency_name", "agency_num")
#                                  ) %>% mutate(in_tif = ifelse(is.na(tif_share), 0, 1))
```

Combine tif and non-tif taxing agencies:

```{r eval=FALSE}


all_taxing_agencies <- rbind(taxing_agencies, tif_taxing_agencies, fill=TRUE) %>% 
  mutate(in_tif = ifelse(is.na(tif_share), 0, 1))


all_taxing_agencies %>%
  group_by(agency_num, agency_name, in_tif) %>% 
  summarize(agency_total_eav = mean(agency_total_eav),
            tif_share = mean(tif_share),
            agency_total_ext=mean(agency_total_ext)
          #  ,summed_eav = sum(eav)
            )
```

Need EAV of each Tax Code. Then we know the TIF eav from TIF tax_codes and other agency EAVs from Tax codes associated with other agencies.

all_taxing_agencies has agency_total_eav but that includes all pin values everywhere that are included in that tax_code (So across all of cook county for Cook County's total eav)

```{r pins-in-tifs}
taxcodes_intifs <- all_taxing_agencies %>% 
  filter(in_tif==1) %>% 
  select(tax_code) %>% 
  distinct() 

taxcodes_intifs

tif_agency_nums <- all_taxing_agencies %>% 
  filter(in_tif==1) %>% 
  select(agency_num) %>% 
  distinct()

tif_agency_nums

```

> Calculate the increment by comparing the total property value in the TIF (in EAV) to the frozen amount. Anything above the frozen amount is taxed by the TIF

Get all pins within TIF area, then compare total equalized assessed value to the value "frozen" by the TIF.

agency_total_eav The total amount of EAV within the taxing district, otherwise known as the "base". This is the denominator when calculating tax rates

agency_total_ext The total extension requested by the taxing district, otherwise known as the "levy". This is the amount the district needs in tax revenue and is the numerator when calculating tax rates

"Changing tax_code_vec "relocates" a PIN by changing the things that are taxing it. This can be useful for counterfactual analysis. For example, if you own property within a school district and want to know what your tax bill would be just outside the district, but otherwise within the same municipality, then you can find the tax code that represents that situation and plug it into tax_bill()"

```{r}

tifs <- lookup_tif(2021, tax_codes)
tifs

```

### Exemption Types in TIFs

```{r}


# has exemptions for pins within TIFs
tif_pins_dt <- lookup_pin(2021, pin = tif_pins_vec) %>% 
  mutate(tax_code = lookup_tax_code(year, pin)) 

tif_pins_dt <- tif_pins_dt %>% left_join(pin_data)

pin_data_summarytable<- tif_pins_dt %>% 
  group_by(muni_name) %>%
#  mutate(inTIF = ifelse(pin %in% tif_pins, 1,0)) %>%
  summarize(#eav = sum(eav),
  exe_homeowner = sum(exe_homeowner),
  exe_senior = sum(exe_senior),
  exe_freeze = sum(exe_freeze),
  exe_longtime_homeowner = sum(exe_longtime_homeowner),
  exe_disabled = sum(exe_disabled),
  exe_vet_returning = sum(exe_vet_returning),
  exe_vet_dis = sum(exe_vet_dis_lt50+exe_vet_dis_50_69+exe_vet_dis_ge70),
  exe_abate = sum(exe_abate)
  
) %>%
  mutate(TOTAL = sum(exe_homeowner + exe_senior + exe_freeze + exe_longtime_homeowner + 
                       exe_disabled + exe_vet_returning + exe_vet_dis + exe_abate))  %>% 
  pivot_longer(cols = exe_homeowner:TOTAL, 
               values_to = "exemptions", 
               names_to = "exemption_type")

pin_data_summarytable


```


## Exporting data into Excel

```{r eval=FALSE}
library(openxlsx)

dataset_names <- list('Taxing Agencies from Tax Bills' = taxing_agencies, #tifs not listed as agencies
                      'Agency Levies' = agency_levies, 

                      # comparable to JD taxing agency table
                      'JD Taxing Agencies Comparison' = nonTIF_taxingagencies_table, 
                      
                      'Total District Revenue' = district_revenue,
                      'Total Revenue' = total_revenue_table,
                      'Revenue by District&TIF' = rev_by_district_and_tif,
                      'Revenue by Class' = revenue_by_class,
                     # 'Taxcodes in TIFs' == tif_tax_codes2021, creates problem with excel file?
                      'TIF pins' = tif_pins_dt,
                      'EAV in and out of TIF' = EAV_inandout_TIF,
                      'Exemptions by Class' = exemptions_by_class,
                      'Joined Table' = joined_table,
                     
                      'EAV in Muni'= eav_by_class_inMUNI, # PRE exemptions! 
                      'eav in tifs'= eav_by_class_inTIFS,
                      'TIF increment&frozenTotal'= tif_pins_sum,
                      'tif taxing agencies' = tif_taxing_agencies,
                     
                      'exemption in tifs'= pin_data_summarytable
                      
                      )

#write.xlsx(dataset_names, file = 'Bridgeview_2021.xlsx')
```
