==> old_versions/11_Township_Burden_Shift.Rmd <==
---
title: "Mapping Property Tax Burden Shift due to Exemptions"
subtitle: "Cook County Townships"
author: "Alea Wilbur"
date: "`r Sys.Date()`"
output:
  html_document:
    toc: yes
    toc_float: yes
    df_print: paged
    code_folding: hide
    code_download: true
---

# Getting Data

```{r setup, include=TRUE, message = FALSE, warning = FALSE}
knitr::opts_chunk$set(echo = TRUE, warning = FALSE, message = FALSE)


library(tidyverse)
library(DBI)
library(data.table)
library(ggspatial)
library(gstat)
library(here)
library(httr)
library(jsonlite)
library(ptaxsim)
library(sf)
library(stars)
library(glue)

# Create the DB connection with the default name expected by PTAXSIM functions
ptaxsim_db_conn <- DBI::dbConnect(RSQLite::SQLite(), "./ptaxsim.db/ptaxsim-2021.0.4.db")


options(digits=4, scipen = 999)

library(sf)
library(jsonlite)
library(httr)

# link to the API output as a JSON file
muni_shp <- read_sf("https://gis.cookcountyil.gov/traditional/rest/services/politicalBoundary/MapServer/2/query?outFields=*&where=1%3D1&f=geojson")

# nicknames <- readxl::read_excel("muni_shortnames.xlsx")


# elem_schools_shp <- read_sf("https://gis.cookcountyil.gov/traditional/rest/services/clerkTaxDistricts/MapServer/2/query?outFields=*&where=1%3D1&f=geojson") %>%
#   mutate(agency_num = str_pad(AGENCY, width = "9" ,side = "left", pad = "0"))


township_shp <- read_sf("https://gis.cookcountyil.gov/traditional/rest/services/politicalBoundary/MapServer/3/query?outFields=*&where=1%3D1&f=geojson")
```

**Data Prep**


```{r}



base_url <- "https://datacatalog.cookcountyil.gov/resource/tx2p-k2g9.json"

# Grab all PINs with township variables

pins_2021 <- GET(
  base_url,
  query = list(
    year = 2021,
    `$select` = paste0(c("pin", "class", "township_code", "township_name",
                         "nbhd_code", "census_puma_geoid"),
                        # "ward_num", 
                        # "ward_chicago_2021",
                       #  "chicago_community_area_name", "chicago_community_area_num",
                       #  "tax_cook_municipality_name", "tax_cook_municipality_number"
                     #  ),
   collapse = ","),
    `$limit` = 500000000L
  )
)

pins_2021 <- fromJSON(rawToChar(pins_2021$content))

```


```{r}
joined_pins <- read_csv("./Output/4C_joined_PINs_bills_and_exemptions.csv")

joined_pins <- left_join(joined_pins, pins_2021, by = c("pin"))

table(joined_pins$township_name)
```

# Proportion of Residential Land in Townships {.tabset .tabset-pills}


__Exemptions within each Townships:__


Percent of EAV that is Class 2 Residential represents all C2 EAV / Township EAV

Percent of taxable Class 2 EAV is non-TIF increment EAV that is taxable by taxing agencies (i.e. exempt EAV from exemptions have been subtracted from totals)

```{r}
### Shows the amount in rental vs owner occupied residential properties ### 

# grouped_exemptions <- exemptions_by_class_per_TC %>% 
#   filter(!is.na(agency_name)) %>%
#   # group_by(agency_name, major_class_code, major_class_type, ResidentialProps, PropType) %>%
#   group_by(PropType, agency_name, agency_num.x, Alea_cat)%>%
#   summarize(eav = sum(eav),
#             exempt_EAV = sum(exempt_EAV, exe_abate, na.rm=TRUE),
#             tax_base_current = sum(tax_base_current, na.rm=TRUE),
#             tax_base_noexemps = sum(tax_base_noexemps, na.rm=TRUE)) %>% ungroup() %>% 
#   select(agency_name, eav, exempt_EAV, everything()) %>% arrange(agency_name)
# 
# grouped_exemptions
```


```{r}
township_shp <- township_shp %>% 
  mutate(name = str_to_title(NAME)) %>%
  mutate(name = case_when( 
    name == "North" ~ "North Chicago",
    name == "South" ~ "South Chicago",
    name == "West" ~ "West Chicago",
    .default = as.character(name)) 
  )


```

```{r pct-exempt}
pct_exempt_MC <- joined_pins %>% 
  group_by(township_code, township_name) %>%
  mutate(total_eav = sum(eav), # eav from pin table. "original" eav before exemptions or tifs
         nonTIF_EAV_post_exemps = sum(final_tax_to_dist/(tax_code_rate/100)),
        # nonTIF_EAV_pre_exemps = sum(final_tax_to_dist +  )
         pin_count  = n(),
         township_exemptEAV = sum(all_exemptions)) %>%
  ungroup() %>%
  group_by(township_code, township_name, major_class_code, total_eav, pin_count, township_exemptEAV, nonTIF_EAV_post_exemps) %>%
  summarize(MC_eav =  sum(eav),
            mc_pc = n(),
            MC_exemptions = sum(all_exemptions),
            MC_nonTIF_EAV_post_exemps = sum(final_tax_to_dist / (tax_code_rate/100) ),
            
) %>%
  mutate(
         pct_eav_MC = MC_eav / total_eav,
       #  pct_pins_w_exe = mc_pc / pin_count,
        exemps_per_resPIN =  MC_exemptions / mc_pc)

pct_exempt_MC

pct_exempt_MC %>% 
  filter(major_class_code == 2) %>%
  full_join(township_shp, by = c( "township_name" = "name") ) %>%
  ggplot(aes(fill = exemps_per_resPIN)) + 
  geom_sf(aes(geometry = geometry), color = "black") + 
  labs(title = "Average Exempt EAV for Residential PINs", 
     #  caption = "The median township has 67.2% of its EAV from Class 2 Residential Properties"
     ) +
  theme_void() + 
  theme(axis.ticks = element_blank(), axis.text = element_blank())+# +#+
  scale_fill_steps2(
    high = "darkblue", low = "black",  #  guide = "legend",
    #midpoint = median(pct_residential$pct_eav_MC),
    show.limits=TRUE,
    na.value = NA,
    name = "Exempt EAV",
    labels = scales::dollar)
```


```{r}
pct_eav_MC <- joined_pins %>% 
  group_by(township_code, township_name) %>%
  mutate(total_eav =  sum(eav)) %>%
  ungroup() %>%
  group_by(township_code, township_name, major_class_code, total_eav) %>%
  summarize(MC_eav =  sum(eav)) %>%
  mutate(
         pct_eav_MC = MC_eav / total_eav )

pct_residential <- pct_eav_MC %>% filter(major_class_code == 2)
  
pct_eav_MC %>% 
  filter(major_class_code == 2) %>%
  full_join(township_shp, by = c( "township_name" = "name") ) %>%
  ggplot(aes(fill = pct_eav_MC)) + 
  geom_sf(aes(geometry = geometry), color = "black") + 
  labs(title = "Residential EAV /  Total EAV in Township", 
       caption = "The median township has 67.2% of its EAV from Class 2 Residential Properties") +
  theme_void() + 
  theme(axis.ticks = element_blank(), axis.text = element_blank())+# +#+
  scale_fill_steps2(
    high = "darkblue", low = "black",  #  guide = "legend",
    midpoint = median(pct_residential$pct_eav_MC),
    show.limits=TRUE,
    na.value = NA,
    name = "% Residential",
    labels = scales::percent)
```




# Taxable Base

Tax base in the tables above are the total EAV outside of TIF areas.

```{r}

TS_taxrates <- joined_pins %>% 
  group_by(township_name, township_code) %>%
  summarize(TownshipLevy = sum(final_tax_to_dist, na.rm = TRUE), # amount billed by tax districts with current exemptions in place
            nonTIF_EAV_post_exemps = sum(final_tax_to_dist/(tax_code_rate/100), na.rm = TRUE),
            TIF_increment_EAV = sum(final_tax_to_tif/(tax_code_rate/100), na.rm=TRUE),  
            Exempt_EAV = sum(tax_amt_exe/(tax_code_rate/100), na.rm=TRUE), 
            Total_EAV = sum((tax_amt_exe+final_tax_to_dist+final_tax_to_tif)/(tax_code_rate/100), na.rm = TRUE)) %>%

  mutate(tax_rate_current = TownshipLevy/nonTIF_EAV_post_exemps,
         nonTIF_EAV_pre_exemps = nonTIF_EAV_post_exemps + Exempt_EAV,
         taxrate_new = TownshipLevy/nonTIF_EAV_pre_exemps,
         taxrate_change = tax_rate_current-taxrate_new) %>% 
  select(township_name, taxrate_change, tax_rate_current, taxrate_new, everything()) %>% 
  arrange(desc(tax_rate_current))

TS_taxrates


# write_csv(TS_taxrates, "./Output/11_Township_Composite_Taxrates.csv")
```

Current composite tax rates for each township are above. Table also includes the new tax rate if there were no exemptions, the levy (aka amount collected by the district from final_tax_to_dist variable), EAV outside of TIFs, amount of exempt EAV, and additional variables.

```{r township-current-taxrates}

TS_taxrates %>%
  full_join(township_shp, by = c( "township_name" = "name") ) %>%
  ggplot(aes(fill = tax_rate_current)) + 
  geom_sf(aes(geometry = geometry), color = "black") + 
  labs(title = "Residential EAV /  Total EAV in Township", 
       caption = "The median composite tax rate at the township level is 10.28%.") +
  theme_void() + 
  theme(axis.ticks = element_blank(), axis.text = element_blank())+# +#+
  scale_fill_steps2(
    high = "darkblue", low = "black",  #  guide = "legend",
    midpoint = median(TS_taxrates$tax_rate_current),
    show.limits=TRUE,
    na.value = NA,
    name = "Current Tax Rate",
    labels = scales::percent)
```

```{r}

burden_table <- TS_taxrates %>% 
  left_join(pct_exempt_MC) %>%
  left_join(pct_eav_MC) %>%
  mutate(rev_collected_current = MC_nonTIF_EAV_post_exemps * tax_rate_current,
         rev_collected_new = (MC_nonTIF_EAV_post_exemps + MC_exemptions )*taxrate_new,
         burden_current = rev_collected_current/TownshipLevy,
         burden_noexemps = rev_collected_new/TownshipLevy, 
         burden_change = burden_noexemps- burden_current) %>%
  mutate(burden_current = ifelse(burden_current>1, 1, burden_current))


```

```{r}
burden_table %>% 

  full_join(township_shp, by = c("township_name" = "name")) %>%
  ggplot(aes(fill = tax_rate_current)) + 
  geom_sf(aes(geometry = geometry), color = "black") + 
  labs(title = "Composite Tax Rate  for Townships", 
    #   caption = "The median township taxrate outside of Chicago has a composite tax rate around 11.1%"
    ) +
    theme_void() + 
  theme(axis.ticks = element_blank(), axis.text = element_blank())+# +#+
    scale_fill_steps2(
    high = "darkblue", low = "black",  #  guide = "legend",
    #  midpoint = median(Current_Taxrates$tax_rate_current),
  #                      n.breaks = 6,
  show.limits=TRUE,
  na.value = NA,
                        name = "Composite Tax Rate",
  labels = scales::percent)
```


```{r}
burden_table %>% 
  filter(major_class_code == 2) %>%
  full_join(township_shp, by = c("township_name" = "name")) %>%
  ggplot(aes(fill = burden_current )) + 
  geom_sf(aes(geometry = geometry), color = "black") + 
  labs(title = "Current C2 Tax Burden", 
       caption = "Tax Burden is the amount of the Levy revenue collected paid by C2 properties"
    ) +
    theme_void() + 
  theme(axis.ticks = element_blank(), axis.text = element_blank())+# +#+
    scale_fill_steps2(
    high = "darkblue", low = "black",  #  guide = "legend",
    #  midpoint = median(Current_Taxrates$tax_rate_current),
  #                      n.breaks = 6,
  show.limits=TRUE,
  na.value = NA,
                        name = "Tax Burden",
  labels = scales::percent)
```


### Chicago by Ward

```{r}
#ward_link <- "https://gis.cookcountyil.gov/hosting/rest/services/Hosted/Parcel2021_enhancedAll/FeatureServer/0/query?where=chicagoward%20%3D%20'49'&outFields=*&outSR=4326&f=json"
```
